### 2022-10-05 #4819 @editorialbot
[PRE REVIEW]: Metrics As Scores: A Tool- and Analysis Suite and Interactive Application for Exploring Context-Dependent Distributions
**Submitting author:** @mrshoenel (<a href="http://orcid.org/0000-0001-7937-1645">Sebastian Hönel</a>)
**Repository:** https://github.com/MrShoenel/metrics-as-scores
**Branch with paper.md** (empty if default branch): paper
**Version:** v1.0.8
