### 2022-10-06 d25123e
Remove badge from TOC
