### 2022-12-05 9e9a252
Add a toggle-legend button to webapp
