### 2022-12-16 MrShoenel/metrics-as-scores#1 @mdhaber
Meta-issue for mdhaber review of JOSS submission
This is a detailed list of notes corresponding with openjournals/joss-reviews#4913. The checklist below may be modified as the review progresses. I'll create a separate issue for any items that require substantial discussion.

  - [x] Yes, but paper title in `readme.md` should match title of JOSS submission. _Update: I don't remember exactly where I was looking before, but the readme doesn't seem to refer to this paper anymore._
  - Yes, GPLv3.
  - [x] Yes, @MrShoenel is the only contributor. But in that case, how did the other authors contribute to the project?
- [x] **Substantial scholarly effort:** Does this submission meet the scope eligibility described in the [JOSS guidelines](https://joss.readthedocs.io/en/latest/submitting.html#substantial-scholarly-effort)?
  - See "Details" for specific considerations. Question for the editor: most of the commits took place over a period of two months, but at least one commit toward the beginning (266d01068c12b79e8467e28aa56913bc629d86fc) suggests that there was substantial work before this. Does this meet the standard that "As a rule of thumb, JOSS’ minimum allowable contribution should represent not less than three months of work for an individual"?

  - Age of software (is this a well-established software project) / length of commit history.
    - Most the commits occured in August and September
  - Number of commits.
    - 160
  - Number of authors
    - 1
  - Total lines of code (LOC). Submissions under 1000 LOC will usually be flagged, those under 300 LOC will be desk rejected.
    - Nearly 2000 Python lines in src plus web app. Much of the Python code wraps existing code, though. Are there key algorithmic parts I should look at?
  - Whether the software has already been cited in academic papers.
    - I don't see any
  - Whether the software is sufficiently useful that it is likely to be cited by your peer group.
    - TBD
  - In addition, JOSS requires that software should be feature-complete (i.e., no half-baked solutions)
    - This doesn't look half-baked
  - packaged appropriately according to common community standards for the programming language being used (e.g., Python, R),
    - yes, it's PIP-installable
  - designed for maintainable extension (not one-off modifications of existing tools).
    - yes
  - “Minor utility” packages, including “thin” API clients, and single-function packages are not acceptable.
    - This is not a single-function package

  - [x] ~~I don't think there is any original data. The Qualitas.class corpus itself is not claimed as part of this paper. Is this accurate?~~ _Update: there are [three datasets ](https://github.com/MrShoenel/metrics-as-scores#known-datasets) mentioned in the Readme, and one is mentioned in the paper. All are available and documented by accompanying PDFs._
  - [x] I don't think any original results are claimed. Is this accurate? _Update: yes. "There are **no** new results claimed in the JOSS paper, just the dataset as its own publication. (https://github.com/MrShoenel/metrics-as-scores/issues/1#issuecomment-1437451432)"_

  - `pip install metrics-as-scores` seems to have completed successfully. I don't see any instructions for testing the installation or running the software locally, though.
  - ~~I did not attempt the ["Stand-alone Usage / Development Setup"](https://github.com/MrShoenel/metrics-as-scores#2-stand-alone-usage--development-setup).~~ Development installation worked.
  - Maybe. I would like the authors to list the functional claims of the software concisely before I judge this. Since I am not finding instructions for interacting with the software locally, I am relying on https://metrics-as-scores.ml/webapp. There, I see lots of probability density functions overlayed on the same graph. IIUC, each of them was generated by fitting some data to ~120 distributions and keeping only the best fit. But I have many questions.
    - What does the data represent? I'm still not sure what the Qualitas.class corpus data is. (I'd suggest showing less by default. This is a lot to be confronted with.)
    - There are a few fitting metrics that are listed - which is considered when selecting the best fit? (Looks like KS 2-sample. Why? This will be stochastic, and there are deterministic statistics available.)
    - Are there any claims about the statistical interpretation of the results, or are statistical methods being used for convenience? As an obvious example of what I'm looking for: I don't think the software claims that the observed data _were_ drawn from the fitted distribution, but if it did, I would say that this has not been confirmed because it would be an abuse of goodness-of-fit tests to make such a claim.
    - The paper mentions lots of functionality that does not seem to be demonstrated by the web interface - ANOVA, TukeyHSD, etc.
  - Authors - I don't see anything that I would consider to be a performance claim ("fast", "memory-efficient", "accurate"). Is that correct?
  - Addressed by https://github.com/MrShoenel/metrics-as-scores/issues/1#issuecomment-1430527332

  - It is stated, but IMO the language used is too abstract to be easily interpreted by a general audience or even a computational statistics developer (myself). I would suggest adding a very concrete example to demonstrate what is meant by key terms "raw data", "metrics", "scores", "distance", "context". I think that in a domain-independent statistics context, I would call "raw data" -> "sample(s)", "metric" -> "statistic" (because I do not think it satisfies the mathematical definition of a "metric"), "scores" -> something related to a CDF fitted either parametrically or nonparametrically. I confident that I understand "distance" and "context" correctly, though.
  - Is there a way to distinguish between "metric" as a mathematical function and "metric" as the numerical value that function assumes given particular "raw data"? Similar question for "score".
  - Maybe? The example [website](https://metrics-as-scores.ml/webapp) seems to show how the software works on real-world data, but I don't understand what that data is. The ["Use your own data"](https://github.com/MrShoenel/metrics-as-scores#3-use-your-own-data) might satisfy this criterion, but IMO it should include a much simpler example with a minimum number of "raw data" samples, and it should show each of the claimed features individually (e.g. empirical distribution, KDE, MLE, ANOVA, TukeyHSD). See #5.
  - I don't see any rendered API documentation, and AFAICT, none of the Python code has docstrings.  See #2.
  - There are some tests. Pytest is not listed as a dependency, and there are no instructions for running the tests. I am not sure whether there are adequate tests because there is not documentation for public functions. See #4.
- [x] **Community guidelines:** Are there clear guidelines for third parties wishing to 1) Contribute to the software 2) Report issues or problems with the software 3) Seek support.
  - ~~I don't see any. See #3.~~ *Update: this has been resolved.*

  - ~~There is a summary, but I don't think it's clear to non-specialists. A simple, concrete example would help. I'll link to a separate issue about this.~~
  - This has been improved.
  - ~~I think that once the introduction is more accessible, the statement of need will satisfy this criterion with minor adustments.~~
  -   This has been improved.
  - No, I don't think so. See https://github.com/MrShoenel/metrics-as-scores/issues/9
  - Yes, mostly. I can make an issue with copyediting suggestions when the paper is closer to its final form.
  - It is extensive. I'll have to check for citation syntax.
  - See https://github.com/MrShoenel/metrics-as-scores/issues/8
