### 2022-12-16 MrShoenel/metrics-as-scores#1 @mdhaber
To check off remaining boxes, I would appreciate more information from the authors. Let's take this a few items at a time to keep things organized.

@MrShoenel can you answer the following questions:

1.  How did the four authors of the paper contribute to the project, assuming @MrShoenel is the GH handle for only one of the authors?
2. Is there any original data that is claimed to be associated with this paper? "Qualitas.class" is the only data set I see, and is not claimed as a contribution of this paper.
3. Do you consider the paper's claim to include the results, e.g. those displayed at https://metrics-as-scores.ml/webapp? Or is the paper primarily the methods/code for producing results given data? (I think the latter.)
4. Are there any performance claims ("fast", "memory-efficient", "accurate") that I am missing? That's fine if there are none; I just wanted to make sure I didn't miss any.
