### 2023-01-06 MrShoenel/metrics-as-scores#1 @MrShoenel
Thanks for the detailed feedback :) I think we can answer some parts directly, others will require some effort. I believe that we will begin next week and have it finished the week after.
