### 2023-02-01 #4913 @MrShoenel
There were some unexpected delays on my side, but I had begun work on issues #4 (automated tests) and #2 (API documentation). I will complete some of it this week and reference the changes in these issues then.
I am confident to have addressed all of @mdhaber's comments by next week :)
I will post here then, too!
