### 2023-02-02 1fe8bd2
Read the package's version directly from pyproject.toml
