### 2023-02-05 b011f1e
Start to add proper documentation for fitting problems
