+++ src/metrics_as_scores/distribution/fitting_problems.py
@@ -1,3 +1,14 @@
+"""
+This module contains :code:`pymoo` fitting problems that allow fitting
+distributions to almost arbitrary discrete data. The discrete random
+variables in :code:`scipy` do not have a :code:`fit()`-method, as their fitting
+often requires a global search. Also, many distributions require
+discrete parameters, or a mixture of real and integer parameters.
+The problems in this module provide a generalized way for `pymoo` to
+find parameters for all of :code:`scipy`'s discrete random variables.
+
+"""
+
 import numpy as np
 from nptyping import Float, NDArray, Shape
 from scipy.stats._discrete_distns import rv_discrete, bernoulli_gen, betabinom_gen, binom_gen, boltzmann_gen, dlaplace_gen, geom_gen, hypergeom_gen, logser_gen, nbinom_gen, nchypergeom_fisher_gen, nchypergeom_wallenius_gen, nhypergeom_gen, planck_gen, poisson_gen, randint_gen, skellam_gen, yulesimon_gen, zipf_gen, zipfian_gen
@@ -8,7 +19,32 @@ from pymoo.core.variable import Real, Integer
 
 
 class MixedVariableDistributionFittingProblem(ElementwiseProblem):
+    """
+    This is the base class for fitting all of ``scipy``'s discrete random variables.
+    Therefore, it accepts a dictionary of parameters for each distribution to find
+    optimal values for.
+    """
     def __init__(self, dist: rv_discrete, data: NDArray[Shape['*'], Float], vars: dict[str, Variable], n_ieq_constr: int=0, **kwargs):
+        """
+        Constructor for a fitting any discrete random variable with one or more
+        parameters that can be of any type as supported by ``pymoo.core.variable.Variable`` (e.g., ``Integer``, ``Real``, etc.).
+
+        Parameters
+        ----------
+        dist: ``rv_discrete``
+            An instance of the concrete discrete random variable that should be fit to the data.
+        
+        vars: ``dict[str, Variable]``
+            An ordered dictionary of named variables to optimize. These must correspond
+            one to one with the variable names of those defined for the random variable.
+
+        data: ``NDArray[Shape['*'], Float]``
+            The data the distribution should be fit to.
+        
+        n_ieq_constr: ``int``
+            Number of inequality constraints. If there are any, then the problem also
+            overrides ``_evaluate()`` and sets values for each constraint.
+        """
         self.ext = int(np.max(data) - np.min(data))
         self.dist = dist
         # All of them have 'loc':
@@ -19,6 +55,30 @@ class MixedVariableDistributionFittingProblem(ElementwiseProblem):
         self.data = data
     
     def _evaluate(self, X, out, *args, **kwargs) -> dict:
+        r"""
+        This is an internal method that evaluates the discrete random variable's
+        negative log likelihood, given the currently set values for all of its
+        variables (stored in ``X``). This method is called by ``pymoo``, so be
+        sure to check out their references, too.
+        Note that the ``X``-dictionary is used to build :math:`\theta`, the vector
+        of parameters for the random variable. The order of the parameters is that
+        vector depends on the order of ``self.vars``.
+        This method usually does not need to be overridden, except for when, e.g.,
[…]
