### 2023-02-05 c8bd647
Remove workaround for bug #17026 in old scipy
