+++ docs/metrics_as_scores.data.rst
@@ -1,0 +1,45 @@
+metrics\_as\_scores.data package
+================================
+
+Submodules
+----------
+
+metrics\_as\_scores.data.metrics module
+---------------------------------------
+
+.. automodule:: metrics_as_scores.data.metrics
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+metrics\_as\_scores.data.pregenerate module
+-------------------------------------------
+
+.. automodule:: metrics_as_scores.data.pregenerate
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+metrics\_as\_scores.data.pregenerate\_distns module
+---------------------------------------------------
+
+.. automodule:: metrics_as_scores.data.pregenerate_distns
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+metrics\_as\_scores.data.pregenerate\_fit module
+------------------------------------------------
+
+.. automodule:: metrics_as_scores.data.pregenerate_fit
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+Module contents
+---------------
+
+.. automodule:: metrics_as_scores.data
+   :members:
+   :undoc-members:
+   :show-inheritance:
