+++ docs/metrics_as_scores.cli.rst
@@ -1,0 +1,69 @@
+metrics\_as\_scores.cli package
+===============================
+
+Submodules
+----------
+
+metrics\_as\_scores.cli.Cli module
+----------------------------------
+
+.. automodule:: metrics_as_scores.cli.Cli
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+metrics\_as\_scores.cli.CreateDataset module
+--------------------------------------------
+
+.. automodule:: metrics_as_scores.cli.CreateDataset
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+metrics\_as\_scores.cli.KnownDatasets module
+--------------------------------------------
+
+.. automodule:: metrics_as_scores.cli.KnownDatasets
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+metrics\_as\_scores.cli.LocalWebserver module
+---------------------------------------------
+
+.. automodule:: metrics_as_scores.cli.LocalWebserver
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+metrics\_as\_scores.cli.MainWorkflow module
+-------------------------------------------
+
+.. automodule:: metrics_as_scores.cli.MainWorkflow
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+metrics\_as\_scores.cli.Workflow module
+---------------------------------------
+
+.. automodule:: metrics_as_scores.cli.Workflow
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+metrics\_as\_scores.cli.helpers module
+--------------------------------------
+
+.. automodule:: metrics_as_scores.cli.helpers
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+Module contents
+---------------
+
+.. automodule:: metrics_as_scores.cli
+   :members:
+   :undoc-members:
+   :show-inheritance:
