### 2023-02-05 0ba1902
Add CLI-related helpers and datatypes
