### 2023-02-05 b63db81
Create base-class work CLI-based workflows using questionary and rich
