### 2023-02-05 30d515a
Create override 'askt' that can use tuples
