+++ src/metrics_as_scores/cli/Workflow.py
@@ -1,4 +1,4 @@
-from typing import TypeVar
+from typing import Any, TypeVar
 from rich.console import Console
 import questionary
 from questionary import Choice
@@ -14,11 +14,16 @@ class Workflow:
         self.style_mas = 'bold fg:#8B8000 bg:#000'
     
     def ask(self, options: list[str], prompt: str='You now have the following options:', rtype: T=int) -> T:
-        choices = list([Choice(title=options[idx], value=idx) for idx in range(len(options))])
-        res = self.q.select(message=prompt, choices=choices, use_shortcuts=True).ask()
+        res = self.askt(
+            options=list([(options[idx], idx) for idx in range(len(options))]),
+            prompt=prompt)
         if rtype == str:
             return options[res]
         return res
+    
+    def askt(self, options: list[tuple[str, T]], prompt: str='You now have the following options:') -> T:
+        choices = list([Choice(title=options[idx][0], value=options[idx][1]) for idx in range(len(options))])
+        return self.q.select(message=prompt, choices=choices, use_shortcuts=True).ask()
     
     def print_info(self, text_normal: str, text_vital: str=None, end: str=None) -> None:
         kwargs = {}
