### 2023-02-05 179cfe8
Add main workflow for CLI (only few things implemented)
