### 2023-02-05 4caf7c5
Add CLI workflow for local webserver (process-based and internal)
