### 2023-02-05 a92c9a1
Add CLI workflow to list available online datasets
