### 2023-02-05 66103c7
Fix a problem where the version was not correctly retrieved
