+++ src/metrics_as_scores/__version__.py
@@ -2,4 +2,4 @@ from toml import load
 from pathlib import Path
 
 proj_dir = Path(__file__).parent.parent.parent
-__version__ = load(proj_dir.joinpath('./pyproject.toml')).get('version')
+__version__ = load(proj_dir.joinpath('./pyproject.toml'))['tool']['poetry']['version']
