+++ src/metrics_as_scores/cli/CreateDataset.py
@@ -1,0 +1,185 @@
+from typing import Literal, Union
+from Workflow import Workflow
+from helpers import LocalDataset, isint, isnumeric
+from re import match
+from os import makedirs
+from json import dump
+import numpy as np
+import pandas as pd
+
+from pathlib import Path
+this_dir = Path(__file__).resolve().parent
+datasets_dir = this_dir.parent.parent.parent.joinpath('./datasets')
+
+
+class CreateDatasetWorkflow(Workflow):
+    """
+    This workflow creates an own, local dataset from a single data source that can
+    be read by Pandas from file or URL. The original dataset needs three columns:
+    For a single sample, one column holds the numeric observation, one holds the
+    ordinal type, and one holds the context associated with it. A dataset can hold
+    one or more types, but should hold at least two contexts in order to compare
+    distributions of a single sample type across contexts.
+    This workflow creates the entire dataset: The manifest JSON, the parametric fits,
+    the pre-generated distributions. When done, the dataset is installed, such that it
+    can be discovered and used by the local Web-Application. If you wish to publish
+    your dataset so that it can be used by others, start the Bundle-workflow from the
+    previous menu afterwards.
+    """
+    def __init__(self) -> None:
+        super().__init__()
+        self.regex_id = r'^[a-z]+[a-z-_\d]*?[a-z\d]+?$'
+    
+
+    def _set_ideal_values(self, qtypes: list[str]) -> dict[str, Union[float,int]]:
+        type_dict = {}
+        for t in qtypes:
+            type_dict[t] = None
+        
+        ready = False
+        while not ready:
+            use_type_idx = self.ask(
+                rtype=int,
+                prompt='Select a type to set a default value for:',
+                options=list([f'{item[0]}={"None" if item[1] is None else item[1]}' for item in type_dict.items()]))
+            ideal = self.q.text(message=f'Set the ideal value for {qtypes[use_type_idx]} (leave blank for None):', validate=lambda s: s == '' or isnumeric(s)).ask()
+            if ideal == '':
+                ideal = None
+            else:
+                if isint(ideal):
+                    ideal = int(ideal)
+                else:
+                    ideal = float(ideal)
+            type_dict[qtypes[use_type_idx]] = ideal
+
+            ready = not self.q.confirm(message='Continue setting ideal values?', default=True).ask()
+        
+        return type_dict
+    
+
+    def _read_csv(self, path_like: str) -> pd.DataFrame:
+        sep = self.q.text(message='What is the separator used?', default=',', validate=lambda s: s in [',', ';', ' ', '\t']).ask()
+        dec = self.q.text(message='What is the decimal point?', default='.', validate=lambda s: s in ['.', ',']).ask()
+        return pd.read_csv(filepath_or_buffer=path_like, sep=sep, decimal=dec)
+    
+    def _read_excel(self, path_like: str) -> pd.DataFrame:
+        sheet = self.q.text(message='Which sheet would you like to read?', default='0').ask()
+        if isint(sheet):
+            sheet = int(sheet)
+        header = self.q.text('Zero-based index of header column?', default='0', validate=isint).ask()
+        return pd.read_excel(io=path_like, sheet_name=sheet, header=header)
+    
+
+    def _is_feature_discrete(self, df: pd.DataFrame, col_data: str, col_type: str, use_type: str) -> bool:
+        temp = df.loc[df[col_type] == use_type, :]
+        vals = temp[col_data].to_numpy()
+        # Check if all values are integer.
+        return np.all(np.mod(vals, 1) == 0)
+
+    
+
+    def _create_manifest(self) -> tuple[LocalDataset, pd.DataFrame]:
[…]
