### 2023-02-05 658abcc
Add main CLI entry point and 'mas' runnable script
