### 2023-02-05 19810da
Improve documentation on fitting problems
