+++ src/metrics_as_scores/distribution/fitting_problems.py
@@ -61,7 +61,7 @@ class MixedVariableDistributionFittingProblem(ElementwiseProblem):
         variables (stored in ``X``). This method is called by ``pymoo``, so be
         sure to check out their references, too.
         Note that the ``X``-dictionary is used to build :math:`\theta`, the vector
-        of parameters for the random variable. The order of the parameters is that
+        of parameters for the random variable. The order of the parameters in that
         vector depends on the order of ``self.vars``.
         This method usually does not need to be overridden, except for when, e.g.,
         it is required to evaluate (in-)equality constraints (that is, whenever
@@ -94,10 +94,11 @@ class Fit_bernoulli_gen(MixedVariableDistributionFittingProblem):
 
     Notes
     -----
-    Does not override `_evaluate()` and does not have any (in-)equality constraints.
+    Does not override :meth:`MixedVariableDistributionFittingProblem._evaluate()`
+    and does not have any (in-)equality constraints.
     Calls the super constructor with these variables (in this order):
 
-    - ``p``: (``int``) :math:`\left[0,1\right]]`
+    - ``p``: (``int``) :math:`\left[0,1\right]`
     """
     def __init__(self, data: NDArray[Shape['*'], Float], **kwargs):
         super().__init__(dist=bernoulli_gen(), data=data, vars={
@@ -112,7 +113,8 @@ class Fit_betabinom_gen(MixedVariableDistributionFittingProblem):
 
     Notes
     -----
-    Does not override `_evaluate()` and does not have any (in-)equality constraints.
+    Does not override :meth:`MixedVariableDistributionFittingProblem._evaluate()`
+    and does not have any (in-)equality constraints.
     Calls the super constructor with these variables (in this order):
 
     - ``n``: (``int``) :math:`\left(0,1e^{4}\right)`
@@ -134,7 +136,8 @@ class Fit_binom_gen(MixedVariableDistributionFittingProblem):
 
     Notes
     -----
-    Does not override `_evaluate()` and does not have any (in-)equality constraints.
+    Does not override :meth:`MixedVariableDistributionFittingProblem._evaluate()`
+    and does not have any (in-)equality constraints.
     Calls the super constructor with these variables (in this order):
 
     - ``n``: (``int``) :math:`\left(1,25e^{3}\right)`
@@ -154,7 +157,8 @@ class Fit_boltzmann_gen(MixedVariableDistributionFittingProblem):
 
     Notes
     -----
-    Does not override `_evaluate()` and does not have any (in-)equality constraints.
+    Does not override :meth:`MixedVariableDistributionFittingProblem._evaluate()`
+    and does not have any (in-)equality constraints.
     Calls the super constructor with these variables (in this order):
 
     - ``lambda`` [:math:`\lambda`]: (``float``) :math:`\left(0,1e^{5}\right)`
@@ -174,7 +178,8 @@ class Fit_dlaplace_gen(MixedVariableDistributionFittingProblem):
 
     Notes
     -----
-    Does not override `_evaluate()` and does not have any (in-)equality constraints.
+    Does not override :meth:`MixedVariableDistributionFittingProblem._evaluate()`
+    and does not have any (in-)equality constraints.
     Calls the super constructor with these variables (in this order):
 
     - ``a``: (``float``) :math:`\left(5e^{-324},1e^4\right)`
@@ -192,7 +197,8 @@ class Fit_geom_gen(MixedVariableDistributionFittingProblem):
 
     Notes
     -----
-    Does not override `_evaluate()` and does not have any (in-)equality constraints.
[…]
