### 2023-02-07 e27ad3c
Delete enum MetricID and class QualitasCorpusMetricsExtractor
