+++ src/metrics_as_scores/data/metrics.py
@@ -1,104 +1,0 @@
-from typing import Union
-from typing import Iterable
-from xml.etree.cElementTree import parse, Element, iterparse
-from strenum import StrEnum
-from re import match
-from os import walk
-from typing import Iterable
-import pandas as pd
-
-
-# Complexity metrics: A metrics suite for object oriented design
-# Software package metrics: https://en.wikipedia.org/wiki/Software_package_metrics and Robert Cecil Martin (2002). Agile Software Development: Principles, Patterns and Practices. Pearson Education. ISBN 0-13-597444-5.
-# Counts are low-level stuff
-# In some cases, a user-defined "normality" is more useful such that new code fits right in with the existing one
-
-class MetricID(StrEnum):
-    TLOC = 'Total Lines of Code' # None -> user/context preference
-    NOP = 'Number of Packages' # None
-    NOC = 'Number of Classes' # None
-    NOI = 'Number of Interfaces' # None
-    NOM = 'Number of Methods' # None
-    NOF = 'Number of Attributes' # None
-    NORM = 'Number of Overridden Methods' # None
-    PAR = 'Number of Parameters' # None
-    NSM = 'Number of Static Methods' # None
-    NSF = 'Number of Static Attributes' # None
-    WMC = 'Weighted Methods per Class' # [0]; is a complexity measure; however, it depends on the definition of c, since it's \sum_{i=1}^{n} c_i, and c is deliberately left undefined; In the qualitas.class corpus, inf{c}=0, because we actually observed it
-    DIT = 'Depth of Inheritance Tree' # None
-    NSC = 'Number of Children' # None
-    LCOM = 'Lack of Cohesion in Methods' # 0; also complexity; Zero lack would be best; however, this is also a candidate for arbitrary user-defined ideal values -> maybe you are writing a class that is a collection of static methods, like C# extensions methods, so LCOM would be high.
-    MLOC = 'Method Lines of Code' # None
-    SIX = 'Specialization Index' # 0; because a high specialization is undesirable; SIX = (NORM * DIT) / ([No. of Added Methods] + [No. of Inherited methods] + NORM) -> we have observed 0
-    VG = 'McCabe Cyclomatic Complexity' # 1, lowest possible
-    NBD = 'Nested Block Depth' # [0], or minimum possible; less nesting is considered better. In our case the average is returned and we observed values as low as ~0.3, so I guess 0 would be the theoretically lowest. However, also subject to user pref. I recommend NOT setting an ideal value since it is also a counting metric of some type.
-    RMD = 'Normalized Distance' # 0; -> Distance from the main sequence (D): The perpendicular distance of a package from the idealized line A + I = 1. D is calculated as D = | A + I - 1 |. This metric is an indicator of the package's balance between abstractness and stability. A package squarely on the main sequence is optimally balanced with respect to its abstractness and stability. Ideal packages are either completely abstract and stable (I=0, A=1) or completely concrete and unstable (I=1, A=0). The range for this metric is 0 to 1, with D=0 indicating a package that is coincident with the main sequence and D=1 indicating a package that is as far from the main sequence as possible.
-    CA = 'Afferent Coupling' # 0; but also user-preference
-    CE = 'Efferent Coupling' # 0; but also user-preference
-    RMI = 'Instability' # 0 -> Instability (I): The ratio of efferent coupling (Ce) to total coupling (Ce + Ca) such that I = Ce / (Ce + Ca). This metric is an indicator of the package's resilience to change. The range for this metric is 0 to 1, with I=0 indicating a completely stable package and I=1 indicating a completely unstable package.
[…]
