### 2023-02-07 b5c24c7
Remove dummy 'Dataset' class and move JSON dataset definitions out
