### 2023-02-07 5615d13
Move Literal type StatTest_Types
