### 2023-02-07 e90c977
Add new CLI workflow for fitting parametric RVs (not finished)
