+++ src/metrics_as_scores/cli/FitParametric.py
@@ -1,0 +1,140 @@
+from typing import Iterable, Union
+from Workflow import Workflow
+from pathlib import Path
+from json import load
+from helpers import get_local_datasets
+from metrics_as_scores.distribution.distribution import LocalDataset
+from scipy.stats._distn_infrastructure import rv_continuous, rv_discrete
+from metrics_as_scores.distribution.fitting import Fitter, FitterPymoo, Continuous_RVs, Discrete_Problems
+from questionary import Choice
+from rich.progress import Progress
+from json import dump
+import pandas as pd
+
+this_dir = Path(__file__).resolve().parent
+datasets_dir = this_dir.parent.parent.parent.joinpath('./datasets')
+
+
+
+class FitParametricWorkflow(Workflow):
+    def __init__(self) -> None:
+        super().__init__()
+        self.use_ds: LocalDataset = None
+        self.df: pd.DataFrame = None
+        self.fits_dir: Path = None
+        self.use_fitter: type[Fitter] = FitterPymoo
+    
+
+    def _select_continuous_rvs(self) -> Iterable[rv_continuous]:
+        self.q.print('''
+You can now select the continuous random variables that you want to
+attempt to fit to the data. Select all in case you intend to re-
+distribute and publicize your dataset. It is often not worth to de-
+select distributions unless you have a specific reason to do so,
+because fitting a continuous distribution is comparatively cheap and
+fast.'''.strip())
+        return self.q.checkbox(message='Select continuous random variables:', choices=[Choice(title=type(rv).__name__, value=rv, checked=True) for rv in Continuous_RVs]).ask()
+    
+    def _select_discrete_rvs(self) -> Iterable[rv_discrete]:
+        self.q.print('''
+You can now select the discrete random variables that you want to
+additionally attempt to fit to discrete quantity types. Select all
+in case you intend to redistribute and publicize your dataset.
+While there are much fewer discrete random variables, their fitting
+is dramatically more computationally expensive to compute, since a
+global search has to be performed.'''.strip())
+        return self.q.checkbox(message='Select discrete random variables:', choices=[Choice(title=dp[0], value=dp[1], checked=True) for dp in Discrete_Problems.items()]).ask()
+    
+
+    def _iterable_fits(self, rvs: Iterable[Union[rv_continuous, rv_discrete]]) -> Iterable[tuple[Path, str, str, Union[rv_continuous, rv_discrete]]]:
+        from itertools import product
+        for comb in product(self.use_ds['qtypes'].keys(), self.use_ds['contexts'], rvs):
+            fit_file = f'{comb[0]}_{comb[1]}_{"c" if issubclass(type(comb[2]), rv_continuous) else "d"}_{type(comb[2]).__name__}.json'
+            fit_path = self.fits_dir.joinpath(f'./{fit_file}')
+            if fit_path.exists():
+                self.print_info(text_normal='Fit already exists, skipping: ', text_vital=fit_file)
+            else:
+                yield (fit_path, comb[0], comb[1], comb[2])
+    
+
+    def _fit_continuous_rvs(self, rvs: Iterable[rv_continuous]) -> None:
+        """
+        For each quantity type, and for each context, fit each random variable
+        once, and save the result in "`qtype_context_rv`". If that file exists,
+        skip to next.
+        """
+        perms = list(self._iterable_fits(rvs=rvs))
+        with Progress() as progress:
+            task1 = progress.add_task('[cyan]Fitting continuous distributions...', total=len(perms))
+            for fit_path, qtype, ctx, rv in perms:
+                data = self.df[(self.df[self.use_ds['colname_type']] == qtype) & (self.df[self.use_ds['colname_context']] == ctx)][self.use_ds['colname_data']].to_numpy()
+                fitter = self.use_fitter(dist=type(rv))
[…]
