### 2023-02-07 91c317a
Use tuples and short ids in the MainWorkflow; also add FitParametricWorkflow
