### 2023-02-07 d96107a
Remove obsolete static method Dataset::fit_parametric()
