### 2023-02-07 94462c5
Remove now-obsolete method Dataset::analyze_distr()
