+++ src/metrics_as_scores/distribution/distribution.py
@@ -54,7 +54,7 @@ class KnownDataset(JsonDataset):
 
 
 class Density(ABC):
-    def __init__(self, range: tuple[float, float], pdf: Callable[[float], float], cdf: Callable[[float], float], ppf: Callable[[float], float]=None, ideal_value: float=None, dist_transform: DistTransform=DistTransform.NONE, transform_value: float=None, metric_id: MetricID=None, domain: str=None, **kwargs) -> None:
+    def __init__(self, range: tuple[float, float], pdf: Callable[[float], float], cdf: Callable[[float], float], ppf: Callable[[float], float]=None, ideal_value: float=None, dist_transform: DistTransform=DistTransform.NONE, transform_value: float=None, qtype: str=None, context: str=None, **kwargs) -> None:
         self.range = range
         self._pdf = pdf
         self._cdf = cdf
@@ -62,8 +62,8 @@ class Density(ABC):
         self._ideal_value = ideal_value
         self._dist_transform = dist_transform
         self._transform_value: float = None
-        self._metric_id = metric_id
-        self._domain = domain
+        self._qtype = qtype
+        self._context = context
         self._practical_domain: tuple[float, float] = None
         self._practical_range_pdf: tuple[float, float] = None
 
@@ -78,12 +78,12 @@ class Density(ABC):
     
 
     @property
-    def metric_id(self) -> Union[MetricID, None]:
-        return self._metric_id
+    def qtype(self) -> Union[str, None]:
+        return self._qtype
     
     @property
-    def domain(self) -> Union[str, None]:
-        return self._domain
+    def context(self) -> Union[str, None]:
+        return self._context
 
     @property
     def ideal_value(self) -> Union[float, None]:
@@ -165,7 +165,7 @@ class Density(ABC):
 
 
 class KDE_integrate(Density):
-    def __init__(self, data: NDArray[Shape["*"], Float], ideal_value: float=None, dist_transform: DistTransform=DistTransform.NONE, transform_value: float=None, metric_id: MetricID=None, domain: str=None, **kwargs) -> None:
+    def __init__(self, data: NDArray[Shape["*"], Float], ideal_value: float=None, dist_transform: DistTransform=DistTransform.NONE, transform_value: float=None, qtype: str=None, context: str=None, **kwargs) -> None:
         self._kde = gaussian_kde(dataset=np.asarray(data))
         lb, ub = np.min(data), np.max(data)
         ext = np.max(data) - lb
@@ -182,11 +182,11 @@ class KDE_integrate(Density):
         
         ppf = cdf_to_ppf(cdf=np.vectorize(cdf), x=data, y_left=np.min(data), y_right=np.max(data))
 
-        super().__init__(range=(m_lb.x, m_ub.x), pdf=pdf, cdf=cdf, ppf=ppf, ideal_value=ideal_value, dist_transform=dist_transform, transform_value=transform_value, metric_id=metric_id, domain=domain, kwargs=kwargs)
+        super().__init__(range=(m_lb.x, m_ub.x), pdf=pdf, cdf=cdf, ppf=ppf, ideal_value=ideal_value, dist_transform=dist_transform, transform_value=transform_value, qtype=qtype, context=context, kwargs=kwargs)
 
 
 class KDE_approx(Density):
-    def __init__(self, data: NDArray[Shape["*"], Float], resample_samples: int=200_000, compute_ranges: bool=False, ideal_value: float=None, dist_transform: DistTransform=DistTransform.NONE, transform_value: float=None, metric_id: MetricID=None, domain: str=None, **kwargs) -> None:
[…]
