### 2023-02-07 d14c119
Remove old getters for QCC-specific properties and refactor ANOVA analysis
