+++ src/metrics_as_scores/distribution/distribution.py
@@ -466,83 +466,89 @@ class Dataset:
         return (transform_value, data)
 
 
-    @lru_cache(maxsize=None)
-    @staticmethod
-    def domains(include_all_domain: bool=True) -> list[str]:
-        systems_domains_df = pd.read_csv('./files/systems-domains.csv')
-        domains: list[str] = systems_domains_df.Domain.unique().tolist()
-        if include_all_domain:
-            domains.append('__ALL__')
-        return domains
+    def analyze_ANOVA(self, qtypes: Iterable[str], contexts: Iterable[str], unique_vals: bool=True) -> pd.DataFrame:
+        """
+        For each given type of quantity, this method performs an ANOVA across all
+        given contexts.
 
-
-    @lru_cache(maxsize=None)
-    @staticmethod
-    def systems_for_domain(domain: str) -> list[str]:
-        systems_domains_df = pd.read_csv('./files/systems-domains.csv')
-        systems_domains = dict(zip(systems_domains_df.System, systems_domains_df.Domain))
-        systems_qc_names = dict(zip(systems_domains_df.System, systems_domains_df.System_QC_name))
-
-        if domain == '__ALL__':
-            return list(systems_qc_names.values())
+        qtypes: `Iterable[str]`
+            An iterable of quantity types to conduct the analysis for. For each given
+            type, a separate analysis is performed and the result appended to the
+            returned data frame.
         
-        # Gather all systems with the selected domain.
-        temp = filter(lambda di: di[1] == domain, systems_domains.items())
-        # Map the names to the Qualitas compiled corpus:
-        return list(map(lambda di: systems_qc_names[di[0]], temp))
-    
+        contexts: `Iterable[str]`
+            An iterable of contexts across which each of the quantity types shall be
+            analyzed.
+        
+        unique_vals: `bool`
+            Passed to :meth:`self.data()`. If true, than small, random, and unique
+            noise is added to the data before it is analyzed. This will effectively
+            deduplicate any samples in the data (if any).
 
-    @lru_cache(maxsize=None)
-    @staticmethod
-    def domain_for_system(system: str, is_qc_name: bool=False) -> str:
-        systems_domains_df = pd.read_csv('./files/systems-domains.csv')
-        domain_dict = dict(zip(systems_domains_df[('System_QC_name' if is_qc_name else 'System')], systems_domains_df.Domain))
-        return domain_dict[system]
-    
+        :rtype: ``pd.DataFrame``
+        
+        :return: A data frame with the columns ``qtype`` (name of the quantity type),
+            ``stat`` (ANOVA test statistic), ``pval``, and ``across_contexts``, where
+            the latter is a semicolon-separated list of contexts the quantity type was
+            compared across.
+        """
 
-    @lru_cache(maxsize=None)
-    @staticmethod
-    def system_qc_to_system(system_qc: str) -> str:
-        systems_domains_df = pd.read_csv('./files/systems-domains.csv')
-        domain_dict = dict(zip(systems_domains_df.System_QC_name, systems_domains_df.System))
-        return domain_dict[system_qc]
-
-
-    def analyze_ANOVA(self, metric_ids: Iterable[MetricID], domains: Iterable[str], unique_vals: bool=True) -> pd.DataFrame:
         # We first have to build the data; f_oneway requires *args, where each
         # arg is a data series.
-        if len(list(metric_ids)) < 1 or len(list(domains)) < 2:
-            raise Exception('Requires one or metrics and two or more domains.')
+        if len(list(qtypes)) < 1 or len(list(contexts)) < 2:
[…]
