### 2023-02-07 8a4c7d1
Refactor class Dataset to be compatible with new, generic structure
