### 2023-02-08 72ae076
Remove metrics module from documentation
