### 2023-02-08 56bdbf1
Add FitParametric module to documentation
