### 2023-02-08 b7c74cb
Allow the user to cancel the main workflow
