### 2023-02-08 efae8c1
Register package paths in init of mas
