### 2023-02-08 8d6ac24
Clean up CLI main entry point
