+++ src/metrics_as_scores/cli/FitParametric.py
@@ -1,15 +1,23 @@
-from typing import Iterable, Union
-from Workflow import Workflow
+from typing import Any, Iterable
 from pathlib import Path
-from json import load
-from helpers import get_local_datasets
-from metrics_as_scores.distribution.distribution import LocalDataset
+from nptyping import Float, NDArray, Shape
 from scipy.stats._distn_infrastructure import rv_continuous, rv_discrete
+from metrics_as_scores.cli.helpers import get_local_datasets, isint
+from metrics_as_scores.cli.Workflow import Workflow
+from metrics_as_scores.distribution.distribution import DistTransform, LocalDataset, Dataset
 from metrics_as_scores.distribution.fitting import Fitter, FitterPymoo, Continuous_RVs, Discrete_Problems
+from metrics_as_scores.data.pregenerate_fit import get_data_tuple
+from metrics_as_scores.data.pregenerate_distns import generate_parametric_fits
 from questionary import Choice
-from rich.progress import Progress
-from json import dump
+from pickle import dump
+from os import cpu_count
+from joblib import Parallel, delayed
+from tqdm import tqdm
 import pandas as pd
+
+from scipy.stats._continuous_distns import norminvgauss_gen, gausshyper_gen, genhyperbolic_gen, geninvgauss_gen, invgauss_gen, studentized_range_gen
+from scipy.stats._discrete_distns import nhypergeom_gen, hypergeom_gen
+
 
 this_dir = Path(__file__).resolve().parent
 datasets_dir = this_dir.parent.parent.parent.joinpath('./datasets')
