### 2023-02-08 55caf0d
Finalize workflow for fitting parametric distributions
