+++ src/metrics_as_scores/cli/FitParametric.py
@@ -28,12 +28,16 @@ class FitParametricWorkflow(Workflow):
     def __init__(self) -> None:
         super().__init__()
         self.use_ds: LocalDataset = None
+        self.ds: Dataset = None
         self.df: pd.DataFrame = None
         self.fits_dir: Path = None
+        self.selected_rvs_c: list[type[rv_continuous]] = None
+        self.selected_rvs_d: list[type[rv_discrete]] = None
         self.use_fitter: type[Fitter] = FitterPymoo
+        self.num_cpus = 1
     
 
-    def _select_continuous_rvs(self) -> Iterable[rv_continuous]:
+    def _select_continuous_rvs(self) -> Iterable[type[rv_continuous]]:
         self.q.print('''
 You can now select the continuous random variables that you want to
 attempt to fit to the data. Select all in case you intend to re-
@@ -41,9 +45,10 @@ distribute and publicize your dataset. It is often not worth to de-
 select distributions unless you have a specific reason to do so,
 because fitting a continuous distribution is comparatively cheap and
 fast.'''.strip())
-        return self.q.checkbox(message='Select continuous random variables:', choices=[Choice(title=type(rv).__name__, value=rv, checked=True) for rv in Continuous_RVs]).ask()
+        recommend_ignore = [norminvgauss_gen, gausshyper_gen, genhyperbolic_gen, geninvgauss_gen, invgauss_gen, studentized_range_gen]
+        return self.q.checkbox(message='Select continuous random variables:', choices=[Choice(title=type(rv).__name__, value=type(rv), checked=not type(rv) in recommend_ignore) for rv in Continuous_RVs]).ask()
     
-    def _select_discrete_rvs(self) -> Iterable[rv_discrete]:
+    def _select_discrete_rvs(self) -> Iterable[type[rv_discrete]]:
         self.q.print('''
 You can now select the discrete random variables that you want to
 additionally attempt to fit to discrete quantity types. Select all
@@ -51,37 +56,93 @@ in case you intend to redistribute and publicize your dataset.
 While there are much fewer discrete random variables, their fitting
 is dramatically more computationally expensive to compute, since a
 global search has to be performed.'''.strip())
-        return self.q.checkbox(message='Select discrete random variables:', choices=[Choice(title=dp[0], value=dp[1], checked=True) for dp in Discrete_Problems.items()]).ask()
+        # Note how we will use the RV's type, not the problem's!
+        from scipy.stats import _discrete_distns
+        recommended_ignore = list(rv.__name__ for rv in [nhypergeom_gen, hypergeom_gen])
+        return self.q.checkbox(message='Select discrete random variables:', choices=[Choice(title=dp[0], value=getattr(_discrete_distns, dp[0]), checked=not dp[0] in recommended_ignore) for dp in Discrete_Problems.items()]).ask()
     
 
-    def _iterable_fits(self, rvs: Iterable[Union[rv_continuous, rv_discrete]]) -> Iterable[tuple[Path, str, str, Union[rv_continuous, rv_discrete]]]:
-        from itertools import product
-        for comb in product(self.use_ds['qtypes'].keys(), self.use_ds['contexts'], rvs):
-            fit_file = f'{comb[0]}_{comb[1]}_{"c" if issubclass(type(comb[2]), rv_continuous) else "d"}_{type(comb[2]).__name__}.json'
-            fit_path = self.fits_dir.joinpath(f'./{fit_file}')
-            if fit_path.exists():
-                self.print_info(text_normal='Fit already exists, skipping: ', text_vital=fit_file)
-            else:
-                yield (fit_path, comb[0], comb[1], comb[2])
[…]
