### 2023-02-08 a2d4ac7
Refactor generate_parametric_fits() to work with qtypes and contexts, and remove __name__==__main__
