+++ src/metrics_as_scores/data/pregenerate_distns.py
@@ -10,11 +10,9 @@ from metrics_as_scores.data.pregenerate_fit import fit, Fitter
 from sklearn.model_selection import ParameterGrid
 
 
-metrics_discrete_df = pd.read_csv('./files/metrics-discrete.csv')
-metrics_discrete = { x: y for (x, y) in zip(map(lambda m: MetricID[m], metrics_discrete_df.Metric), metrics_discrete_df.Discrete) }
 
 
-def generate_parametric_fits(dist_transform: DistTransform, data_dict: dict[str, NDArray[Shape["*"], Float]], transform_values_dict: dict[str, float], data_discrete_dict: dict[str, NDArray[Shape["*"], Float]], transform_values_discrete_dict: dict[str, float]) -> list[dict[str, Any]]:
+def generate_parametric_fits(ds: Dataset, num_jobs: int, fitter_type: type[Fitter], dist_transform: DistTransform, selected_rvs_c: list[type[rv_continuous]], selected_rvs_d: list[type[rv_discrete]], data_dict: dict[str, NDArray[Shape["*"], Float]], transform_values_dict: dict[str, float], data_discrete_dict: dict[str, NDArray[Shape["*"], Float]], transform_values_discrete_dict: dict[str, float]) -> list[dict[str, Any]]:
     # The thinking is this: To each data series we can always fit a continuous distribution,
     # whether it's discrete or continuous data. The same is not true the other way round, i.e.,
     # we should not fit a discrete distribution if the data is known to be continuous.
@@ -23,25 +21,34 @@ def generate_parametric_fits(dist_transform: DistTransform, data_dict: dict[str,
     # - Regardless of the data, always attempt to fit a continuous RV
     # - For all discrete data, also attempt to fit a discrete RV
     #
-    # That means that for discrete data, we will have to kinds of fitted RVs.
+    # That means that for discrete data, we will have two kinds of fitted RVs.
     # Also, when fitting a continuous RV to discrete data, we will add jitter to the data.
 
-    domains = Dataset.domains(include_all_domain=True)
-
-    from scipy.stats._continuous_distns import norminvgauss_gen, gausshyper_gen, genhyperbolic_gen, geninvgauss_gen, invgauss_gen, studentized_range_gen
-    from scipy.stats._discrete_distns import nhypergeom_gen, hypergeom_gen
-    ignored_dists = list([rv.__name__ for rv in [norminvgauss_gen, gausshyper_gen, genhyperbolic_gen, geninvgauss_gen, invgauss_gen, studentized_range_gen, nhypergeom_gen, hypergeom_gen]])
-
-    param_grid = { 'domain': domains, 'metric': list(map(lambda m: m.name, MetricID)), 'rv': list([x for x in Continuous_RVs_dict.keys() if x not in ignored_dists]), 'type': ['continuous'], 'dist_transform': [dist_transform] }
+    param_grid = {
+        'context': list(ds.contexts(include_all_domain=True)),
+        'qtype': ds.quantity_types, # Fit continuous to all
+        'rv': list([rv.__name__ for rv in selected_rvs_c]), # continuous here, discrete below
+        'type': ['continuous'],
+        'dist_transform': [dist_transform] }
     expanded_grid = pd.DataFrame(ParameterGrid(param_grid=param_grid))
-    param_grid = { 'domain': domains, 'metric': list([x[0].name for x in metrics_discrete.items() if x[1]]), 'rv': list([x for x in Discrete_RVs_dict.keys() if x not in ignored_dists]), 'type': ['discrete'], 'dist_transform': [dist_transform] }
-    expanded_grid = pd.concat([expanded_grid, pd.DataFrame(ParameterGrid(param_grid=param_grid))])
+
[…]
