### 2023-02-08 f373c74
Add Json type for a fitting result
