### 2023-02-08 a7276e0
Refactor the actual fitting function for parametric fits to work with generic qtypes and contexts
