+++ src/metrics_as_scores/data/pregenerate_fit.py
@@ -15,22 +15,52 @@ from timeit import default_timer as timer
 
 
 Continuous_RVs_dict: dict[str, type[rv_continuous]] = { x: y for (x, y) in zip(map(lambda rv: type(rv).__name__, Continuous_RVs), map(lambda rv: type(rv), Continuous_RVs)) }
+"""
+Dictionary of continuous random variables that are supported by `scipy.stats`.
+Note this is a dictionary of types, rather than instances.
+"""
 Discrete_RVs_dict: dict[str, type[rv_discrete]] = { x: y for (x, y) in zip(map(lambda rv: type(rv).__name__, Discrete_RVs), map(lambda rv: type(rv), Discrete_RVs)) }
+"""
+Dictionary of discrete random variables that are supported by `scipy.stats`.
+Note this is a dictionary of types, rather than instances.
+"""
 
 
 
-def get_data_tuple(dist: Dataset, metric_id: MetricID, dist_transform: DistTransform, continuous_transform: bool=True) -> list[tuple[str, NDArray[Shape["*"], Float]]]:
+def get_data_tuple(ds: Dataset, qtype: str, dist_transform: DistTransform, continuous_transform: bool=True) -> list[tuple[str, NDArray[Shape["*"], Float]]]:
+    """
+    This method is part of the workflow for computing parametric fits.
+    For a specific type of quantity and transform, it creates datasets
+    for all available contexts.
+
+    ds: ``Dataset``
+
+    qtype: ``str``
+        The type of quantity to get datasets for.
+    
+    dist_transform: ``DistTransform``
+        The chosen distribution transform.
+
+    continuous_transform: ``bool``
+        Whether the transform is real-valued or must be converted to integer.
+
+
+    :rtype: ``list[tuple[str, NDArray[Shape["*"], Float]]]``
+
+    :return: A list of tuples of three elements. The first element is a key
+        that identifies the context, the quantity type, and whether the data
+        was computed using unique values (see :meth:`Dataset.transform()`).
+    """
     l = []
-    for dom in Dataset.domains(include_all_domain=True):
+    for ctx in ds.contexts(include_all_domain=True):
         for unique_vals in [True, False]:
-            data = dist.data(metric_id=metric_id, domain=(None if dom == '__ALL__' else dom), unique_vals=unique_vals, sub_sample=25_000)
+            data = ds.data(qtype=qtype, context=(None if ctx == '__ALL__' else ctx), unique_vals=unique_vals, sub_sample=25_000)
             transform_value, data = Dataset.transform(data=data, dist_transform=dist_transform, continuous_value=continuous_transform)
-            key = f"{dom}_{metric_id.name}{('_u' if unique_vals else '')}"
+            key = f"{ctx}_{qtype}{('_u' if unique_vals else '')}"
             l.append((key, data, transform_value))
     return l
 
 
-def fit(grid_idx: int, row, metrics_discrete: dict[MetricID, bool], the_data: NDArray[Shape["*"], Float], the_data_unique: NDArray[Shape["*"], Float], transform_value: Union[float, None], dist_transform: DistTransform, write_temporary_results: bool=True) -> dict[str, Any]:
 class FitResult(TypedDict):
     grid_idx: int
     dist_transform: str
@@ -45,13 +75,15 @@ class FitResult(TypedDict):
 
     stat_tests: StatisticalTestJson
 
+
[…]
