### 2023-02-08 8cefb65
Add Json types for the statistical test
