### 2023-02-08 61fda86
Always add entry for a statistical test, even if it fails
