### 2023-02-08 b266dca
Fix how continuous and discrete random variables are found in scipy
