### 2023-02-08 66f436f
Add proper typing and some checks
