### 2023-02-08 1939e1a
Fix error in Anova analysis, some typings, and cosmetic changes
