### 2023-02-08 53aed78
Add method to perform statistical tests on own datasets and store those
