### 2023-02-08 5973897
Add description properties to LocalDataset and more getters to class Dataset
