### 2023-02-08 55c5e8d
Replace of usages 324 with 308 (powers of 10)
