### 2023-02-08 9f9bbd9
Clean up obsolete code from workflow
