### 2023-02-08 5147601
Refactor workflow for asking columns such that order is retained
