### 2023-02-08 f2c4b06
Completely refactor pre-generation of densities to work with generic qtype and context
