+++ src/metrics_as_scores/data/pregenerate.py
@@ -1,22 +1,11 @@
-from os import cpu_count
-from sys import path
 from pathlib import Path
-
-this_dir = Path(__file__).resolve().parent
-src_dir = this_dir.parent.parent
-path.append(str(src_dir.resolve()))
-
-
-
 import pandas as pd
-import numpy as np
 from typing import Any, Union
 from warnings import warn
 from os.path import exists
 from pickle import dump, load
 from joblib import Parallel, delayed
 from metrics_as_scores.tools.funcs import flatten_dict
-from metrics_as_scores.data.metrics import MetricID
 from metrics_as_scores.data.pregenerate_fit import Continuous_RVs_dict, Discrete_RVs_dict
 from metrics_as_scores.distribution.distribution import Dataset, Density, DistTransform, Dataset, Empirical, Empirical_discrete, KDE_approx, Parametric, Parametric_discrete
 from metrics_as_scores.distribution.fitting import StatisticalTest
@@ -26,43 +15,33 @@ from sklearn.model_selection import ParameterGrid
 
 
 def generate_densities(dataset: Dataset, clazz: type[Density]=Empirical, unique_vals: bool=None, resample_samples=250_000, dist_transform: DistTransform=DistTransform.NONE) -> dict[str, Density]:
-    metrics_ideal_df = pd.read_csv('./files/metrics-ideal.csv')
-    metrics_ideal_df.replace({ np.nan: None }, inplace=True)
-    metrics_ideal = { x: y for (x, y) in zip(map(lambda m: MetricID[m], metrics_ideal_df.Metric), metrics_ideal_df.Ideal) }
-
-    domains = Dataset.domains()
-    param_grid = { 'domain': domains, 'metric': list(map(lambda m: m.name, MetricID)) }
     contexts = list(dataset.contexts(include_all_contexts=True))
+    param_grid = { 'context': contexts, 'qtype': dataset.quantity_types }
     expanded_grid = pd.DataFrame(ParameterGrid(param_grid=param_grid))
 
     def get_density(grid_idx: int) -> tuple[str, Density]:
         row = expanded_grid.iloc[grid_idx,]
-        domain = row.domain
-        systems = Dataset.systems_for_domain(domain=domain)
-        metric_id = MetricID[row.metric]
+        context = row.context
+        qtype = row.qtype
         
         # Only use unique values if explicitly requested or CDF-type is ECDF!
         uvals = True if unique_vals else (clazz == Empirical)
-        data = dataset.data(metric_id=metric_id, unique_vals=uvals, systems=systems)
+        data = dataset.data(qtype=qtype, context=None if context == '__ALL__' else context, unique_vals=uvals)
 
         # Do transformation manually for other types of DensityFunc
         transform_value, data = Dataset.transform(data=data, dist_transform=dist_transform)
 
-        return (f'{domain}_{row.metric}', clazz(data=data, resample_samples=resample_samples, compute_ranges=True, ideal_value=metrics_ideal[metric_id], dist_transform=dist_transform, transform_value=transform_value, metric_id=metric_id, domain=domain))
+        return (f'{context}_{qtype}', clazz(data=data, resample_samples=resample_samples, compute_ranges=True, ideal_value=dataset.ideal_values[qtype], dist_transform=dist_transform, transform_value=transform_value, qtype=qtype, context=context))
 
     cdfs = Parallel(n_jobs=-1)(delayed(get_density)(i) for i in range(len(expanded_grid.index)))
     return dict(cdfs)
 
 
-def fits_to_MaS_densities(dataset: Dataset, distns_dict: dict[int, dict[str, Any]], dist_transform: DistTransform, use_continuous: bool) -> dict[str, Union[Parametric, Parametric_discrete]]:
[…]
