### 2023-02-08 d9c33bd
Remove __name__==__main__ as its now in a proper workflow
