### 2023-02-08 b0a2e1c
Allow specifying the arrow in print_info()
