### 2023-02-08 05fa2ff
Add pre-generation of densities to main workflow
