### 2023-02-08 3f6ab4f
Refactor and clean up web app to work with generic datasets
