+++ src/metrics_as_scores/webapp/main.py
@@ -1,21 +1,18 @@
-''' Present an interactive explorer for metrics as scores.
+''' Present an interactive explorer for Metrics As Scores.
 
 Use the ``bokeh serve`` command to run the example by executing:
 
-    bokeh serve scoring.py
+    ``bokeh serve src/metrics_as_scores/webapp/ dataset=qcc``
 
 at your command prompt. Then navigate to the URL
 
     http://localhost:5678/webapp
 
-in your browser.
+in your browser (the port might be different).
 
 '''
-from os import environ, getcwd
-from os.path import join
-from sys import path
-from pathlib import Path
-path.append(f'{Path(join(getcwd(), "src")).resolve()}')
+
+from os import environ
 
 if 'BOKEH_VS_DEBUG' in environ and environ['BOKEH_VS_DEBUG'] == 'true':
     import ptvsd
@@ -27,52 +24,50 @@ if 'BOKEH_VS_DEBUG' in environ and environ['BOKEH_VS_DEBUG'] == 'true':
 
 import pandas as pd
 import numpy as np
-from metrics_as_scores.webapp.exception import PlotException
+from pathlib import Path
 from typing import Union
-from metrics_as_scores.data.metrics import MetricID
+from metrics_as_scores.webapp.exception import PlotException
 from metrics_as_scores.tools.funcs import nonlinspace
-from metrics_as_scores.distribution.distribution import Empirical, DistTransform, Empirical_discrete, KDE_approx, Parametric, Parametric_discrete
+from metrics_as_scores.distribution.distribution import Empirical, DistTransform, Empirical_discrete, KDE_approx, Parametric, Parametric_discrete, Dataset
 from collections import OrderedDict
 from bokeh.events import MenuItemClick
 from bokeh.io import curdoc
 from bokeh.layouts import column, row
 from bokeh.models import ColumnDataSource, Dropdown, CheckboxGroup, Button, DataTable, TableColumn
 from bokeh.models.widgets.inputs import Spinner
+# TODO: Use best palette
 from bokeh.palettes import Category20_12
 from bokeh.plotting import figure
 from bokeh.models.widgets.markups import Div
 from numbers import Integral
 
 
+this_dir = Path(__file__).resolve().parent
+
+# Import the data; Note this is process-static!
+from metrics_as_scores.webapp import data
+ds: Dataset = data.ds
+cdfs = data.cdfs
+contexts = list(ds.contexts(include_all_contexts=True))
+
 def cap(s: str) -> str:
     return s[:1].upper() + s[1:]
 
 
 
-# Data and UI control for the systems' domains:
-systems_domains_df = pd.read_csv('./files/systems-domains.csv')
-systems_domains = dict(zip(systems_domains_df.System, systems_domains_df.Domain))
-systems_qc_names = dict(zip(systems_domains_df.System, systems_domains_df.System_QC_name))
-temp = systems_domains_df.groupby('Domain').count()
-domains = OrderedDict(zip(temp.index.to_list(), temp['System'].to_list()))
-domains_labels = OrderedDict({ key: f'{cap(key)} ({val})' for (key, val) in domains.items() })
-domains['__ALL__'] = int(temp.System.sum())
-domains_labels['__ALL__'] = '[All domain types combined]'
+# Data and UI for the qtype/score types
+dd_scores_items: list[tuple[str, str]] = []
+if len(ds.quantity_types_discrete) > 0:
+    dd_scores_items.append(('------------ Discrete (integral) Quantities', '-'))
+    dd_scores_items += list([(f'[{qtype}] {ds.qytpe_desc(qtype=qtype)}', qtype) for qtype in ds.quantity_types_discrete])
+if len(ds.quantity_types_continuous) > 0:
[…]
