### 2023-02-08 787566e
Also create a 'web' directory for new datasets
