### 2023-02-08 df53cba
Export the dataset directory as well
