### 2023-02-08 2f638ab
Load about and references from current dataset and embed them into header and footer
