### 2023-02-08 63259f3
Change the used color palette to the maximum possible using 20 colors
