### 2023-02-08 493b456
Commit previously forgotten and completely implemented workflow for generating densities
