+++ src/metrics_as_scores/cli/GenerateDensities.py
@@ -1,0 +1,99 @@
+import pandas as pd
+from pathlib import Path
+from os import cpu_count
+from metrics_as_scores.cli.Workflow import Workflow
+from metrics_as_scores.cli.helpers import isint, get_local_datasets
+from metrics_as_scores.distribution.distribution import Parametric, Parametric_discrete, DistTransform, Dataset, LocalDataset, Empirical, Empirical_discrete, KDE_approx
+from metrics_as_scores.data.pregenerate import generate_parametric, generate_empirical, generate_empirical_discrete
+from sklearn.model_selection import ParameterGrid
+from joblib import Parallel, delayed
+
+
+this_dir = Path(__file__).resolve().parent
+datasets_dir = this_dir.parent.parent.parent.joinpath('./datasets')
+
+
+class GenerateDensitiesWorkflow(Workflow):
+    def __init__(self) -> None:
+        super().__init__()
+        self.use_ds: LocalDataset = None
+        self.ds: Dataset = None
+        self.fits_dir: Path = None
+        self.densities_dir: Path = None
+        self.num_cpus = 1
+    
+
+    def _generate_parametric(self) -> None:
+        grid = dict(
+            clazz = [Parametric, Parametric_discrete],
+            transform = list(DistTransform))
+        expanded_grid = pd.DataFrame(ParameterGrid(param_grid=grid))
+        Parallel(n_jobs=min(self.num_cpus, len(expanded_grid.index)))(delayed(generate_parametric)(self.ds, self.densities_dir, self.fits_dir, expanded_grid.iloc[i,]['clazz'], expanded_grid.iloc[i,]['transform']) for i in range(len(expanded_grid.index)))
+    
+    def _generate_empirical_kde(self) -> None:
+        grid = dict(
+            clazz = [Empirical, KDE_approx],
+            transform = list(DistTransform))
+        expanded_grid = pd.DataFrame(ParameterGrid(param_grid=grid))
+        Parallel(n_jobs=min(self.num_cpus, len(expanded_grid.index)))(delayed(generate_empirical)(self.ds, self.densities_dir, expanded_grid.iloc[i,]['clazz'], expanded_grid.iloc[i,]['transform']) for i in range(len(expanded_grid.index)))
+    
+    def _generate_empirical_discrete(self) -> None:
+        grid = dict(
+            clazz = [Empirical_discrete],
+            transform = list(DistTransform))
+        expanded_grid = pd.DataFrame(ParameterGrid(param_grid=grid))
+        Parallel(n_jobs=min(self.num_cpus, len(expanded_grid.index)))(delayed(generate_empirical_discrete)(self.ds, self.densities_dir, expanded_grid.iloc[i,]['transform']) for i in range(len(expanded_grid.index)))
+
+    
+
+    def pre_generate(self) -> None:
+        self.q.print('\n' + 10*'-')
+        self.q.print('''
+This workflow generates density-related functions that are used by the
+Web Application. While those can be large, generating them on-the-fly is
+usually not possible in acceptable time. Using pre-generated functions
+is a trade-off between space and user experience, where we sacrifice the
+former as it is cheaper.
+
+For each quantity and each context, we pre-generate functions for the
+probability density (PDF), the cumulative distribution (CDF) and its
+complement (CCDF), as well as the quantile (or percent point) function
+(PPF). So for one quantity and one context, we pre-generate one density
+that unites those four functions.
+
+There are 5 primary classes of densities: Parametric, Parametric_discrete,
+Empirical, Empirical_discrete, and KDE_approx. Please refer to the
+documentation for details about these. Generating parametric densities
+uses the computed fits from another workflow, is cheap, fast, and does
[…]
