### 2023-02-09 6ff8378
Fix a couple of smaller bugs
