### 2023-02-09 2de6184
Remove obsolete code and make some cosmetical changes
