+++ datasets/_default/readme.md
@@ -1,0 +1,7 @@
+# Dataset Skeleton
+
+This directory contains templates that are copied into the directories of newly created datasets.
+The purpose is to give you something extendable to start with.
+Some of the files, like `about.html` and `references.html` are even required.
+There is also a document template for `Quarto` that you can render once you finished creating your dataset.
+Once rendered, it generates the required `About.pdf` that you must include in your dataset, in case you intend to publicly share it.
