### 2023-02-09 e1ce483
Revert "Remove now-obsolete method Dataset::analyze_distr()"
This reverts commit 94462c5a035fcb6ab0b94308bda134344267d0a5.
