### 2023-02-10 365f142
Make method Dataset::data() properly handle non- or __ALL__-contexts
