### 2023-02-10 fc6b93a
Adapt implementation of KS2 and Welch-t tests to generic qtype and context
