+++ src/metrics_as_scores/cli/CreateDataset.py
@@ -210,7 +210,7 @@ cannot be resumed.
         dataset_dir = datasets_dir.joinpath(f'./{manifest["id"]}')
         fits_dir = dataset_dir.joinpath('./fits')
         densities_dir = dataset_dir.joinpath('./densities')
-        tests_dir = dataset_dir.joinpath('./tests')
+        tests_dir = dataset_dir.joinpath('./stat-tests')
         web_dir = dataset_dir.joinpath('./web')
         for dir in [dataset_dir, fits_dir, densities_dir, tests_dir, web_dir]:
             if not dir.exists():
