+++ datasets/_default/refs.bib
@@ -1,0 +1,51 @@
+@INCOLLECTION{Chambers2017-fu,
+	title     = "Statistical Models",
+	booktitle = "Statistical Models in {S}",
+	author    = "Chambers, John M and Hastie, Trevor J",
+	publisher = "Routledge",
+	pages     = "13--44",
+	month     =  nov,
+	year      =  2017,
+	doi       = {10.1201/9780203738535},
+	isbn      = {9780203738535},
+	isbn13    = {9780203738535}
+}
+
+@article{tukey1949hsd,
+	ISSN      = {0006341X, 15410420},
+	URL       = {http://www.jstor.org/stable/3001913},
+	author    = {John W. Tukey},
+	journal   = {Biometrics},
+	number    = {2},
+	pages     = {99--114},
+	publisher = {[Wiley, International Biometric Society]},
+	title     = {Comparing Individual Means in the Analysis of Variance},
+	urldate   = {2022-09-05},
+	volume    = {5},
+	year      = {1949}
+}
+
+@inproceedings{honel2022qrs,
+  author    = {Sebastian Hönel and Morgan Ericsson and Welf Löwe and Anna wingkvist},
+  title     = {Contextual Operationalization of Metrics As Scores: Is My Metric Value Good?},
+  booktitle = {22nd {IEEE} International Conference on Software Quality, Reliability and Security, {QRS} 2022, Guangzhou, China, December 5-9, 2022},
+  pages     = {333--343},
+  publisher = {{IEEE}},
+  year      = {2022},
+  url       = {https://doi.org/10.1109/QRS57517.2022.00042},
+  doi       = {10.1109/QRS57517.2022.00042}
+}
+
+@article{stephens1974kstest,
+    author = { M. A.   Stephens },
+    title = {{EDF} Statistics for Goodness of Fit and Some Comparisons},
+    journal = {Journal of the American Statistical Association},
+    volume = {69},
+    number = {347},
+    pages = {730-737},
+    year  = {1974},
+    publisher = {Taylor \& Francis},
+    doi = {10.1080/01621459.1974.10480196},
+    URL = {https://www.tandfonline.com/doi/abs/10.1080/01621459.1974.10480196},
+    eprint = {https://www.tandfonline.com/doi/pdf/10.1080/01621459.1974.10480196}
+}
