### 2023-02-10 dce8741
Number sections by default in About template
