### 2023-02-10 a0006e5
Add generating results for ANOVA to About template
