### 2023-02-10 a260f14
Install tabulate (needed for report generation in quarto)
