+++ pyproject.toml
@@ -43,6 +43,7 @@ sphinx-rtd-theme = "^1.1.1"
 toml = "^0.10.2"
 rich = "^13.3.1"
 questionary = "^1.10.0"
+tabulate = "^0.9.0"
 
 
 [tool.poetry.group.dev.dependencies]
