### 2023-02-10 d7c2290
Prematurely upgrade project to v2.0.0 and add section about it to readme
