### 2023-02-11 dfd2c1f
Finish implementing showing some common/extreme values
