### 2023-02-11 847046a
Implement TukeyHSD report generation in default template
