+++ datasets/_default/refs.bib
@@ -26,7 +26,7 @@
 }
 
 @inproceedings{honel2022qrs,
-  author    = {Sebastian Hönel and Morgan Ericsson and Welf Löwe and Anna wingkvist},
+  author    = {Sebastian Hönel and Morgan Ericsson and Welf Löwe and Anna Wingkvist},
   title     = {Contextual Operationalization of Metrics As Scores: Is My Metric Value Good?},
   booktitle = {22nd {IEEE} International Conference on Software Quality, Reliability and Security, {QRS} 2022, Guangzhou, China, December 5-9, 2022},
   pages     = {333--343},
