### 2023-02-12 3aa37d0
Limit the number of samples when performing the exact KS2 test
