### 2023-02-12 85c45a2
Properly install scikit-learn and remove sklearn
