### 2023-02-12 b176cee
Add function for natural string sorting
