### 2023-02-12 3d7b2ba
Sort qtype and contexts; also cache results from example values
