+++ datasets/_default/About.qmd
@@ -48,21 +48,30 @@ dataset = Dataset(ds=manifest, df=pd.DataFrame())
 ```
 
 
+# Description
+
+<!-- You can put more description here, into the following diplayed markup, or below. -->
+
 ```{python}
 #| echo: false
 
 dm(f"""
-# Description
 {manifest['desc']}
 """)
 ```
+
 
 This dataset has the following **Quantity Types**:
 
 ```{python}
 #| echo: false
 
-temp = [f'- *{qtype}*: {dataset.qytpe_desc(qtype=qtype)}' for qtype in dataset.quantity_types]
+from metrics_as_scores.tools.funcs import natsort
+
+qtypes = dataset.quantity_types
+qtypes.sort(key=natsort)
+
+temp = [f'- *{qtype}*: {dataset.qytpe_desc(qtype=qtype)}{" [discrete]" if dataset.is_qtype_discrete(qtype=qtype) else ""}' for qtype in qtypes]
 
 dm('\n'.join(temp))
 ```
@@ -71,6 +80,7 @@ dm('\n'.join(temp))
 #| echo: false
 
 contexts = list(dataset.contexts(include_all_contexts=False))
+contexts.sort(key=natsort)
 
 dm(f'''
 It has a total of {len(contexts)} **Contexts**: {format_comma_and(list([f'*{c}*' for c in contexts]))}.''')
@@ -138,6 +148,7 @@ Therefore, the ANOVA test also indicates whether quantity types' values are diff
 #| echo: false
 
 contexts = list(dataset.contexts(include_all_contexts=False))
+contexts.sort(key=natsort)
 row = df_anova.iloc[0,:].to_dict()
 
 
@@ -204,21 +215,20 @@ from math import ceil
 
 temp = df_ks2samp[df_ks2samp.pval >= use_alpha]
 cnts = list([
-	temp[(temp.qtype == qtype)].shape[0]
-	for qtype in dataset.quantity_types
+	temp[(temp.qtype == qtype)].shape[0] for qtype in qtypes
 ])
 
 fig, ax = plt.subplots()
 fig.set_figwidth(8)
 fig.set_figheight(2.5)
-bars = ax.bar(dataset.quantity_types, cnts, width=.5, color=Category20_20)
+bars = ax.bar(qtypes, cnts, width=.5, color=Category20_20)
 
 ax.set_ylim(bottom=0, top=r + ceil(r*.1))
 ax.bar_label(bars)
 ax.set_axisbelow(True)
 ax.grid(color='#dddddd', linestyle='dashed')
 angle = 0 if len(contexts) < 10 else (45 if len(contexts) < 20 else 90)
-ax.set_xticklabels(ax.get_xticklabels(), rotation=angle)
+ax.set_xticklabels(ax.get_xticklabels(), rotation=angle, ha='right', rotation_mode='anchor')
 
 plt.rcParams.update({
     "text.usetex": True,
@@ -237,7 +247,7 @@ plt.show()
 #| echo: false
 
 n = len(contexts) + 1
-m = len(dataset.quantity_types)
+m = len(qtypes)
 r = (n-1) * m
 
 dm(f'''
@@ -246,7 +256,7 @@ Consider the set of contexts $\\mathit{{C}}$ and the set of quantity types $\\ma
 Given a context $c_i$ and a quantity type $q_j$, this test reports all other contexts $\\mathit{{C}}^{{\\setminus c_i}}$ that have a statistically significantly different distribution for $q_j$.
 However, for $c_i$, we aggregate these counts across all quantity types.
 We have a total of $n\\tight{{=}}{n-1}+1$ contexts (including the virtual context).
[…]
