### 2023-02-12 78f9a43
Show progress when pre-generating densities
