### 2023-02-12 da87c4a
Fix a problem where sometimes the parameter 'lambda' would end with an underscore
