### 2023-02-12 7bdf842
Add size and size_extracted to known datasets
