### 2023-02-12 f618596
Add completely new workflow for bundling and zipping a dataset
