+++ src/metrics_as_scores/cli/BundleOwn.py
@@ -1,0 +1,130 @@
+import pandas as pd
+from os.path import commonpath
+from metrics_as_scores.cli.Workflow import Workflow
+from metrics_as_scores.cli.helpers import get_local_datasets
+from metrics_as_scores.distribution.distribution import DistTransform, Parametric, Parametric_discrete, Empirical, Empirical_discrete, KDE_approx, Dataset, LocalDataset
+from itertools import product
+from pathlib import Path
+from zipfile import ZipFile, ZIP_DEFLATED
+
+
+this_dir = Path(__file__).resolve().parent
+datasets_dir = this_dir.parent.parent.parent.joinpath('./datasets')
+
+
+class BundleDatasetWorkflow(Workflow):
+    def __init__(self) -> None:
+        super().__init__()
+        self.use_ds: LocalDataset = None
+        self.ds: Dataset = None
+        self.ds_dir: Path = None
+
+        self.dirs_required: list[Path] = None
+        self.files_required: list[Path] = None
+    
+
+    def _check_files(self) -> None:
+        for d in self.dirs_required:
+            self.print_info(text_normal='Checking directory: ', text_vital=str(d))
+            if not d.exists():
+                raise Exception(f'Directory does not exist: {str(d)}')
+            elif not d.is_dir():
+                raise Exception(f'Not a directory: {str(d)}')
+        
+        for f in self.files_required:
+            self.print_info(text_normal='Checking file: ', text_vital=str(f))
+            if not f.exists():
+                raise Exception(f'File does not exist: {str(f)}')
+            elif not f.is_file():
+                raise Exception(f'Not a file: {str(f)}')
+    
+
+    def _make_zip(self) -> Path:
+        zip_file = self.ds_dir.joinpath('./dataset.zip')
+        with ZipFile(file=str(zip_file), mode='x', compression=ZIP_DEFLATED, compresslevel=9) as zf:
+            for path in self.files_required:
+                fs_path = str(path.resolve())
+                base = commonpath([fs_path, str(self.ds_dir.resolve())])
+                zf.write(filename=fs_path, arcname=fs_path.removeprefix(base))
+        return zip_file
+    
+
+    def bundle(self) -> None:
+        self.q.print('\n' + 10*'-')
+        self.q.print('''
+This workflow bundles a manually created dataset into a single Zip file
+that can be uploaded to, e.g., Zenodo, and registered with Metrics As
+Scores in order to make it available to others as a known dataset.
+
+In order for an own dataset to be publishable, it needs to have all
+parametric fits, generated densities, references, about, etc. This
+workflow will check that all requirements are fulfilled.
+
+For an example dataset, check out https://doi.org/10.5281/zenodo.TODO
+'''.strip())
+        self.q.print('')
+        self.q.print(10*'-')
+        
+
+        datasets = list(get_local_datasets())
+        self.use_ds = self.askt(
+            prompt='Select the local dataset you want to pre-generate densities for:',
+            options=list([(f'{ds["name"]} [{ds["id"]}] by {", ".join(ds["author"])}', ds) for ds in datasets]))
+        self.ds_dir = datasets_dir.joinpath(f'./{self.use_ds["id"]}')
+        self.ds = Dataset(ds=self.use_ds, df=pd.read_csv(str(self.ds_dir.joinpath('./org-data.csv'))))
+
+        dir_densities = self.ds_dir.joinpath(f'./densities')
+        dir_fits = self.ds_dir.joinpath(f'./fits')
+        dir_stests = self.ds_dir.joinpath(f'./stat-tests')
+        dir_web = self.ds_dir.joinpath(f'./web')
+
+        self.dirs_required = [
+            dir_densities,
+            dir_fits,
+            dir_stests,
+            dir_web
+        ]
+
+        self.files_required = list([
[…]
