### 2023-02-12 155298d
Add methods for getting known datasets (local remote) and formatting file size
