### 2023-02-13 55be895
Add new workflow to show locally available datasets
