+++ src/metrics_as_scores/cli/Download.py
@@ -1,5 +1,7 @@
 from metrics_as_scores.cli.Workflow import Workflow
 from metrics_as_scores.cli.helpers import get_known_datasets, get_local_datasets, KNOWN_DATASETS_FILE
+from requests import get
+from shutil import copyfileobj, unpack_archive
 from pathlib import Path
 
 
@@ -35,6 +37,19 @@ Known datasets are loaded from: {KNOWN_DATASETS_FILE}
             return
         
         use_ds = known_ds[id]
+        dataset_dir = datasets_dir.joinpath(f'./{use_ds["id"]}')
+        dataset_dir.mkdir(exist_ok=False)
+
+        self.print_info(text_normal='Downloading archive from: ', text_vital=use_ds['download'])
+        zip_file = dataset_dir.joinpath('./dataset.zip')
+        with get(url=use_ds['download'], stream=True) as resp:
+            with open(file=str(zip_file), mode='wb') as fp:
+                copyfileobj(fsrc=resp.raw, fdst=fp)
+        self.q.print('Download complete. Extracting ...')
+
+        unpack_archive(filename=str(zip_file), extract_dir=str(dataset_dir))
+        self.q.print('Done! You can now use this dataset!')
+        zip_file.unlink()
 
     
 
