### 2023-02-13 7ef886b
Update configuration for docs and remove webapp module
