+++ docs/metrics_as_scores.webapp.rst
@@ -1,45 +1,0 @@
-metrics\_as\_scores.webapp package
-==================================
-
-Submodules
-----------
-
-metrics\_as\_scores.webapp.app\_hooks module
---------------------------------------------
-
-.. automodule:: metrics_as_scores.webapp.app_hooks
-   :members:
-   :undoc-members:
-   :show-inheritance:
-
-metrics\_as\_scores.webapp.data module
---------------------------------------
-
-.. automodule:: metrics_as_scores.webapp.data
-   :members:
-   :undoc-members:
-   :show-inheritance:
-
-metrics\_as\_scores.webapp.exception module
--------------------------------------------
-
-.. automodule:: metrics_as_scores.webapp.exception
-   :members:
-   :undoc-members:
-   :show-inheritance:
-
-metrics\_as\_scores.webapp.main module
---------------------------------------
-
-.. automodule:: metrics_as_scores.webapp.main
-   :members:
-   :undoc-members:
-   :show-inheritance:
-
-Module contents
----------------
-
-.. automodule:: metrics_as_scores.webapp
-   :members:
-   :undoc-members:
-   :show-inheritance:
