### 2023-02-13 2bf0b96
Create build-doc.yml action for documentation automation
