+++ .github/workflows/build-doc.yml
@@ -6,34 +6,16 @@ on:
       - 'master'
 
 jobs:
-  docs:
+  build:
+    
     runs-on: ubuntu-latest
+
     steps:
-    - uses: actions/checkout@v1
-    - uses: mrshoenel/metrics-as-scores@master
+    - uses: actions/setup-python@v4.5.0
       with:
-        build-command: sphinx-apidoc -o docs src/metrics_as_scores "**/webapp/*" --force
-        docs-folder: "docs/"
-    - uses: actions/upload-artifact@v1
-      with:
-        name: DocumentationHTML
-        path: docs/_build/html/
-    - name: Commit documentation changes
-      run: |
-        git clone https://github.com/mrshoenel/metrics-as-scores.git --branch gh-pages --single-branch gh-pages
-        cp -r docs/_build/html/* gh-pages/
-        cd gh-pages
-        touch .nojekyll
-        git config --local user.email "action@github.com"
-        git config --local user.name "GitHub Action"
-        git add .
-        git commit -m "Update documentation" -a || true
-        # The above command will fail if no changes were present, so we ignore
-        # that.
-    - name: Push changes
-      uses: ad-m/github-push-action@master
-      with:
-        branch: gh-pages
-        directory: gh-pages
-        github_token: ${{ secrets.GITHUB_TOKEN }}
-    # ===============================
+        python-version: '3.10'
+    - run: pip install poetry
+    - run: poetry install
+    - run: sphinx-apidoc -o docs src\metrics_as_scores "**/webapp/*" --force
+    - run: docs/make html
+
