+++ .github/workflows/build-doc.yml
@@ -11,6 +11,7 @@ jobs:
     runs-on: ubuntu-latest
 
     steps:
+    - uses: actions/checkout@v3
     - uses: actions/setup-python@v4.5.0
       with:
         python-version: '3.10'
