+++ .github/workflows/build-doc.yml
@@ -17,6 +17,6 @@ jobs:
         python-version: '3.10'
     - run: pip install poetry
     - run: poetry install
-    - run: sphinx-apidoc -o docs src\metrics_as_scores "**/webapp/*" --force
+    - run: sphinx-apidoc -o docs src/metrics_as_scores "**/webapp/*" --force
     - run: docs/make html
 
