### 2023-02-13 583d785
Update build-doc with all manual steps
