+++ .github/workflows/build-doc.yml
@@ -11,12 +11,26 @@ jobs:
     runs-on: ubuntu-latest
 
     steps:
-    - uses: actions/checkout@v3
     - uses: actions/setup-python@v4.5.0
       with:
         python-version: '3.10'
+    - run: apt-get update -y && apt-get install python3.10-venv python3.10-dev
+    - run: git clone https://github.com/MrShoenel/metrics-as-scores.git
+    - run: cd metrics-as-scores
+    - run: python3.10 -m venv venv
+    - run: source venv/bin/activate
     - run: pip install poetry
     - run: poetry install
     - run: sphinx-apidoc -o docs src/metrics_as_scores "**/webapp/*" --force
-    - run: docs/make html
+    - run: sphinx-build docs/ docs/_build
 
+    - name: Commit documentation changes
+      run: |
+        mkdir gh-pages
+        cp -r docs/_build/html/* gh-pages/
+        cd gh-pages
+        touch .nojekyll
+        git config --local user.email "action@github.com"
+        git config --local user.name "GitHub Action"
+        git add .
+        git commit -m "Update documentation" -a || true
