### 2023-02-13 a768298
Add sudo for apt-get commands
