+++ .github/workflows/build-doc.yml
@@ -14,7 +14,7 @@ jobs:
     - uses: actions/setup-python@v4.5.0
       with:
         python-version: '3.10'
-    - run: apt-get update -y && apt-get install python3.10-venv python3.10-dev
+    - run: sudo apt-get update -y && sudo apt-get install python3.10-venv python3.10-dev
     - run: git clone https://github.com/MrShoenel/metrics-as-scores.git
     - run: cd metrics-as-scores
     - run: python3.10 -m venv venv
