### 2023-02-13 1a44918
Use checkout action instead of manual cloning
