+++ .github/workflows/build-doc.yml
@@ -14,15 +14,25 @@ jobs:
     - uses: actions/setup-python@v4.5.0
       with:
         python-version: '3.10'
-    - run: sudo apt-get update -y && sudo apt-get install python3.10-venv python3.10-dev
-    - run: git clone https://github.com/MrShoenel/metrics-as-scores.git
-    - run: cd metrics-as-scores
-    - run: python3.10 -m venv venv
-    - run: source venv/bin/activate
-    - run: pip install poetry
-    - run: poetry install
-    - run: sphinx-apidoc -o docs src/metrics_as_scores "**/webapp/*" --force
-    - run: sphinx-build docs/ docs/_build
+
+    - name: Install deps
+      run: |
+        sudo apt-get update -y
+        sudo apt-get install -y python3.10-venv python3.10-dev
+    
+    - uses: actions/checkout@v3
+
+    - name: Prepare venv and install project
+      run: |
+        python3.10 -m venv venv
+        source venv/bin/activate
+        pip install poetry
+        poetry install
+    
+    - name: Build api-doc
+      run: |
+        sphinx-apidoc -o docs src/metrics_as_scores "**/webapp/*" --force
+        sphinx-build docs/ docs/_build
 
     - name: Commit documentation changes
       run: |
