+++ .github/workflows/build-doc.yml
@@ -5,44 +5,13 @@ on:
     branches:
       - 'master'
 
+
 jobs:
   build:
-    
+    name: Push Sphinx Pages
     runs-on: ubuntu-latest
-
     steps:
-    - uses: actions/setup-python@v4.5.0
+    - uses: seanzhengw/sphinx-pages@master
       with:
-        python-version: '3.10'
-
-    - name: Install deps
-      run: |
-        sudo apt-get update -y
-        sudo apt-get install -y python3.10-venv python3.10-dev
-    
-    - uses: actions/checkout@v3
-
-    - name: Prepare venv and install project
-      run: |
-        python3.10 -m venv venv
-        source venv/bin/activate
-        pip install poetry
-        poetry install
-    
-    - name: Build api-doc
-      run: |
-        pwd
-        ll
-        ./venv/bin/sphinx-apidoc -o docs src/metrics_as_scores "**/webapp/*" --force
-        ./venv/bin/sphinx-build docs/ docs/_build
-
-    - name: Commit documentation changes
-      run: |
-        mkdir gh-pages
-        cp -r docs/_build/html/* gh-pages/
-        cd gh-pages
-        touch .nojekyll
-        git config --local user.email "action@github.com"
-        git config --local user.name "GitHub Action"
-        git add .
-        git commit -m "Update documentation" -a || true
+        github_token: ${{ secrets.GITHUB_TOKEN }}
+        create_readme: true
