+++ .github/workflows/build-doc.yml
@@ -17,9 +17,9 @@ jobs:
         uses: actions/checkout@v2.3.4
 
       - name: Set up Python
-        uses: actions/setup-python@v2.2.1
+        uses: actions/setup-python@v4.5.0
         with:
-          python-version: 3.10
+          python-version: '3.10'
  
       - name: Install dependencies
         run: |
