+++ .github/workflows/build-doc.yml
@@ -40,13 +40,13 @@ jobs:
           git init
           touch .nojekyll
           git add -A
+          git config --global --add safe.directory /github/workspace/docs/_build/html
           git config --local user.email "action@github.com"
           git config --local user.name "GitHub Action"
           git commit -m 'deploy'
       - name: Force push to destination branch
-        uses: ad-m/github-push-action@v0.5.0
+        uses: ad-m/github-push-action@master
         with:
           github_token: ${{ secrets.GITHUB_TOKEN }}
           branch: gh-pages
-          force: true
           directory: ./docs/_build/html
