+++ .github/workflows/build-doc.yml
@@ -34,19 +34,29 @@ jobs:
           sphinx-apidoc -o docs src/metrics_as_scores --force
           make -C docs html
       
-      - name: Init new repo in dist folder and commit generated files
+      - uses: actions/upload-artifact@v1
+        with:
+          name: DocumentationHTML
+          path: docs/_build/html/
+      
+      # Publish built docs to gh-pages branch.
+      # ===============================
+      - name: Commit documentation changes
         run: |
-          cd docs/_build/html/
-          git init
+          git clone https://github.com/MrShoenel/metrics-as-scores.git --branch gh-pages --single-branch gh-pages
+          cp -r docs/_build/html/* gh-pages/
+          cd gh-pages
           touch .nojekyll
-          git add -A
-          git config --global --add safe.directory /github/workspace/docs/_build/html
           git config --local user.email "action@github.com"
           git config --local user.name "GitHub Action"
-          git commit -m 'deploy'
-      - name: Force push to destination branch
+          git add .
+          git commit -m "Update documentation" -a || true
+          # The above command will fail if no changes were present, so we ignore
+          # that.
+      - name: Push changes
         uses: ad-m/github-push-action@master
         with:
-          github_token: ${{ secrets.GITHUB_TOKEN }}
           branch: gh-pages
-          directory: ./docs/_build/html
+          directory: gh-pages
+          github_token: ${{ secrets.GITHUB_TOKEN }}
+      # ===============================
