+++ .github/workflows/build-doc.yml
@@ -44,6 +44,7 @@ jobs:
       - name: Commit documentation changes
         run: |
           git clone https://github.com/MrShoenel/metrics-as-scores.git --branch gh-pages --single-branch gh-pages
+          mkdir gh-pages
           cp -r docs/_build/html/* gh-pages/
           cd gh-pages
           touch .nojekyll
