+++ .github/workflows/build-doc.yml
@@ -43,7 +43,8 @@ jobs:
       # ===============================
       - name: Commit documentation changes
         run: |
-          git clone https://github.com/MrShoenel/metrics-as-scores.git --branch gh-pages --single-branch gh-pages
+          git clone https://github.com/MrShoenel/metrics-as-scores.git --branch gh-pages
+          cd metrics-as-scores
           mkdir gh-pages
           cp -r docs/_build/html/* gh-pages/
           cd gh-pages
