+++ .github/workflows/build-doc.yml
@@ -43,11 +43,10 @@ jobs:
       # ===============================
       - name: Commit documentation changes
         run: |
-          git clone https://github.com/MrShoenel/metrics-as-scores.git --branch gh-pages
-          cd metrics-as-scores
-          mkdir gh-pages
-          cp -r docs/_build/html/* gh-pages/
-          cd gh-pages
+          git clone https://github.com/MrShoenel/metrics-as-scores.git --branch gh-pages --single-branch temp
+          mkdir temp/gh-pages
+          cp -r docs/_build/html/* temp/gh-pages/
+          cd temp/gh-pages
           touch .nojekyll
           git config --local user.email "action@github.com"
           git config --local user.name "GitHub Action"
