+++ .github/workflows/build-doc.yml
@@ -58,6 +58,6 @@ jobs:
         uses: ad-m/github-push-action@master
         with:
           branch: gh-pages
-          directory: gh-pages
+          directory: temp/gh-pages
           github_token: ${{ secrets.GITHUB_TOKEN }}
       # ===============================
