### 2023-02-13 bc0b01d
Install poetry deps and generate docs in folder docs
