+++ .github/workflows/build-doc.yml
@@ -13,28 +13,29 @@ jobs:
       GITHUB_PAT: ${{ secrets.GITHUB_TOKEN }}
  
     steps:
-      - name: Checkout
-        uses: actions/checkout@v2.3.4
-
-      - name: Set up Python
-        uses: actions/setup-python@v4.5.0
+      - uses: actions/setup-python@v4.5.0
         with:
           python-version: '3.10'
- 
-      - name: Install dependencies
+
+      - name: Install deps
         run: |
-          python -m pip install poetry
-          poetry install
-          python -m pip install -U sphinx
-          python -m pip install sphinx-rtd-theme
-          python -m pip install sphinxcontrib-apidoc
-          python -m pip install sphinx-autoapi
+          sudo apt-get update -y
+          sudo apt-get install -y python3.10-venv python3.10-dev
       
-      - name: make the sphinx docs
+      - uses: actions/checkout@v3
+
+      - name: Prepare venv and install project
         run: |
-          make -C docs clean 
-          sphinx-apidoc -o docs src/metrics_as_scores --force
-          make -C docs html
+          python3.10 -m venv venv
+          source venv/bin/activate
+          pip install poetry
+          poetry install
+          poetry add sphinxcontrib-apidoc sphinx-autoapi
+      
+      - name: Build api-doc
+        run: |
+          sphinx-apidoc -o docs src/metrics_as_scores "**/webapp/*" --force
+          sphinx-build docs/ docs/_build
       
       - uses: actions/upload-artifact@v1
         with:
