+++ .github/workflows/build-doc.yml
@@ -13,27 +13,27 @@ jobs:
       GITHUB_PAT: ${{ secrets.GITHUB_TOKEN }}
  
     steps:
-      - uses: actions/setup-python@v4.5.0
+      - name: Checkout
+        uses: actions/checkout@v2.3.4
+
+      - name: Set up Python
+        uses: actions/setup-python@v4.5.0
         with:
           python-version: '3.10'
-      
-      - uses: actions/checkout@v3
-
-      - name: Prepare venv and install project
+ 
+      - name: Install dependencies
         run: |
-          python3.10 -m venv venv
-          source venv/bin/activate
-          pip install poetry
+          python -m pip install poetry
           poetry install
           python -m pip install -U sphinx
           python -m pip install sphinx-rtd-theme
           python -m pip install sphinxcontrib-apidoc
           python -m pip install sphinx-autoapi
       
-      - name: Make the sphinx docs
+      - name: make the sphinx docs
         run: |
           make -C docs clean 
-          sphinx-apidoc -o docs src/metrics_as_scores "**/webapp/*" --force
+          sphinx-apidoc -o docs src/metrics_as_scores --force
           make -C docs html
       
       - uses: actions/upload-artifact@v1
