+++ paper/paper.md
@@ -1,130 +1,0 @@
----
-affiliations: 
-  - 
-    index: 1
-    name: "Department of Computer Science and Media Technology, Linnaeus University, Sweden"
-authors: 
-  - 
-    affiliation: 1
-    name: "Sebastian Hönel^[corresponding author]"
-    orcid: 0000-0001-7937-1645
-  - 
-    affiliation: 1
-    equal-contrib: false
-    name: "Morgan Ericsson"
-    orcid: 0000-0003-1173-5187
-  - 
-    affiliation: 1
-    equal-contrib: false
-    name: "Welf Löwe"
-    orcid: 0000-0002-7565-3714
-  - 
-    affiliation: 1
-    equal-contrib: false
-    name: "Anna Wingkvist"
-    orcid: 0000-0002-0835-823X
-bibliography: refs.bib
-date: "29 September 2022"
-tags: 
-  - Python
-  - "Distribution fitting"
-  - "Inverse sampling"
-  - "Empirical distributions"
-  - "Kernel density estimation"
-title: "Metrics As Scores: A Tool- and Analysis Suite and Interactive Application for Exploring Context-Dependent Distributions"
----
-
-
-\newcommand{\mas}{\textsf{MAS}\xspace}
-\newcommand\tightto{\!\to\!}
-\newcommand\tightmapsto{\!\mapsto\!}
-\newcommand{\tight}[1]{\,{#1}\,}
-\newcommand{\utight}[1]{{#1}\,}
-
-
-# Summary
-<!-- A summary describing the high-level functionality and purpose of the software for a diverse, non-specialist audience. -->
-<!-- Mention (if applicable) a representative set of past or ongoing research projects using the software and recent scholarly publications enabled by it. -->
-
-<!-- Short introduction into the problem, then -->
-Raw data can be aggregated according to some well-defined standard, method, or calculation, which makes it a metric.
-In software processes, such aggregations are often counts of events or certain properties [@carleton1999].
-However, without the aggregation that is done in a quality model, raw data and metrics are rarely of great value to analysts and decision-makers. This is because quality models are conceived to establish a connection between metrics and certain quality goals [@kaner2004software].
-It is, therefore, difficult to answer the question ``is my metric value good?''.
-With \emph{Metrics As Scores} (\mas) we present an approach that, given some \emph{ideal} value, can transform any metric into a score, given an observation of sufficiently many relevant metrics values.
-While such ideal values were previously attempted to be derived from, e.g., experience or surveys [@benlarbi2000thresh], benchmarks [@alves2010thresh], or by setting practical values [@grady1992practical], with \mas we suggest deriving ideal values additionally in non-parametric, statistical ways.
-To do so, data first needs to be captured in a \emph{relevant} context. A metric value might be good in one context, while it is less so in another.
-Therefore, we suggest generalizing and contextualizing the approach taken by [@UlanLEW21], in which a score is defined to always have a range of $[0,1]$ and linear behavior.
-This means that scores can now also be compared and that a unit change in any score is equally valuable among scores.
-This is not the case for raw data or metrics.
-
-
-\mas consists of a tool- and analysis suite and an interactive application that allows researchers to explore and understand differences in scores across contexts.
-The operationalization of raw data or metrics as scores lies in gathering values that are context-specific (usual), determining an ideal value non-parametrically or by user preference, and then transforming the observed values into distances.
-\mas enables this procedure by unifying the way of obtaining probability densities/masses and conducting appropriate statistical tests.
-More than $120$ different parametric distributions (approx. $20$ of which are discrete) are fitted through a common interface.
-While fitting continuous distributions is straightforward using maximum likelihood estimation, many discrete distributions have integral parameters. For these, \mas solves a mixed-variable global optimization problem using a genetic algorithm and Pymoo [@pymoo].
[…]
