+++ src/metrics_as_scores/webapp/exception.py
@@ -1,4 +1,0 @@
-class PlotException(Exception):
-    def __init__(self, msg: str, *args: object) -> None:
-        self.msg = msg
-        super().__init__(*args)
