+++ _sources/metrics_as_scores.distribution.rst.txt
@@ -1,0 +1,37 @@
+metrics\_as\_scores.distribution package
+========================================
+
+Submodules
+----------
+
+metrics\_as\_scores.distribution.distribution module
+----------------------------------------------------
+
+.. automodule:: metrics_as_scores.distribution.distribution
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+metrics\_as\_scores.distribution.fitting module
+-----------------------------------------------
+
+.. automodule:: metrics_as_scores.distribution.fitting
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+metrics\_as\_scores.distribution.fitting\_problems module
+---------------------------------------------------------
+
+.. automodule:: metrics_as_scores.distribution.fitting_problems
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+Module contents
+---------------
+
+.. automodule:: metrics_as_scores.distribution
+   :members:
+   :undoc-members:
+   :show-inheritance:
