+++ _sources/metrics_as_scores.rst.txt
@@ -1,0 +1,21 @@
+metrics\_as\_scores package
+===========================
+
+Subpackages
+-----------
+
+.. toctree::
+   :maxdepth: 4
+
+   metrics_as_scores.cli
+   metrics_as_scores.data
+   metrics_as_scores.distribution
+   metrics_as_scores.tools
+
+Module contents
+---------------
+
+.. automodule:: metrics_as_scores
+   :members:
+   :undoc-members:
+   :show-inheritance:
