+++ _sources/modules.rst.txt
@@ -1,0 +1,7 @@
+metrics_as_scores
+=================
+
+.. toctree::
+   :maxdepth: 4
+
+   metrics_as_scores
