+++ .github/workflows/static.yml
@@ -30,6 +30,8 @@ jobs:
     steps:
       - name: Checkout
         uses: actions/checkout@v3
+        with:
+          ref: 'gh-pages'
       - name: Setup Pages
         uses: actions/configure-pages@v3
       - name: Upload artifact
