### 2023-02-13 85f9e0f
Add new file about contributing to the application and new datasets
