### 2023-02-13 f2fc80e
Add new build script that copies over a pyproject.toml
