### 2023-02-13 5b00c50
Read version from root first, then locally
