### 2023-02-13 9b6d723
Only add src-dir to path
