### 2023-02-13 e14df9e
Allow package to determine if it runs local or as package and export some directories
