### 2023-02-13 10e948b
Change path to known datasets and use DATASETS_DIR
