+++ src/metrics_as_scores/cli/helpers.py
@@ -3,12 +3,8 @@ from typing import Iterable
 from pathlib import Path
 from json import load, loads
 from urllib.request import urlopen
+from metrics_as_scores.__init__ import DATASETS_DIR
 from metrics_as_scores.distribution.distribution import LocalDataset, KnownDataset
-
-this_dir = Path(__file__).resolve().parent
-datasets_dir = this_dir.parent.parent.parent.joinpath('./datasets')
-
-
 
 
 def isint(s: str) -> bool:
@@ -27,12 +23,12 @@ def isnumeric(s: str) -> bool:
 
 
 def get_local_datasets() -> Iterable[LocalDataset]:
-    for d in scandir(path=str(datasets_dir.resolve())):
+    for d in scandir(path=str(DATASETS_DIR.resolve())):
         if d.is_dir():
             if d.name == '_default':
                 continue
             try:
-                manifest = datasets_dir.joinpath(f'./{d.name}/manifest.json')
+                manifest = DATASETS_DIR.joinpath(f'./{d.name}/manifest.json')
                 if manifest.exists():
                     with open(file=str(manifest), mode='r', encoding='utf-8') as fp:
                         manifest: LocalDataset = load(fp=fp)
@@ -41,11 +37,11 @@ def get_local_datasets() -> Iterable[LocalDataset]:
                 pass
 
 
-KNOWN_DATASETS_FILE = 'https://raw.githubusercontent.com/MrShoenel/metrics-as-scores/master/datasets/known-datasets.json'
+KNOWN_DATASETS_FILE = 'https://raw.githubusercontent.com/MrShoenel/metrics-as-scores/master/src/metrics_as_scores/datasets/known-datasets.json'
 
 def get_known_datasets(use_local_file: bool=False) -> list[KnownDataset]:
     if use_local_file:
-        with open(file=str(datasets_dir.joinpath(f'./known-datasets.json')), mode='r', encoding='utf-8') as fp:
+        with open(file=str(DATASETS_DIR.joinpath(f'./known-datasets.json')), mode='r', encoding='utf-8') as fp:
             return load(fp=fp)
     return loads(urlopen(url=KNOWN_DATASETS_FILE).read().decode('utf-8'))
 
