### 2023-02-13 6af06fc
Start re-writing readme using quarto and add references
