+++ README.qmd
@@ -1,0 +1,265 @@
+---
+title: "Metrics As Scores [![DOI](https://zenodo.org/badge/524333119.svg)](https://zenodo.org/badge/latestdoi/524333119) [![status](https://joss.theoj.org/papers/eb549efe6c0111490395496c68717579/status.svg)](https://joss.theoj.org/papers/eb549efe6c0111490395496c68717579)"
+top-level-division: section
+jupyter: python3
+  
+format:
+  gfm:
+    bibliography: refs.bib
+    citeproc: true
+    toc: true
+    toc-depth: 3
+    number-sections: true
+---
+
+
+```{python}
+#| echo: false
+
+from IPython.display import display, Markdown
+from pathlib import Path
+from sys import path
+from os import getcwd
+
+root = Path(getcwd())
+path.append(str(root.joinpath('./src').resolve()))
+
+from metrics_as_scores.__version__ import __version__ as MAS_VERSION
+
+dm = lambda s: display(Markdown(s))
+```
+
+-----------------
+
+```{python}
+#| echo: false
+
+dm(f'''
+**Please Note**: Metrics As Scores changed considerably between versions [**`v1.0.8`**](/MrShoenel/metrics-as-scores/tree/v1.0.8) and **`v2.xx`**.
+
+The current version is `v{MAS_VERSION}`.
+'''.strip())
+```
+
+
+Version **`v2.x.x`** has the following new features:
+
+- Textual User Interface (TUI)
+	- Download and install known (3rd-party) datasets
+	- Create own datasets: Start with a single CSV and get a fully-fledged dataset, ready for publication, including a report with statistical analyses. After publication, the dataset will be added to the known ones (if you wish) and is discoverable and installable by others.
+	- Start built-in Web Application as local server
+- Proper documentation and testing
+- New version on PyPI. Install the package and run the command line interface by typing **`mas`**!
+
+
+-----------------
+
+Contains the data and scripts needed for the application __`Metrics as Scores`__, check out <https://metrics-as-scores.ml/>.
+
+```{python}
+#| echo: false
+
+dm(f'''
+This package accompanies the paper entitled "_Contextual Operationalization of Metrics As Scores: Is My Metric Value Good?_" [@honel2022qrs].
+It seeks to answer the question whether or not the context a metric was captured in, matters.
+It enables the user to compare contexts and to understand their differences.
+In order to answer the question of whether a metric value is actually good, we need to transform it into a **score**.
+Scores are normalized **and rectified** distances, that can be compared in an apples-to-apples manner, across contexts.
+The same metric value might be good in one context, while it is not in another.
+To borrow an example from the context of software: It is much more acceptable (or common) to have large applications (in terms of lines of code) in the contexts/domains of games and databases than it is for the domains of IDEs and SDKs.
+Given an *ideal* value for a metric (which may also be user-defined), we can transform observed metrics values to distances from that value and then use the cumulative distribution function to map distances to scores.
+'''.strip())
+```
+
+
+
+<!--
+Jump to:
+
+
+- [1. Installation](#1-installation)
+- [2. Stand-alone Usage / Development Setup](#2-stand-alone-usage--development-setup)
+	- [2.1. Setting Up a Virtual Environment](#21-setting-up-a-virtual-environment)
+	- [2.2. Installing Packages](#22-installing-packages)
+- [3. Use Your Own Data](#3-use-your-own-data)
+	- [3.1. Data Preparation](#31-data-preparation)
+	- [3.2. \[Optional\] Computing Fits For Parametric Distributions](#32-optional-computing-fits-for-parametric-distributions)
+	- [3.3. Pre-generating Distributions](#33-pre-generating-distributions)
[…]
