+++ refs.bib
@@ -1,0 +1,10 @@
+@inproceedings{honel2022qrs,
+  author    = {Sebastian Hönel and Morgan Ericsson and Welf Löwe and Anna Wingkvist},
+  title     = {Contextual Operationalization of Metrics As Scores: Is My Metric Value Good?},
+  booktitle = {22nd {IEEE} International Conference on Software Quality, Reliability and Security, {QRS} 2022, Guangzhou, China, December 5-9, 2022},
+  pages     = {333--343},
+  publisher = {{IEEE}},
+  year      = {2022},
+  url       = {https://doi.org/10.1109/QRS57517.2022.00042},
+  doi       = {10.1109/QRS57517.2022.00042}
+}
