### 2023-02-13 fd50ba1
Bump version and make ure copied-over pyproject.toml is included
