+++ pyproject.toml
@@ -1,6 +1,6 @@
 [tool.poetry]
 name = "metrics-as-scores"
-version = "2.1.2"
+version = "2.1.3"
 description = "Interactive web application, tool- and analysis suite for approximating, exploring, understanding, and sampling from conditional distributions."
 authors = ["Sebastian Hönel <development@hoenel.net>"]
 license = "Dual-licensed under GNU General Public License v3 (GPLv3) and closed-source"
@@ -8,6 +8,7 @@ readme = "README.md"
 homepage = "https://github.com/mrshoenel/metrics-as-scores"
 repository = "https://github.com/mrshoenel/metrics-as-scores/issues"
 packages = [{include = "metrics_as_scores", from = "src"}]
+include = ["src/metrics_as_scores/pyproject.toml"]
 keywords = ["distribution fitting", "statistical tests", "context-dependent", "metrics", "quality", "score"]
 classifiers = [
     "Programming Language :: Python :: 3",
