### 2023-02-14 926e00f
Add convenience methods to print Workflow's class doc and wait for continue
