### 2023-02-14 8985dff
Remove unused properties and methods from class Density
