+++ src/metrics_as_scores/webapp/main.py
@@ -1,17 +1,16 @@
-''' Present an interactive explorer for Metrics As Scores.
+"""
+This is the Web Application of Metrics As Scores. It should be run by
+using the command line interface and selecting a dataset. It is an
+interactive explorer for the selected dataset.
 
-Use the ``bokeh serve`` command to run the example by executing:
+During development, this may also be run using the following commands:
 
-    ``bokeh serve src/metrics_as_scores/webapp/ dataset=qcc``
+    ``bokeh serve src/metrics_as_scores/webapp/ --port 5678 --allow-websocket-origin * --args dataset=qcc``
 
-at your command prompt. Then navigate to the URL
-
-    http://localhost:5678/webapp
-
-in your browser (the port might be different).
-
-'''
-
+This command should be run from the project's root. It will start a
+web server on port `5678` using the dataset `qcc` (use the ID of the
+dataset, it is shown by listing all locally available datasets in the CLI).
+"""
 from os import environ
 
 if 'BOKEH_VS_DEBUG' in environ and environ['BOKEH_VS_DEBUG'] == 'true':
@@ -28,7 +27,7 @@ from pathlib import Path
 from typing import Union
 from metrics_as_scores.webapp.exception import PlotException
 from metrics_as_scores.tools.funcs import nonlinspace, natsort
-from metrics_as_scores.distribution.distribution import Empirical, DistTransform, Empirical_discrete, KDE_approx, Parametric, Parametric_discrete, Dataset
+from metrics_as_scores.distribution.distribution import Empirical, DistTransform, Empirical_discrete, KDE_approx, Parametric, Parametric_discrete, Dataset, LocalDataset
 from bokeh.events import MenuItemClick
 from bokeh.io import curdoc
 from bokeh.layouts import column, row
@@ -43,16 +42,26 @@ from numbers import Integral
 # Import the data; Note this is process-static!
 from metrics_as_scores.webapp import data
 ds: Dataset = data.ds
+if ds is None:
+    dummy_manifest: LocalDataset = dict(
+        name = '', desc = '', author = '',
+        qtypes = dict(QTY1 = 'continuous'), contexts = [],
+        desc_qtypes = dict(QTY1 = ''))
+    ds = Dataset(ds=dummy_manifest, df=pd.DataFrame())
 cdfs = data.cdfs
 contexts = list(ds.contexts(include_all_contexts=False))
 contexts.sort(key=natsort)
 contexts = contexts + ['__ALL__']
 
 this_dir = Path(__file__).resolve().parent
-web_dir = data.dataset_dir.joinpath('./web')
+dataset_dir: Path = data.dataset_dir
+if dataset_dir is None:
+    dataset_dir = Path(__file__)
+web_dir = dataset_dir.joinpath('./web')
 
 
 def cap(s: str) -> str:
+    """ :meta private: """
     return s[:1].upper() + s[1:]
 
 
@@ -73,7 +82,9 @@ if len(ds.quantity_types_continuous) > 0:
     dd_scores_items += list([(f'[{qtype}] {ds.qytpe_desc(qtype=qtype)}', qtype) for qtype in qtypes_continuous])
 
 selected_score = dd_scores_items[1]
-"""Has the currently selected quantity type."""
+"""
+:meta private:
+"""
 dd_scores = Dropdown(label=selected_score[0], menu=dd_scores_items)
 
 
@@ -122,6 +133,7 @@ for idx, ctx in enumerate(contexts):
 line_own_source = ColumnDataSource(data=pd.DataFrame(columns=['x', 'y']))
 line_own = plot.line('x', 'y', source=line_own_source, line_alpha=1., color='black', line_width=1.5)
 def update_own_line():
[…]
