### 2023-02-14 f37d545
Document DATASETS_DIR, IS_MAS_LOCAL, and __version__
