### 2023-02-14 80333a6
Add special and private members that should be documented
