+++ src/metrics_as_scores/cli/Workflow.py
@@ -14,9 +14,9 @@ class Workflow:
         self.style_mas = 'bold fg:#8B8000 bg:#000'
         self.style_err = 'bold fg:#8B0000 bg:#000'
     
-    def _print_doc(self) -> None:
+    def _print_doc(self, more: str='') -> None:
         self.q.print(10*'-' + '\n')
-        self.q.print(type(self).__doc__.strip() + '\n')
+        self.q.print(type(self).__doc__.strip() + more + '\n')
         self.q.print(10*'-' + '\n')
     
     def _wait_for(self, what_for: str) -> None:
