### 2023-02-14 0a2bb0f
Refactor how workflows print their class doc-string
