+++ src/metrics_as_scores/cli/KnownDatasets.py
@@ -1,15 +1,24 @@
 from pathlib import Path
-from metrics_as_scores.__init__ import DATASETS_DIR
 from metrics_as_scores.cli.Workflow import Workflow
-from metrics_as_scores.cli.helpers import get_known_datasets, format_file_size
+from metrics_as_scores.cli.helpers import get_known_datasets, format_file_size, KNOWN_DATASETS_FILE
 from metrics_as_scores.distribution.distribution import KnownDataset
 
 
 class KnownDatasetsWorkflow(Workflow):
+    __doc__ = f'''
+This workflow access a curated online list of known datasets that were
+designed to work with Metrics As Scores. The list accessed is at:
+
+{KNOWN_DATASETS_FILE}
+
+If you would like to have your own dataset added to this list, open an
+issue in the Github repository of Metrics As Scores (also, check out the
+contributing guidelines).
+'''.strip()
     def __init__(self) -> None:
         super().__init__()
         self.q.print('\nFetching available datasets ...\n')
-        self.known_datasets = get_known_datasets(use_local_file=False)
+        self.known_datasets: list[KnownDataset] = None
     
     def _print_json_dataset(self, jsd: KnownDataset) -> None:
         self.q.print('     Name: ', style=self.style_mas, end='')
@@ -26,6 +35,12 @@ class KnownDatasetsWorkflow(Workflow):
         self.q.print(f'{format_file_size(jsd["size"])} ({format_file_size(jsd["size_extracted"])} extracted)')
     
     def show_datasets(self) -> None:
+        """Main entry point for this workflow."""
+        self._print_doc()
+
+        self._wait_for(what_for='to fetch and list available datasets')
+        self.known_datasets = get_known_datasets(use_local_file=False)
+
         for jsd in self.known_datasets:
             self.q.print('\nDataset:')
             self.q.print(10*'-')

[…]
