### 2023-02-14 ffc00c9
Document all CLI helpers
