+++ src/metrics_as_scores/cli/helpers.py
@@ -7,14 +7,46 @@ from metrics_as_scores.__init__ import DATASETS_DIR
 from metrics_as_scores.distribution.distribution import LocalDataset, KnownDataset
 
 
+KNOWN_DATASETS_FILE = 'https://raw.githubusercontent.com/MrShoenel/metrics-as-scores/master/src/metrics_as_scores/datasets/known-datasets.json'
+"""
+This is the URL to the curated list of available datasets to be used
+with Metrics As Scores.
+"""
+
 def isint(s: str) -> bool:
+    """
+    Attempts to convert the string to an integer.
+
+    s: ``str``
+        The string to check.
+
+    :rtype: ``bool``
+
+    :returns:
+        `True` if the string can be converted to an `int`; `False`, otherwise.
+    """
     try:
         int(s)
         return True
     except:
         return False
 
+
 def isnumeric(s: str) -> bool:
+    """
+    Attempts to convert a string to a float to check whether it is numeric.
+    This is not the same as :py:meth:`str::isnumeric()`, as this method
+    essentially checks whether ``s`` contains something that looks like a
+    number (int, float, scientific notation, etc).
+
+    s: ``str``
+        The string to check.
+
+    :rtype: ``bool``
+
+    :returns:
+        `True` if the string is numeric; `False`, otherwise.
+    """
     try:
         float(s)
         return True
@@ -23,6 +55,13 @@ def isnumeric(s: str) -> bool:
 
 
 def get_local_datasets() -> Iterable[LocalDataset]:
+    """
+    Opens the dataset directory and looks for locally available datasets.
+    Locally available means datasets that were installed or created manually.
+    A dataset is only considered to be locally available if it has a `manifest.json`.
+
+    :rtype: ``Iterable[LocalDataset]``
+    """
     for d in scandir(path=str(DATASETS_DIR.resolve())):
         if d.is_dir():
             if d.name == '_default':
@@ -37,9 +76,17 @@ def get_local_datasets() -> Iterable[LocalDataset]:
                 pass
 
 
-KNOWN_DATASETS_FILE = 'https://raw.githubusercontent.com/MrShoenel/metrics-as-scores/master/src/metrics_as_scores/datasets/known-datasets.json'
 
 def get_known_datasets(use_local_file: bool=False) -> list[KnownDataset]:
+    """
+    Reads the file :py:data:`KNOWN_DATASETS_FILE` to obtain a list of known datasets.
+
+    use_local_file: ``bool``
+        If true, will attempt to read the known datasets from a local file, instead
+        of the online file. This is only used during development.
+
+    :rtype: ``list[KnownDataset]``
+    """
     if use_local_file:
         with open(file=str(DATASETS_DIR.joinpath(f'./known-datasets.json')), mode='r', encoding='utf-8') as fp:
             return load(fp=fp)
@@ -47,6 +94,18 @@ def get_known_datasets(use_local_file: bool=False) -> list[KnownDataset]:
 
 
 def format_file_size(num_bytes: int) -> str:
+    """
+    Formats bytes into a string with a suffix for bytes, kilobite, etc.
+    The number of bytes in the prefix is less than 1000 and may be rounded
+    to two decimals. For example: 780 B, 1.22 KB, 43 GB, etc. Does NOT use
+    SI-2 suffixes as that would be non-sensical (e.g., what is 1.27 GiB?).
+
+    num_bytes: ``int``
+        Unsigned integer with amount of bytes.
+    
+    :returns:
+        The size in bytes, formatted.
+    """
     suffixes = ['B', 'KB', 'MB', 'GB', 'TB']
     size = float(num_bytes)
 
