### 2023-02-14 0e4f259
Allow to use 'preload' argument and refactor class-docstring usage
