### 2023-02-14 36d8fcc
Catch exception in the CLI and allow printing the error
