+++ src/metrics_as_scores/cli/Workflow.py
@@ -1,4 +1,4 @@
-from typing import Any, TypeVar
+from typing import TypeVar
 from rich.console import Console
 import questionary
 from questionary import Choice
@@ -8,6 +8,10 @@ T = TypeVar('T')
 
 
 class Workflow:
+    """
+    This is the base class for all workflows. It features a few common methods
+    that we use in the derived workflows.
+    """
     def __init__(self) -> None:
         self.c = Console()
         self.q = questionary
@@ -24,6 +28,23 @@ class Workflow:
         return self.q.text(message='', qmark='').ask()
     
     def ask(self, options: list[str], prompt: str='You now have the following options:', rtype: T=int) -> T:
+        """
+        Common method to ask for a selection of options among a list of choices.
+        Options are indexed starting from `0`. If the chosen return type is int,
+        the index is returned; the option as a string, otherwise.
+
+        options: ``list[str]``
+            A list of options to choose (select) from.
+        
+        prompt: ``str``
+            The prompt shown to the user.
+        
+        :rtype: ``T``
+            Returns either int (the index) or the option itself (of any type)
+        
+        :return:
+            The index of the chosen option or the chosen option itself.
+        """
         res = self.askt(
             options=list([(options[idx], idx) for idx in range(len(options))]),
             prompt=prompt)
@@ -35,10 +56,41 @@ class Workflow:
         return res
     
     def askt(self, options: list[tuple[str, T]], prompt: str='You now have the following options:') -> T:
+        """
+        Wrapper around :py:meth:`ask()` that can use any type associated with
+        an option.
+
+        options: ``list[tuple[str, T]]``
+            The options, the text to show and the associated value for each
+        
+        prompt: ``str``
+            The prompt shown to the user.
+        
+        :rtype: ``T``
+
+        :return:
+            Returns the selected option's associated value.
+        """
         choices = list([Choice(title=options[idx][0], value=options[idx][1]) for idx in range(len(options))])
         return self.q.select(message=prompt, choices=choices, use_shortcuts=True).ask()
     
     def print_info(self, text_normal: str, text_vital: str=None, end: str=None, arrow: str=' -> ') -> None:
+        """
+        Used to print an info that consists of a normal text (without extra styles)
+        and a vital text that has some extra styling applied to emphasize it.
+
+        text_normal: ``str``
+            The text that does not have extra styling
+        
+        text_vital: ``str``
+            The text with extra styling for emphasis.
+        
+        end: ``str``
+            The string to print at the end of the info.
+        
+        arrow: ``str``
+            The string to print at the beginning of the info.
+        """
         kwargs = {}
         if end != None:
             kwargs['end'] = end
