### 2023-02-14 0a5beaa
Add documentation to all functions
