+++ src/metrics_as_scores/data/pregenerate.py
@@ -1,13 +1,13 @@
 from pathlib import Path
 import pandas as pd
-from typing import Any, Union
+from typing import Union
 from warnings import warn
 from os.path import exists
 from pickle import dump, load
 from joblib import Parallel, delayed
 from tqdm import tqdm
 from metrics_as_scores.tools.funcs import flatten_dict
-from metrics_as_scores.data.pregenerate_fit import Continuous_RVs_dict, Discrete_RVs_dict
+from metrics_as_scores.data.pregenerate_fit import Continuous_RVs_dict, Discrete_RVs_dict, FitResult
 from metrics_as_scores.distribution.distribution import Dataset, Density, DistTransform, Dataset, Empirical, Empirical_discrete, KDE_approx, Parametric, Parametric_discrete
 from metrics_as_scores.distribution.fitting import StatisticalTest
 from sklearn.model_selection import ParameterGrid
@@ -15,7 +15,42 @@ from sklearn.model_selection import ParameterGrid
 
 
 
-def generate_densities(dataset: Dataset, clazz: type[Density]=Empirical, unique_vals: bool=None, resample_samples=250_000, dist_transform: DistTransform=DistTransform.NONE) -> dict[str, Density]:
+def generate_densities(
+    dataset: Dataset,
+    clazz: type[Density]=Empirical,
+    unique_vals: bool=None,
+    resample_samples=250_000,
+    dist_transform: DistTransform=DistTransform.NONE
+) -> dict[str, Density]:
+    """
+    Generates a set of :py:class:`Density` objects for a certain :py:class:`DistTransform`.
+    For each combination, we will later save one file that is then to be used in the web
+    application, as generating these on-the-fly would take too long.
+
+    dataset: ``Dataset``
+        Required for obtaining quantity types, contexts, and filtered data.
+    
+    clazz: ``type[Density]``
+        A type of empirical density to generate densities for.
+    
+    unique_vals: ``bool``
+        Used to conditionally add some jitter to data to all data points unique. This is
+        automatically set to `True` if the class is :py:class:`Empirical`, because this
+        class is for continuous RVs. If the data is not continuous (real), then setting
+        this to `True` will make it so.
+    
+    resample_samples: ``int``
+        Unsigned integer, passed forward to the type of dict[str, Density].
+    
+    dist_transform: ``DistTransform``
+        The chosen transformation for the data.
+    
+    :rtype: ``dict[str, Density]``
+
+    :return:
+        A dictionary where the key is made of the context and quantity type, and
+        the value is the generated :py:class:`Empirical` density.
+    """
     contexts = list(dataset.contexts(include_all_contexts=True))
     param_grid = { 'context': contexts, 'qtype': dataset.quantity_types }
     expanded_grid = pd.DataFrame(ParameterGrid(param_grid=param_grid))
@@ -38,7 +73,38 @@ def generate_densities(dataset: Dataset, clazz: type[Density]=Empirical, unique_
     return dict(cdfs)
 
 
-def fits_to_MAS_densities(dataset: Dataset, distns_dict: dict[int, dict[str, Any]], dist_transform: DistTransform, use_continuous: bool) -> dict[str, Union[Parametric, Parametric_discrete]]:
+def fits_to_MAS_densities(
+    dataset: Dataset,
+    distns_dict: dict[int, FitResult],
+    dist_transform: DistTransform,
+    use_continuous: bool
+) -> dict[str, Union[Parametric, Parametric_discrete]]:
+    """
+    Converts previously produced parametric fits to :py:class:`Density` objects that can
+    be loaded and used in the web application.
+    Similar to :py:meth:`generate_densities()`, this method also returns a dictionary
+    with generated parametric densities.
+
+    dataset: ``Dataset``
[…]
