### 2023-02-14 973c5f4
Add documentation to generate_parametric_fits()
