+++ src/metrics_as_scores/distribution/distribution.py
@@ -23,16 +23,52 @@ from strenum import StrEnum
 
 
 class DistTransform(StrEnum):
+    """
+    This is an enumeration of transforms applicable to distributions of a quantity.
+    A transform first computes the desired ideal (transform) value from the given
+    density (e.g., the expectation) and then transforms the initial distribution of
+    values into a distribution of distances.
+    """
     NONE = '<none>'
+    """Do not apply any transform."""
+
     EXPECTATION = 'E[X] (expectation)'
+    """
+    Compute the expectation of the random variable.
+    This is similar to :math:`\mathbb{E}[X]=\int_{-\infty}^{\infty}x*f_X(x) dx` for a
+    continuous random variable.
+    """
+
     MEDIAN = 'Median (50th percentile)'
+    """
+    Compute the median (50th percentile) of the random variable. The median is defined
+    as the value that splits a probability distribution into a lower and higher half.
+    """
+    
     MODE = 'Mode (most likely value)'
+    """
+    The mode of a random variable is the most frequently occurring value, i.e., the
+    value with the highest probability (density).
+    """
+
     INFIMUM = 'Infimum (min. observed value)'
+    """
+    The infimum is the lowest observed value of some empirical random variable.
+    """
+
     SUPREMUM = 'Supremum (max. observed value)'
+    """
+    The supremum is the highest observed value of some empirical random variable.
+    """
 
 
 
 class JsonDataset(TypedDict):
+    """
+    This class is the base class for the :py:class:`LocalDataset` and the
+    :py:class:`KnownDataset`. Each manifest should have a name, id, description,
+    and author.
+    """
     name: str
     desc: str
     id: str
@@ -40,6 +76,10 @@ class JsonDataset(TypedDict):
 
 
 class LocalDataset(JsonDataset):
+    """
+    This dataset extends the :py:class:`JsonDataset` and adds properties that
+    are filled out when locally creating a new dataset.
+    """
     origin: str
     colname_data: str
     colname_type: str
@@ -52,6 +92,10 @@ class LocalDataset(JsonDataset):
 
 
 class KnownDataset(JsonDataset):
+    """
+    This dataset extends the :py:class:`JsonDataset` with properties that are
+    known about datasets that are available to Metrics As Scores online.
+    """
     info_url: str
     download: str
     size: int
@@ -60,7 +104,60 @@ class KnownDataset(JsonDataset):
 
 
 class Density(ABC):
-    def __init__(self, range: tuple[float, float], pdf: Callable[[float], float], cdf: Callable[[float], float], ppf: Callable[[float], float]=None, ideal_value: float=None, dist_transform: DistTransform=DistTransform.NONE, transform_value: float=None, qtype: str=None, context: str=None, **kwargs) -> None:
+    """
+    This is the abstract base class for parametric and empirical densities. A
+    :py:class:`Density` represents a concrete instance of some random variable
+    and its PDF, CDF, and PPF. It also stores information about this concrete
+    instance came to be (e.g., by some concrete transform).
+
+    This class provides a set of common getters and setters and also provides
+    some often needed conveniences, such as computing the practical domain.
+    As for the PDF, CDF, and PPF, all known sub-classes have a specific way of
+    obtaining these, and this class' responsibility lies in vectorizing these
+    functions.
+    """
+
+    def __init__(
+        self,
+        range: tuple[float, float],
+        pdf: Callable[[float], float],
+        cdf: Callable[[float], float],
+        ppf: Callable[[float], float]=None,
+        ideal_value: float=None,
+        dist_transform: DistTransform=DistTransform.NONE,
+        transform_value: float=None,
+        qtype: str=None,
+        context: str=None,
+        **kwargs
+    ) -> None:
+        """
+        range: ``tuple[float, float]``
+            The range of the data.
+        
[…]
