### 2023-02-14 d5e0b94
Completely document fitting module
