+++ src/metrics_as_scores/distribution/fitting.py
@@ -71,6 +71,9 @@ class DesignSpaceTerminationFixed(DesignSpaceTermination):
 
 
 class SingleObjectiveTermination(DefaultTermination):
+    """
+    This class is used as termination criterion for the :py:class:`FitterPymoo`.
+    """
     def __init__(self, xtol=1e-8, cvtol=1e-8, ftol=1e-6, period=75, max_time: int=600, **kwargs) -> None:
         x = RobustTermination(DesignSpaceTerminationFixed(xtol), period=period)
         cv = RobustTermination(ConstraintViolationTermination(cvtol, terminate_when_feasible=False), period=period)
@@ -83,6 +86,11 @@ class SingleObjectiveTermination(DefaultTermination):
 
 
 class Fitter:
+    """
+    This class is used as a generic approach to fit any random variable to any kind
+    of data. However, it is recommended to use the more stable :py:class:`FitterPymoo`.
+    """
+
     # All have 'loc' which is an integer, and we will derive it from the data (i.e., use int([min/max](data)))
     Practical_Ranges = dict(
         bernoulli_gen = dict(
@@ -176,6 +184,10 @@ class Fitter:
     """
 
     def __init__(self, dist: type[Union[rv_continuous, rv_discrete]]) -> None:
+        """
+        dist: ``type[Union[rv_continuous, rv_discrete]]``
+            Specify the class of the random variable you want to fit.
+        """
         if not issubclass(dist, rv_continuous) and not issubclass(dist, rv_discrete):
             raise Exception('The distribution given is not continuous or discrete random variable.')
         self.dist = dist
@@ -189,13 +201,26 @@ class Fitter:
 
     @property
     def is_discrete(self) -> bool:
+        """Shortcut getter to return whether the used random variable is discrete."""
         return issubclass(self.dist, rv_discrete)
 
     @property
     def is_continuous(self) -> bool:
+        """Shortcut getter to return whether the used random variable is continuous."""
         return issubclass(self.dist, rv_continuous)
 
-    def fit(self, data) -> dict[str, Union[float, int]]:
+    def fit(self, data: NDArray[Shape["*"], Float]) -> dict[str, Union[float, int]]:
+        """
+        Convenience method to fit the random variable. If it is continuous, calls
+        :py:meth:`rv_continuous.fit()`, which uses maximum likelihood estimation.
+        If it is discrete, uses differential evolution to find an estimate.
+
+        data: ``NDArray[Shape["*"], Float]``
+            The data to fit to.
+        
+        :return:
+            A dictionary with named parameters and their values.
+        """
         params: tuple[float] = None
         if self.is_continuous:
             dist: type[rv_continuous] = self.dist
@@ -213,6 +238,11 @@ class Fitter:
 
 
 class FitterPymoo(Fitter):
+    """
+    This class inherits from :py:class:`Fitter` and is the modern successor to it.
+    This is because it uses :py:mod:`pymoo` to fit discrete random variables, together
+    with a set of specially defined fitting problems.
+    """
     def __init__(self, dist: type[Union[rv_continuous, rv_discrete]]) -> None:
         if not isclass(dist):
             raise Exception(f'The given object {dist} is not a subclass of rv_generic.')
@@ -227,6 +257,37 @@ class FitterPymoo(Fitter):
         super().__init__(dist=dist)
 
     def fit(self, data: NDArray[Shape["*"], Float], max_samples = 10_000, minimize_seeds = [1_337, 0xdeadbeef, 45640321], verbose: bool=True, stop_after_first_res: bool=True) -> dict[str, Union[float, int]]:
+        """
+        Fits the random variable to the given data. For continuous data, calls :py:meth:`Fitter.fit()`.
+        For discrete random variables, however, it solves a mixed-variable problem using a genetic
[…]
