### 2023-02-14 7a10999
Document some funcs and tools, espcially Lazy and SelfResetLazy
