+++ src/metrics_as_scores/tools/lazy.py
@@ -9,7 +9,14 @@ T = TypeVar('T')
 
 
 class Lazy(Generic[T]):
+    """
+    Helper class to lazily compute a value, thread-safe.
+    """
     def __init__(self, fn_create_val: Callable[[], T]) -> None:
+        """
+        fn_create_val: ``Callable[[], T]``
+            Function that produces the desired value.
+        """
         self._val: T = None
         self._has_val = False
         self._semaphore = Semaphore(1)
@@ -17,6 +24,10 @@ class Lazy(Generic[T]):
         self._tpe = ThreadPoolExecutor(max_workers=1)
     
     def unset_value(self) -> Self:
+        """
+        Used to explicitly unset (destroy) the previously produced value.
+        This method is thread-safe.
+        """
         try:
             self._semaphore.acquire()
             self._val = None
@@ -28,10 +39,17 @@ class Lazy(Generic[T]):
 
     @property
     def has_value_volatile(self) -> bool:
+        """
+        Checks, in a volatile manner, if a value is present.
+        """
         return self._has_val
 
     @property
     def has_value(self) -> bool:
+        """
+        Checks whether a value is present, after having acquired the lock.
+        Note that the value may be gone after method exit anyway.
+        """
         try:
             self._semaphore.acquire()
             return self._has_val
@@ -40,10 +58,18 @@ class Lazy(Generic[T]):
 
     @property
     def value_volatile(self) -> Union[None, T]:
+        """
+        Returns, in a volatile manner, the value. The value may or may not exist.
+        """
         return self._val
 
     @property
     def value(self) -> T:
+        """
+        Gets the value. If it was not produced previously, it will be produced first.
+        Therefore, this getter is guaranteed to return a value (provided the factory
+        of the value does not throw).
+        """
         try:
             self._semaphore.acquire()
             if not self._has_val:
@@ -55,6 +81,10 @@ class Lazy(Generic[T]):
 
     @property
     def value_future(self) -> Future[T]:
+        """
+        Gets an awaitable :py:class:`Future` that will yield the produced value
+        once it is available.
+        """
         f = Future()
 
         temp = self._val
@@ -72,7 +102,22 @@ class Lazy(Generic[T]):
 
 
 class SelfResetLazy(Generic[T]):
+    """
+    Similar to :py:class:`Lazy`, this class also automatically destroys its value
+    after some timeout, so that subsequent requests to it force the factory to
+    produce a new instance.
+    """
     def __init__(self, fn_create_val: Callable[[], T], fn_destroy_val: Callable[[T], Any]=None, reset_after: float=None) -> None:
+        """
+        fn_create_val: ``Callable[[], T]``
+            Function that produces the desired value.
+        
+        fn_destroy_val: ``Callable[[T], Any]``
+            Function that will be given the value before it is de-referenced here.
+        
+        reset_after: ``float``
+            Amount of time, in seconds, after which the produced value ought to be destroyed.
+        """
         self._val: T = None
         self._has_val = False
         self._semaphore = Semaphore(1)
@@ -85,6 +130,9 @@ class SelfResetLazy(Generic[T]):
     
     @property
     def reset_after(self):
+        """
+        Thread-safe getter for the reset-after property.
+        """
         try:
             self._semaphore.acquire()
             return self._reset_after
@@ -93,6 +141,9 @@ class SelfResetLazy(Generic[T]):
     
     @reset_after.setter
     def reset_after(self, value: float=None):
+        """
+        Thread-safe setter for the reset-after property.
+        """
         try:
             self._semaphore.acquire()
             self._reset_after = value
@@ -103,6 +154,10 @@ class SelfResetLazy(Generic[T]):
         return self
     
     def unset_value(self):
+        """
[…]
