+++ _sources/metrics_as_scores.rst.txt
@@ -11,6 +11,7 @@ Subpackages
    metrics_as_scores.data
    metrics_as_scores.distribution
    metrics_as_scores.tools
+   metrics_as_scores.webapp
 
 Module contents
 ---------------
