### 2023-02-14 bde4032
Upgrade version to v2.1.4 after documenting everything
