### 2023-02-14 MrShoenel/metrics-as-scores#1 @MrShoenel
Hey, I will answer some of your questions:

1. I am the sole author of the software. The other authors are my supervisors who helped with the ideation of it.
2. There are two datasets that are currently associated with this paper (more to come next week). The two known datasets that can be installed using the CLI are <https://doi.org/10.5281/zenodo.7633950> and <https://doi.org/10.5281/zenodo.7633989>.
3. Both are the case. The paper primarily is concerned with the results derived from using the Qualitas.class corpus (which is displayed at <https://metrics-as-scores.ml/>. However, it also contributes to the methodology and significance of the approach. I have shared the paper privately with you for the time being.
4. There are no such claims. However, the purpose of pre-generating densities is to enable a real-time experience in the web application. I have tried to make that clear within the documentation, and I am going to add something about that to the readme, too!
