### 2023-02-15 8b9f14b
Change the default optimizer from minimize to DIRECT for better results
