### 2023-02-15 791afa0
Add a try/except block to catch unfit parametric RVs
