### 2023-02-15 ebf4932
Do not allow empty grids and raise an exception in this case
