### 2023-02-15 532a134
Remove leftover print statement
