### 2023-02-15 4fe35c3
Add kwargs to fit() so it can ignore superfluous args
