### 2023-02-15 c743f91
Remove args to fit() because they're the default anyway
