### 2023-02-15 de9f80b
Fix an error where parameters were not returned as dict previously
