### 2023-02-16 fef443f
Ignore more discrete RVs by default (these can take a long time)
