### 2023-02-16 87f1fb9
Add a bib file for known datasets that we can use in the readme
