+++ src/metrics_as_scores/datasets/known-datasets.bib
@@ -1,0 +1,36 @@
+@dataset{dataset_qcc,
+  author       = {Sebastian Hönel},
+  title        = {{Metrics As Scores Dataset: Metrics and Domains 
+                   From the Qualitas.class corpus}},
+  month        = feb,
+  year         = 2023,
+  publisher    = {Zenodo},
+  version      = {v1.0},
+  doi          = {10.5281/zenodo.7633949},
+  url          = {https://doi.org/10.5281/zenodo.7633949}
+}
+
+@dataset{dataset_elisa,
+  author       = {Sebastian Hönel},
+  title        = {{Metrics As Scores Dataset: Elisa Spectrophotometer 
+                   Positive Samples}},
+  month        = feb,
+  year         = 2023,
+  publisher    = {Zenodo},
+  version      = {v1.0},
+  doi          = {10.5281/zenodo.7633989},
+  url          = {https://doi.org/10.5281/zenodo.7633989}
+}
+
+@dataset{dataset_diamonds-ideal-if,
+  author       = {Sebastian Hönel},
+  title        = {{Metrics As Scores Dataset: Price, weight, and 
+                   other properties of over 1,200 ideal-cut and best-
+                   clarity diamonds}},
+  month        = feb,
+  year         = 2023,
+  publisher    = {Zenodo},
+  version      = {v1.0},
+  doi          = {10.5281/zenodo.7647596},
+  url          = {https://doi.org/10.5281/zenodo.7647596}
+}
