+++ refs.bib
@@ -8,3 +8,164 @@
   url       = {https://doi.org/10.1109/QRS57517.2022.00042},
   doi       = {10.1109/QRS57517.2022.00042}
 }
+
+@Book{ggplot2,
+	author = {Hadley Wickham},
+	title = {{ggplot2}: Elegant Graphics for Data Analysis},
+	publisher = {Springer-Verlag New York},
+	year = {2016},
+	isbn = {9783319242774},
+	url = {https://ggplot2.tidyverse.org}
+}
+
+@article{gil2016software,
+  author    = {Joseph Yossi Gil and
+               Gal Lalouche},
+  title     = {When do Software Complexity Metrics Mean Nothing? - When Examined out of Context},
+  journal   = {J. Object Technol.},
+  volume    = {15},
+  number    = {1},
+  pages     = {2:1--25},
+  year      = {2016},
+  url       = {https://doi.org/10.5381/jot.2016.15.5.a2},
+  doi       = {10.5381/jot.2016.15.5.a2},
+  timestamp = {Wed, 17 Feb 2021 08:57:52 +0100},
+  biburl    = {https://dblp.org/rec/journals/jot/GilL16.bib},
+  bibsource = {dblp computer science bibliography, https://dblp.org}
+}
+
+@article{terra2013qcc,
+  author    = {Ricardo Terra and
+               Luis Fernando Miranda and
+               Marco T{\'{u}}lio Valente and
+               Roberto da Silva Bigonha},
+  title     = {Qualitas.class corpus: a compiled version of the qualitas corpus},
+  journal   = {{ACM} {SIGSOFT} Softw. Eng. Notes},
+  volume    = {38},
+  number    = {5},
+  pages     = {1--4},
+  year      = {2013},
+  url       = {https://doi.org/10.1145/2507288.2507314},
+  doi       = {10.1145/2507288.2507314},
+  timestamp = {Fri, 19 Aug 2022 14:05:12 +0200},
+  biburl    = {https://dblp.org/rec/journals/sigsoft/TerraMVB13.bib},
+  bibsource = {dblp computer science bibliography, https://dblp.org}
+}
+
+@inproceedings{tempero2010qualitas,
+  author    = {Ewan D. Tempero and
+               Craig Anslow and
+               Jens Dietrich and
+               Ted Han and
+               Jing Li and
+               Markus Lumpe and
+               Hayden Melton and
+               James Noble},
+  editor    = {Jun Han and
+               Tran Dan Thu},
+  title     = {The Qualitas Corpus: {A} Curated Collection of Java Code for Empirical
+               Studies},
+  booktitle = {17th Asia Pacific Software Engineering Conference, {APSEC} 2010, Sydney,
+               Australia, November 30 - December 3, 2010},
+  pages     = {336--345},
+  publisher = {{IEEE} Computer Society},
+  year      = {2010},
+  url       = {https://doi.org/10.1109/APSEC.2010.46},
+  doi       = {10.1109/APSEC.2010.46},
+  timestamp = {Fri, 27 Mar 2020 08:46:40 +0100},
+  biburl    = {https://dblp.org/rec/conf/apsec/TemperoADHLLMN10.bib},
+  bibsource = {dblp computer science bibliography, https://dblp.org}
+}
+
+@article{tukey1949anova,
+ ISSN = {0006341X, 15410420},
+ URL = {http://www.jstor.org/stable/3001913},
+ author = {John W. Tukey},
+ journal = {Biometrics},
+ number = {2},
+ pages = {99--114},
+ publisher = {[Wiley, International Biometric Society]},
+ title = {Comparing Individual Means in the Analysis of Variance},
[…]
