### 2023-02-16 cbad44f
All formatting file size without decimals
