+++ README.md
@@ -1,211 +1,485 @@
-Metrics As Scores [![DOI](https://zenodo.org/badge/524333119.svg)](https://zenodo.org/badge/latestdoi/524333119) [![status](https://joss.theoj.org/papers/eb549efe6c0111490395496c68717579/status.svg)](https://joss.theoj.org/papers/eb549efe6c0111490395496c68717579)
-=================
+Metrics As Scores
+[![DOI](https://zenodo.org/badge/524333119.svg)](https://zenodo.org/badge/latestdoi/524333119)
+[![status](https://joss.theoj.org/papers/eb549efe6c0111490395496c68717579/status.svg)](https://joss.theoj.org/papers/eb549efe6c0111490395496c68717579)
+================
 
+- <a href="#usage" id="toc-usage"><span
+  class="toc-section-number">1</span> Usage</a>
+  - <a href="#sec-tui" id="toc-sec-tui"><span
+    class="toc-section-number">1.1</span> Text-based User Interface
+    (TUI)</a>
+  - <a href="#sec-webapp" id="toc-sec-webapp"><span
+    class="toc-section-number">1.2</span> Web Application</a>
+  - <a href="#sec-dev-setup" id="toc-sec-dev-setup"><span
+    class="toc-section-number">1.3</span> Development Setup</a>
+    - <a href="#setting-up-a-virtual-environment"
+      id="toc-setting-up-a-virtual-environment"><span
+      class="toc-section-number">1.3.1</span> Setting Up a Virtual
+      Environment</a>
+    - <a href="#installing-packages" id="toc-installing-packages"><span
+      class="toc-section-number">1.3.2</span> Installing Packages</a>
+- <a href="#example-usage" id="toc-example-usage"><span
+  class="toc-section-number">2</span> Example Usage</a>
+  - <a href="#software-metrics-example"
+    id="toc-software-metrics-example"><span
+    class="toc-section-number">2.1</span> Software Metrics Example</a>
+  - <a href="#diamonds-example" id="toc-diamonds-example"><span
+    class="toc-section-number">2.2</span> Diamonds Example</a>
+- <a href="#datasets" id="toc-datasets"><span
+  class="toc-section-number">3</span> Datasets</a>
+  - <a href="#use-your-own" id="toc-use-your-own"><span
+    class="toc-section-number">3.1</span> Use Your Own</a>
+  - <a href="#sec-known-datasets" id="toc-sec-known-datasets"><span
+    class="toc-section-number">3.2</span> Known Datasets</a>
+- <a href="#personalizing-the-web-application"
+  id="toc-personalizing-the-web-application"><span
+  class="toc-section-number">4</span> Personalizing the Web
+  Application</a>
+- <a href="#references" id="toc-references">References</a>
 
+------------------------------------------------------------------------
 
-**Version `v2.x.x`** has the following new features:
+**Please Note**: ***Metrics As Scores*** (`MAS`) changed considerably
+between versions
+[**`v1.0.8`**](/MrShoenel/metrics-as-scores/tree/v1.0.8) and
+**`v2.x.x`**.
 
-- Textual User Interface (TUI)
-	- Download and install known (3rd-party) datasets
-	- Create own datasets: Start with a single CSV and get a fully-fledged dataset, ready for publication, including a report with statistical analyses. After publication, the dataset will be added to the known ones (if you wish) and is discoverable and installable by others.
+The current version is `v2.1.4`.
+
+Version **`v2.x.x`** has the following new features:
+
+- [Textual User Interface (TUI)](#sec-tui)
 - Proper documentation and testing
-- New version on PyPI. Install the package and run the command line interface by typing `mas`!
+- New version on PyPI. Install the package and run the command line
+  interface by typing **`mas`**!
 
[…]
