+++ pyproject.toml
@@ -35,7 +35,6 @@ bokeh = "^2.4.3"
 joblib = "^1.2.0"
 jupyterlab = "^3.4.7"
 matplotlib = "^3.6.0"
-nptyping = "^2.3.1"
 ptvsd = "^4.3.2"
 pymoo = "^0.6.0"
 scipy = "^1.9.1"
@@ -49,6 +48,7 @@ rich = "^13.3.1"
 questionary = "^1.10.0"
 tabulate = "^0.9.0"
 scikit-learn = "^1.2.1"
+nptyping = "^2.4.1"
 
 
 [tool.poetry.group.dev.dependencies]
