### 2023-02-17 a5a01a4
Remove outdated import test
