+++ test/test_metrics.py
@@ -1,5 +1,0 @@
-import base
-
-def test_MetricID():
-    from metrics_as_scores.data.metrics import MetricID
-    assert len(list(MetricID)) > 0
