### 2023-02-17 eb75ec8
Add pytest-cov and poethepoet to simply run scripts
