+++ pyproject.toml
@@ -49,6 +49,8 @@ questionary = "^1.10.0"
 tabulate = "^0.9.0"
 scikit-learn = "^1.2.1"
 nptyping = "^2.4.1"
+pytest-cov = "^4.0.0"
+poethepoet = "^0.18.1"
 
 
 [tool.poetry.group.dev.dependencies]
@@ -61,3 +63,16 @@ coverage = "^6.5.0"
 [build-system]
 requires = ["poetry-core"]
 build-backend = "poetry.core.masonry.api"
+
+
+[tool.pytest.ini_options]
+testpaths = ["./test"]
+filterwarnings = ["ignore:::.*nptyping*"]
+addopts = [
+    "--cov=metrics_as_scores",
+    "--cov-report=term-missing"
+]
+
+[tool.poe.tasks]
+test = "pytest"
+cov = "pytest --cov-report html:htmlcov"
