### 2023-02-17 e7320e2
Update readme with how to run tests
