### 2023-02-17 9c41145
Refactor KDE_integrate to only compute a PPF if needed
