### 2023-02-17 4ffa34d
Add documentation for number of decimals
