### 2023-02-17 31d3588
Add some testing for Density, KDE_integrate, and KDE_approx
