### 2023-02-17 aa59369
Override _min_max to handle the case when the density is not fit
