### 2023-02-17 67824bf
Only apply rounding if the data is not NaNs
