### 2023-02-17 2160492
Add tests for Empirical, Empirical_discrete, and Parametric
