### 2023-02-17 0812039
Fix an error in pval and stat (wrong indexing)
