+++ .github/workflows/codecov.yml
@@ -10,7 +10,6 @@ jobs:
     - uses: checkout@v2
       with:
         fetch-depth: ‘2’
-
     - name: Setup Python
       uses: actions/setup-python@master
       with:
@@ -20,4 +19,4 @@ jobs:
         pip install coverage
         coverage run -m unittest
     - name: Upload Coverage to Codecov
-      uses: codecov/codecov-action@v3
+      uses: codecov/codecov-action@v2
