+++ .github/workflows/codecov.yml
@@ -7,7 +7,7 @@ jobs:
       OS: ubuntu-latest
       PYTHON: '3.10'
     steps:
-    - uses: checkout@v2
+    - uses: actions/checkout@v2
       with:
         fetch-depth: ‘2’
     - name: Setup Python
