+++ .github/workflows/codecov.yml
@@ -3,10 +3,12 @@ on: [push, pull_request]
 jobs:
   run:
     runs-on: ubuntu-latest
-    env:
-      OS: ubuntu-latest
-      PYTHON: '3.10'
+
     steps:
+    - name: Set up Python
+      uses: actions/setup-python@v4.5.0
+      with:
+        python-version: '3.10'
     - uses: actions/checkout@v2
       with:
         fetch-depth: ‘2’
