### 2023-02-17 5c32fcf
Remove double Python setup
