+++ .github/workflows/codecov.yml
@@ -9,13 +9,9 @@ jobs:
       uses: actions/setup-python@v4.5.0
       with:
         python-version: '3.10'
-    - uses: actions/checkout@v2
+    - uses: actions/checkout@v2.3.4
       with:
-        fetch-depth: ‘2’
-    - name: Setup Python
-      uses: actions/setup-python@master
-      with:
-        python-version: 3.10
+        fetch-depth: '2'
     - name: Generate Report
       run: |
         pip install coverage
