+++ .github/workflows/codecov.yml
@@ -3,18 +3,23 @@ on: [push, pull_request]
 jobs:
   run:
     runs-on: ubuntu-latest
-
+    
     steps:
+    - name: Checkout
+      uses: actions/checkout@v2.3.4
+
     - name: Set up Python
       uses: actions/setup-python@v4.5.0
       with:
         python-version: '3.10'
-    - uses: actions/checkout@v2.3.4
-      with:
-        fetch-depth: '2'
+
+    - name: Install dependencies
+      run: |
+        python -m pip install poetry
+        poetry install
+
     - name: Generate Report
       run: |
-        pip install coverage
-        coverage run -m unittest
+        poe cov
     - name: Upload Coverage to Codecov
-      uses: codecov/codecov-action@v2
+      uses: codecov/codecov-action@v3
