+++ .github/workflows/codecov.yml
@@ -3,23 +3,19 @@ on: [push, pull_request]
 jobs:
   run:
     runs-on: ubuntu-latest
-    
-    steps:
-    - name: Checkout
-      uses: actions/checkout@v2.3.4
 
+    steps:
     - name: Set up Python
       uses: actions/setup-python@v4.5.0
       with:
         python-version: '3.10'
-
-    - name: Install dependencies
-      run: |
-        python -m pip install poetry
-        poetry install
-
+    - uses: actions/checkout@v2.3.4
+      with:
+        fetch-depth: '2'
     - name: Generate Report
       run: |
-        pytest
+        pip install pytest
+        pip install pytest-codecov
+        pytest --codecov
     - name: Upload Coverage to Codecov
-      uses: codecov/codecov-action@v3
+      uses: codecov/codecov-action@v2
