### 2023-02-17 7257732
Fix broken coverage command
