### 2023-02-17 4098afb
Make the ordinary test command exclude coverage
