### 2023-02-17 35af202
Fix a problem where the exponent could be zero
