### 2023-02-17 41730b1
Add more testing for Parametric_discrete
