+++ test/data/test_Dataset.py
@@ -1,0 +1,150 @@
+import numpy as np
+import pandas as pd
+from pytest import raises
+from json import load
+from pathlib import Path
+from metrics_as_scores.distribution.distribution import Dataset, JsonDataset, KnownDataset, LocalDataset, DistTransform
+from metrics_as_scores.cli.helpers import get_known_datasets, get_local_datasets, isnumeric
+
+
+this_dir = Path(__file__).parent
+qcc_manifest_file = this_dir.joinpath('./qcc-manifest.json')
+elisa_manifest_file = this_dir.joinpath('./elisa-manifest.json')
+elisa_data_file = this_dir.joinpath('./elisa-org-data.csv')
+
+
+def test_Dataset():
+    qcc_manifest: LocalDataset = None
+    with open(file=str(qcc_manifest_file), mode='r', encoding='utf-8') as fp:
+        qcc_manifest = load(fp=fp)
+    # We test the dataset loading the QCC manifest.
+    ds = Dataset(ds=qcc_manifest, df=pd.DataFrame())
+
+    qtypes = set(ds.quantity_types)
+    assert len(qtypes) == 23
+    qtypes_c = set(ds.quantity_types_continuous)
+    qtypes_d = set(ds.quantity_types_discrete)
+
+    assert len(qtypes_c) == 5
+    for qt in ['LCOM', 'RMA', 'RMD', 'RMI', 'SIX']:
+        assert qt in qtypes_c
+    
+    for qt in qtypes_c.union(qtypes_d):
+        assert qt in qtypes
+    
+
+    # There are 11 contexts
+    contexts = set(ds.contexts(include_all_contexts=False))
+    assert len(contexts) == 11
+    assert not '__ALL__' in contexts
+    contexts = set(ds.contexts(include_all_contexts=True))
+    assert len(contexts) == 12
+    assert '__ALL__' in contexts
+
+    
+    # Check type of ideal values: must be int/float or None:
+    for iv in ds.ideal_values.values():
+        assert iv is None or isnumeric(iv)
+    
+    # Check descriptions: None or str:
+    for desc in [ds.qytpe_desc(qtype=qty) for qty in qtypes]:
+        assert desc is None or isinstance(desc, str)
+    for desc in [ds.context_desc(context=ctx) for ctx in ds.contexts(include_all_contexts=False)]:
+        assert desc is None or isinstance(desc, str)
+    # Also:
+    with raises(KeyError):
+        ds.context_desc('__ALL__')
+    
+
+def test_Dataset_data():
+    """
+    Here we'll test subsetting the data.
+    """
+    elisa_manifest: LocalDataset = None
+    with open(file=str(elisa_manifest_file), mode='r', encoding='utf-8') as fp:
+        elisa_manifest = load(fp=fp)
+    # We test the dataset loading the QCC manifest.
+    ds = Dataset(ds=elisa_manifest, df=pd.read_csv(str(elisa_data_file), index_col=False))
+
+    # Data has 3 data points per lot and run
+    data = ds.data(qtype='Lot1')
+    assert data.shape[0] == 15
+    data = ds.data(qtype='Lot1', context=None)
+    assert data.shape[0] == 15
+    data = ds.data(qtype='Lot1', context='__ALL__')
+    assert data.shape[0] == 15
+
+    data = ds.data(qtype='Lot2', context='Run1')
+    assert data.shape[0] == 3
+
+    with raises(Exception, match='The context "FOOBAR" is not known.'):
+        ds.data(qtype='Lot1', context='FOOBAR')
+    
+    # test sub-sampling:
+    data = ds.data(qtype='Lot5', sub_sample=10)
+    assert data.shape[0] == 10
+
+
+def test_Dataset_transform():
[…]
