### 2023-02-17 838819a
Fix potential problems in isint() and isnumeric()
