### 2023-02-17 35a833b
Do not format bytes to float
