### 2023-02-17 8947e46
Add test-suite for cli helpers
