### 2023-02-17 b4b91cf
Remove class Lazy<T> as it's no longer needed
