+++ src/metrics_as_scores/tools/lazy.py
@@ -8,99 +8,6 @@ from gc import collect
 T = TypeVar('T')
 
 
-class Lazy(Generic[T]):
-    """
-    Helper class to lazily compute a value, thread-safe.
-    """
-    def __init__(self, fn_create_val: Callable[[], T]) -> None:
-        """
-        fn_create_val: ``Callable[[], T]``
-            Function that produces the desired value.
-        """
-        self._val: T = None
-        self._has_val = False
-        self._semaphore = Semaphore(1)
-        self._fn_create_val = fn_create_val
-        self._tpe = ThreadPoolExecutor(max_workers=1)
-    
-    def unset_value(self) -> Self:
-        """
-        Used to explicitly unset (destroy) the previously produced value.
-        This method is thread-safe.
-        """
-        try:
-            self._semaphore.acquire()
-            self._val = None
-            self._has_val = False
-        finally:
-            self._semaphore.release()
-
-        return self
-
-    @property
-    def has_value_volatile(self) -> bool:
-        """
-        Checks, in a volatile manner, if a value is present.
-        """
-        return self._has_val
-
-    @property
-    def has_value(self) -> bool:
-        """
-        Checks whether a value is present, after having acquired the lock.
-        Note that the value may be gone after method exit anyway.
-        """
-        try:
-            self._semaphore.acquire()
-            return self._has_val
-        finally:
-            self._semaphore.release()
-
-    @property
-    def value_volatile(self) -> Union[None, T]:
-        """
-        Returns, in a volatile manner, the value. The value may or may not exist.
-        """
-        return self._val
-
-    @property
-    def value(self) -> T:
-        """
-        Gets the value. If it was not produced previously, it will be produced first.
-        Therefore, this getter is guaranteed to return a value (provided the factory
-        of the value does not throw).
-        """
-        try:
-            self._semaphore.acquire()
-            if not self._has_val:
-                self._val = self._fn_create_val()
-                self._has_val = True
-            return self._val
-        finally:
-            self._semaphore.release()
-
-    @property
-    def value_future(self) -> Future[T]:
-        """
-        Gets an awaitable :py:class:`Future` that will yield the produced value
-        once it is available.
-        """
-        f = Future()
-
-        temp = self._val
-        if type(temp) is T and self.has_value_volatile:
-            f.set_result(temp)
-        else:
-            def set_val():
-                try:
-                    f.set_result(self.value)
-                except Exception as e:
-                    f.set_exception(e)
-            self._tpe.submit(set_val)
-
-        return f
-
-
 class SelfResetLazy(Generic[T]):
     """
     Similar to :py:class:`Lazy`, this class also automatically destroys its value
