### 2023-02-17 7bf68f0
Make value_volatile a property
