### 2023-02-17 b49e6fe
Add test-suite for SelfResetLazy
