### 2023-02-17 6fe5349
Refactor getting the ELISA dataset
