### 2023-02-17 24e99cf
Add test-suite for pregenerate_distns
