### 2023-02-17 2ffbd8f
Add test-suite for pregenerate_fit
