### 2023-02-17 027f513
Add test-suite for testing discrete fits (1 example)
