### 2023-02-18 9b20b7b
Fix typo that would lead to basically never returning a non-linear space
