### 2023-02-18 e690464
Add a num_jobs parameter with default None (=-1)
