### 2023-02-18 2b8ccef
Call fits_to_MAS_densities() and check the results
