### 2023-02-18 135ad7f
Add small test-suite for tools/funcs
