### 2023-02-18 45cd2ab
Add test-suite for generate_densities()
