### 2023-02-18 11b5c95
Bump version to v2.2.0 now that we have good testing and documentation
