### 2023-02-18 MrShoenel/metrics-as-scores#1 @MrShoenel
I want to answer some things that were unclear about the following items:

- **Data sharing**: We actually do claim original results. While we used the Qualitas.class corpus of software metrics for the related publication, we claim original results about software metrics and how they are context-sensitive. We took the Qualitas.class corpus and extracted all metric values, then generated parametric fits and empirical densities, conducted ANOVA-, KS2-, and TukeyHSD-tests. The result is a [separate dataset](https://zenodo.org/record/7633950). It can be downloaded now using the new text-based user interface. It contains also the results of the statistical tests and a template that can be rendered to give the exact original results as claimed in the paper (the rendered report can be seen if you follow the link).
- **Reproducibility**: It is straightforward to reproduce all intermediate and aggregated results. For example, you can download the referenced dataset and use the new text-based user interface to generate a dataset from scratch using the contained CSV that can be used by Metrics As Scores, including all statistical tests and the report.
