### 2023-02-18 afdd62d
Add more tests for the top-level functions
