### 2023-02-18 69c11f8
Add PDF output and Zenodo badge as EPS
