+++ src/metrics_as_scores/cli/LocalDatasets.py
@@ -6,7 +6,7 @@ from metrics_as_scores.distribution.distribution import LocalDataset
 class LocalDatasetsWorkflow(Workflow):
     __doc__ = '''
 This workflow lists all locally available datasets. This includes downloaded
-and installed datasets, as well manually created datasets.
+and installed datasets, as well as manually created datasets.
     '''
 
     def __init__(self) -> None:
@@ -22,7 +22,11 @@ and installed datasets, as well manually created datasets.
         """Main entry point for this workflow."""
         self._print_doc()
 
-        for jsd in get_local_datasets():
+        local_datasets = list(get_local_datasets())
+        if len(local_datasets) == 0:
+            self.q.print('\nThere are no local datasets available! You can create or download one.', style=self.style_err)
+
+        for jsd in local_datasets:
             self.q.print('\nDataset:')
             self.q.print(10*'-')
             self._print_json_dataset(jsd=jsd)
