### 2023-02-18 e1f5e3e
Show progress when downloading known datasets
