+++ src/metrics_as_scores/cli/Download.py
@@ -1,8 +1,10 @@
 from metrics_as_scores.__init__ import DATASETS_DIR
 from metrics_as_scores.cli.Workflow import Workflow
 from metrics_as_scores.cli.helpers import get_known_datasets, get_local_datasets, KNOWN_DATASETS_FILE
-from requests import get
-from shutil import copyfileobj, unpack_archive
+from shutil import unpack_archive
+from wget import download
+from rich.progress import BarColumn, MofNCompleteColumn, Progress, TextColumn, TimeElapsedColumn, TimeRemainingColumn
+
 
 
 class DownloadWorkflow(Workflow):
@@ -35,15 +37,20 @@ Known datasets are loaded from: {KNOWN_DATASETS_FILE}
         dataset_dir = DATASETS_DIR.joinpath(f'./{use_ds["id"]}')
         dataset_dir.mkdir(exist_ok=False)
 
-        self.print_info(text_normal='Downloading archive from: ', text_vital=use_ds['download'])
+        self.print_info(text_normal='Downloading archive from: ', text_vital=f"{use_ds['download']}\n", arrow='\n -> ')
         zip_file = dataset_dir.joinpath('./dataset.zip')
-        with get(url=use_ds['download'], stream=True) as resp:
-            with open(file=str(zip_file), mode='wb') as fp:
-                copyfileobj(fsrc=resp.raw, fdst=fp)
+        
+        with Progress(TextColumn("[progress.percentage]{task.percentage:>3.0f}%"), BarColumn(), MofNCompleteColumn(), TextColumn("MB -"), TimeElapsedColumn(), TextColumn("-"), TimeRemainingColumn()) as progress:
+            task = progress.add_task('[darkyellow]Downloading ...', total=int(round(use_ds['size'] / 1e6)))
+            def update(current_bytes: int, total_bytes: int, width: int):
+                progress.update(task_id=task, completed=float(current_bytes) / 1e6)
+            download(url=use_ds['download'], out=str(zip_file), bar=update)
+
         self.q.print('Download complete. Extracting ...')
 
         unpack_archive(filename=str(zip_file), extract_dir=str(dataset_dir))
-        self.q.print('Done! You can now use this dataset!')
+        self.q.print('\nDone! You can now use this dataset!\n')
+        self.q.print(10*'-' + '\n')
         zip_file.unlink()
 
     
