+++ src/metrics_as_scores/datasets/_default/About.qmd
@@ -61,7 +61,7 @@ dm(f"""
 ```
 
 
-This dataset has the following **Quantity Types**:
+This dataset has the following **Features**:
 
 ```{python}
 #| echo: false
@@ -83,7 +83,7 @@ contexts = list(dataset.contexts(include_all_contexts=False))
 contexts.sort(key=natsort)
 
 dm(f'''
-It has a total of {len(contexts)} **Contexts**: {format_comma_and(list([f'*{c}*' for c in contexts]))}.''')
+It has a total of {len(contexts)} **Groups**: {format_comma_and(list([f'*{c}*' for c in contexts]))}.''')
 ```
 
 
@@ -91,25 +91,25 @@ It has a total of {len(contexts)} **Contexts**: {format_comma_and(list([f'*{c}*'
 # Analysis
 
 In this section, results for the analysis of variance (ANOVA) and Tukey's Honest Significance Test (TukeyHSD) are shown.
-These tests will give a first indication as to how different the quantity types are across contexts.
+These tests will give a first indication as to how different the features are across groups.
 These two tests were used in the original paper [@honel2022qrs] that *Metrics As Scores* was initially conceived for.
 
 
 These tests are conducted to help answering related questions, such as:
 
-- Are there significant statistical differences for each type of quantity across all contexts?
-- Is each context in its entirety (i.e., considering all types of quantities) distinguishable from the other contexts?
-- What are good/bad or common/extreme scores for each context of the given dataset?
+- Are there significant statistical differences for each feature across all groups?
+- Is each group in its entirety (i.e., considering all features) distinguishable from the other groups?
+- What are good/bad or common/extreme scores for each group of the given dataset?
 
 
 ## ANOVA
 
 This test analyzes the differences among means [@Chambers2017-fu].
-For each type of quantity, this test analyzes if means of its samples are significantly different across contexts.
+For each feature, this test analyzes if means of its samples are significantly different across groups.
 The null hypothesis of this test is that there are *no* significant differences.
 This test yields a p-value and an F-statistic.
 The latter is the mean square of each independent variable divided by the mean square of the residuals.
-Large F-statistics indicate that the variation among contexts is likely.
+Large F-statistics indicate that the variation among groups is likely.
 The p-value then indicates how likely it is for the F-statistic to have occurred, given the null hypothesis is true.
 
 
@@ -120,14 +120,14 @@ KS2 does not check for a certain type of probability distribution since it uses 
 Its test statistic is the maximum vertical distance between the two CDFs.
 For two samples $\mathbf{x},\mathbf{y}$, the statistic is calculated as $D_{\mathbf{x},\mathbf{y}}\tight{=}\sup_t\,\abs{F_{\mathcal{X}}(t)\tight{-}F_{\mathcal{Y}}(t)}$.
 The null hypothesis is that the samples' CDFs are identical, that is, $F_{\mathcal{X}}\tight{=}F_{\mathcal{Y}}$.
-This test is used to compare one type of quantity between two contexts.
+This test is used to compare one feature between two groups.
 
 
 ## TukeyHSD
 
 This test is used to gain insights into the results of an ANOVA test.
 While the former only allows obtaining the amount of corroboration for the null hypothesis, TukeyHSD performs all pairwise comparisons [@tukey1949hsd].
-For example, by choosing a certain type of quantity and context, we obtain a list of other contexts that are significantly statistically different.
+For example, by choosing a certain feature and group, we obtain a list of other groups that are significantly statistically different.
[…]
