### 2023-02-19 8e73049
Use the words feature and group instead of quantity type and context in CLI workflows
