### 2023-02-20 2413bef
Add a bunch of pragma: no cover for code we cannot possibly test
