### 2023-02-20 9ffbfc8
Add a _val_type for proper volatile type/instance checking
