### 2023-02-20 a6c4764
Add more testing for generating empirical discrete densities
