### 2023-02-20 0a09873
Test that exceptions are raised for empty grids
