### 2023-02-20 fbfb6ee
Add extra tests for Fitter and FitterPymoo
