### 2023-02-20 cdf4a53
Test nonlinspace() and natsort()
