### 2023-02-20 d85946d
Test all discrete fitting problems
