### 2023-02-20 32ffdbe
Update coverage action to use token explicitly; also use v3
