+++ .github/workflows/codecov.yml
@@ -12,9 +12,11 @@ jobs:
     - uses: actions/checkout@v2.3.4
       with:
         fetch-depth: '2'
+    - name: Install Dependencies
+      run: pip install .
     - name: Generate Report
-      run: |
-        pip install .
-        pytest --cov-report html
+      run: pytest --cov-report html
     - name: Upload Coverage to Codecov
-      uses: codecov/codecov-action@v2
+      uses: codecov/codecov-action@v3
+      with:
+        token: ${{ secrets.CODECOV_TOKEN }}
