### 2023-02-20 d40128e
Add pragma: no cover for defining root variables
