### 2023-02-20 862b127
Remove try/catch from opening manifests and let errors propagate
