### 2023-02-20 e1fac7f
Make a proper test for local datasets, finally
