+++ paper/paper.md
@@ -27,6 +27,7 @@ bibliography: refs.bib
 date: "29 September 2022"
 tags: 
   - Python
+  - "Multiple ANOVA"
   - "Distribution fitting"
   - "Inverse sampling"
   - "Empirical distributions"
@@ -45,67 +46,92 @@ title: "Metrics As Scores: A Tool- and Analysis Suite and Interactive Applicatio
 # Summary
 <!-- A summary describing the high-level functionality and purpose of the software for a diverse, non-specialist audience. -->
 <!-- Mention (if applicable) a representative set of past or ongoing research projects using the software and recent scholarly publications enabled by it. -->
-
 <!-- Short introduction into the problem, then -->
-Raw data can be aggregated according to some well-defined standard, method, or calculation, which makes it a metric.
+
+\emph{Metrics As Scores} can be thought of as an interactive, \emph{multiple} analysis of variance (ANOVA; @chambers2017statistical).
+An ANOVA might be used to estimate the goodness-of-fit of a statistical model.
+Beyond ANOVA, which is used to analyze the differences among hypothesized group means for a single feature, Metrics As Scores seeks to answer the question of whether a sample of a certain quantity (feature) is more or less common across groups.
+For each group, we can determine what might constitute a good/bad, acceptable/alarming, or common/extreme value, and how distant the sample is from that value.
+This is expressed in terms of a percentile (a standardized scale of $\left[0,1\right]$), which we call \textbf{\emph{score}}.
+Considering all available features among the existing groups furthermore allows to assess how different the groups are from each other, or whether they are indistinguishable from one another.
+
+
+The name \emph{Metrics As Scores} was derived from its initial application: Examining differences of software metrics across application domains [@honel2022mas].
+A software metric is an aggregation of a raw quantity according to some well-defined standard, method, or calculation.
 In software processes, such aggregations are often counts of events or certain properties [@carleton1999].
-However, without the aggregation that is done in a quality model, raw data and metrics are rarely of great value to analysts and decision-makers. This is because quality models are conceived to establish a connection between metrics and certain quality goals [@kaner2004software].
+However, without the aggregation that is done in a quality model, raw data and software metrics are rarely of great value to analysts and decision-makers. This is because quality models are conceived to establish a connection between software metrics and certain quality goals [@kaner2004software].
 It is, therefore, difficult to answer the question ``is my metric value good?''.
-With \emph{Metrics As Scores} (\mas) we present an approach that, given some \emph{ideal} value, can transform any metric into a score, given an observation of sufficiently many relevant metrics values.
-While such ideal values were previously attempted to be derived from, e.g., experience or surveys [@benlarbi2000thresh], benchmarks [@alves2010thresh], or by setting practical values [@grady1992practical], with \mas we suggest deriving ideal values additionally in non-parametric, statistical ways.
-To do so, data first needs to be captured in a \emph{relevant} context. A metric value might be good in one context, while it is less so in another.
-Therefore, we suggest generalizing and contextualizing the approach taken by [@UlanLEW21], in which a score is defined to always have a range of $[0,1]$ and linear behavior.
+
+
+With Metrics As Scores we present an approach that, given some \emph{ideal} value, can transform any feature into a score, given an observation of sufficiently many relevant values.
+While such ideal values for software metrics were previously attempted to be derived from, e.g., experience or surveys [@benlarbi2000thresh], benchmarks [@alves2010thresh], or by setting practical values [@grady1992practical], with Metrics As Scores we suggest deriving ideal values additionally in non-parametric, statistical ways.
+To do so, data first needs to be captured in a \emph{relevant} context (group).
+A feature value might be good in one context, while it is less so in another.
[…]
