### 2023-02-20 MrShoenel/metrics-as-scores#1 @MrShoenel
I have made changes to the paper and altered the introduction and statement of need.
I hope it is more accessible now since I use the terms *feature* and *group* and make sure that *metrics* refers to software metrics.

I will leave the closing of the issues to you, please do so once all criteria are satisfied :)
