### 2023-02-20 #4913 @MrShoenel
Thanks for your patience. I have made substantial changes in the meantime (please check the README.md), to address all of @mdhaber's comments (thanks for these!).
I did update the paper, trying to make it more accessible.

My to-do list is now empty :) Of course, I absolutely welcome additional comments :)
