### 2023-02-20 MrShoenel/metrics-as-scores#1 @mdhaber
1. Great.

Re: 2-4, in one of your replies:

> I have shared **the paper** privately with you for the time being.

This suggests that your responses may have been partially about the paper sent by email rather than this JOSS paper. I'll try rephrasing my questions to make sure it is clear which paper we're discussing, and I'd appreciate it if you'd rephrase your responses accordingly.

2. Is there any original data that is claimed as part of this JOSS paper that is not published otherwise? If so, please let me know when all of it is mentioned in the JOSS paper so that I can confirm that it is accessible.
3. Are there results (e.g. conclusions at https://metrics-as-scores.ml/webapp) claimed as part of this JOSS paper, or does this JOSS paper primarily about the the code behind the results? (If the results are claimed as part of this JOSS paper, what is the relationship between this JOSS paper and the paper you shared by email? In other words, can you contrast the contributions of the two? Note that if "results" are claimed in this JOSS paper, they all need to be "entirely reproducible by reviewers" - how would I do that?)
4. Does this JOSS paper make any performance claims ("fast", "memory-efficient", "accurate") about the software?

Some new questions so I can check off other boxes:

5. What work toward this JOSS paper precedes your first commit to the repository? (i.e. how long did it take to write the first commit, or were there other draft versions of the software that weren't committed)
6. Can you point to the part of the paper that describes "how this software compares to other commonly-used packages?"
