### 2023-02-20 MrShoenel/metrics-as-scores#1 @MrShoenel
2. Yes, after adapting Metrics As Scores so it can work with arbitrary datasets properly, I have produced [**3** datasets](https://github.com/MrShoenel/metrics-as-scores#known-datasets), one of which is now referenced in the **JOSS** paper and claimed as original data in there (this one: _Metrics and Domains From the Qualitas.class corpus_. <https://doi.org/10.5281/zenodo.7633949>. The other two datasets, while original, are not mentioned or referenced in the paper.
3. This JOSS paper is primarily about the code behind the results. There are no results claimed as part of this JOSS paper. In the section "Application", however, I do refer briefly to some results that were claimed in the referred QRS paper (that I shared with you privately). The relationship between the QRS and the JOSS paper is this:
    - The QRS paper makes use of the web application to derive some original results from the Qualitas.class corpus. Those results are only claimed there, and there is a brief reference to them at the end of the JOSS paper.
    - The JOSS paper introduces the application and analysis suite called _Metrics As Scores_, but other than the software package, it does not claim any original results.
    - As part of improving the application (making it dataset-agnostic), the dataset that was derived from the Qualitas.class corpus was externalized and published separately. It is referenced in the JOSS paper. There are **no** new results claimed in the JOSS paper, just the dataset as its own publication.
4. [see the previous response; it was about the JOSS paper]
5. There were no preceding draft versions. While the first commit starts with an empty repository, a few hours were invested prior to this for obtaining the Qualitas.class corpus and transforming it into a dataset that can be used with the to-be-implemented application. I do not have precise estimates for that, but suppose that it was approx. a man-day.
6. I am not aware of any other packages that could be compared to this one. However, the first two paragraphs in the section "MAS -- The Tool- and Analysis Suite" describe the usage of common analyses that can be used to partially achieve what Metrics As Scores can achieve. In the section "Statement Of Need" it is pointed out that the software enables a novel way of exploring differences among groups.
