### 2023-02-20 MrShoenel/metrics-as-scores#1 @mdhaber
Thanks, I think those questions are answered, and I checked off several boxes. I've created two new bug reports and replied to some of the existing issues.
