### 2023-02-21 9f24981
Move getting and validating required files for own datasets to helpers
