+++ src/metrics_as_scores/cli/BundleOwn.py
@@ -1,10 +1,8 @@
-import pandas as pd
 from os.path import commonpath
 from metrics_as_scores.__init__ import DATASETS_DIR
 from metrics_as_scores.cli.Workflow import Workflow
-from metrics_as_scores.cli.helpers import get_local_datasets
-from metrics_as_scores.distribution.distribution import DistTransform, Parametric, Parametric_discrete, Empirical, Empirical_discrete, KDE_approx, Dataset, LocalDataset
-from itertools import product
+from metrics_as_scores.cli.helpers import get_local_datasets, required_files_folders_local_dataset, validate_local_dataset_files, PathStatus
+from metrics_as_scores.distribution.distribution import LocalDataset
 from pathlib import Path
 from zipfile import ZipFile, ZIP_DEFLATED
 
@@ -25,33 +23,33 @@ For an example dataset, check out https://doi.org/10.5281/zenodo.7633950
     def __init__(self) -> None:
         super().__init__()
         self.use_ds: LocalDataset = None
-        self.ds: Dataset = None
         self.ds_dir: Path = None
-
-        self.dirs_required: list[Path] = None
-        self.files_required: list[Path] = None
     
 
-    def _check_files(self) -> None:
-        for d in self.dirs_required:
-            self.print_info(text_normal='Checking directory: ', text_vital=str(d))
-            if not d.exists():
-                raise Exception(f'Directory does not exist: {str(d)}')
-            elif not d.is_dir():
-                raise Exception(f'Not a directory: {str(d)}')
+    def _check_dirs_and_files(self) -> None:
+        dirs_required, files_required = required_files_folders_local_dataset(local_ds_id=self.use_ds['id'])
+        dirs_dict, files_dict = validate_local_dataset_files(dirs=dirs_required, files=files_required)
+
+        for dir in dirs_dict.keys():
+            self.print_info(text_normal='Checking directory: ', text_vital=str(dir))
+            if dirs_dict[dir] == PathStatus.DOESNT_EXIST:
+                raise Exception(f'Directory does not exist: {str(dir)}')
+            elif dirs_dict[dir] == PathStatus.NOT_A_DIRECTORY:
+                raise Exception(f'Not a directory: {str(dir)}')
         
-        for f in self.files_required:
-            self.print_info(text_normal='Checking file: ', text_vital=str(f))
-            if not f.exists():
-                raise Exception(f'File does not exist: {str(f)}')
-            elif not f.is_file():
-                raise Exception(f'Not a file: {str(f)}')
+        for file in files_dict.keys():
+            self.print_info(text_normal='Checking file: ', text_vital=str(file))
+            if files_dict[file] == PathStatus.DOESNT_EXIST:
+                raise Exception(f'File does not exist: {str(file)}')
+            if files_dict[file] == PathStatus.NOT_A_FILE:
+                raise Exception(f'Not a file: {str(file)}')
     
 
     def _make_zip(self) -> Path:
         zip_file = self.ds_dir.joinpath('./dataset.zip')
+        _, files_required = required_files_folders_local_dataset(local_ds_id=self.use_ds['id'])
         with ZipFile(file=str(zip_file), mode='x', compression=ZIP_DEFLATED, compresslevel=9) as zf:
-            for path in self.files_required:
+            for path in files_required:
                 fs_path = str(path.resolve())
                 base = commonpath([fs_path, str(self.ds_dir.resolve())])
                 zf.write(filename=fs_path, arcname=fs_path.removeprefix(base))
@@ -68,42 +66,9 @@ For an example dataset, check out https://doi.org/10.5281/zenodo.7633950
[…]
