### 2023-02-21 b425122
Make the exception message for missing densities more helpful
