+++ src/metrics_as_scores/webapp/app_hooks.py
@@ -18,7 +18,7 @@ def unpickle(file: str):
         with open(file=file, mode='rb') as f:
             return load(f)
     except Exception as e:
-        raise Exception('The webapp relies on precomputed results. Please generate them using the file pregenerate.py before running this webapp.') from e
+        raise Exception('The webapp requires pre-generated densities. You can generate these for your own datasets by using the corresponding wizard of the TUI before running this webapp.') from e
 
 
 def load_data(dataset_id: str, preload: bool=False):
