### 2023-02-21 95f7ad6
Fix a problem where the p-value and test statistic were accessed wrong
