### 2023-02-21 6f6b529
Add a new line before showing the main menu
