+++ src/metrics_as_scores/cli/MainWorkflow.py
@@ -31,6 +31,7 @@ class MainWorkflow(Workflow):
         """
 
         # The main options/Functions for M-a-S:
+        self.q.print('')
         res = self.askt(options=[
             ('Show Installed Datasets', 'show_local'),
             ('Show List of Known Datasets Available Online That Can Be Downloaded', 'show_known'),
