### 2023-02-21 cea29d5
Observe successful start of process and only allow internal if installed locally
