+++ src/metrics_as_scores/cli/LocalWebserver.py
@@ -1,15 +1,23 @@
 from concurrent.futures import ThreadPoolExecutor
+from threading import Semaphore, Thread
+from typing import Literal, Union
 from bokeh.command.util import build_single_handler_application
 from bokeh.server.server import Server
 from pathlib import Path
+from metrics_as_scores.__init__ import IS_MAS_LOCAL
 from metrics_as_scores.cli.Workflow import Workflow
-from metrics_as_scores.cli.helpers import isint, get_local_datasets
+from metrics_as_scores.cli.helpers import PathStatus, isint, get_local_datasets, required_files_folders_local_dataset, validate_local_dataset_files
 from metrics_as_scores.distribution.distribution import LocalDataset
 from questionary import Choice
 
 this_dir = Path(__file__).resolve().parent
 webapp_dir = this_dir.parent.joinpath('./webapp')
 root_dir = this_dir.parent.parent.parent
+
+
+
+
+
 
 class LocalWebserverWorkflow(Workflow):
     __doc__ = '''
@@ -19,53 +27,109 @@ Metrics As Scores, using one of the locally available datasets.
     def __init__(self) -> None:
         super().__init__()
         self.preload: bool = None
+        self.port: int = None
+        self.use_ds: LocalDataset = None
     
 
     def _port_and_dataset(self) -> tuple[int, LocalDataset]:
+        class LocalDatasetWithMissingFiles(LocalDataset):
+            missing_files: list[Path]
+        
         def check_port(p: str) -> bool:
             if not isint(p):
                 return False
             r = int(p)
             return r > 0 and r < 65536
+        
+        def check_files(ds: LocalDatasetWithMissingFiles) -> LocalDatasetWithMissingFiles:
+            ds['missing_files'] = []
+            _, files = validate_local_dataset_files(*required_files_folders_local_dataset(local_ds_id=ds['id']))
 
-        return (
-            int(self.q.text(message='Enter a Port number:', default=f'{5678}', validate=check_port).ask()),
-            self.q.select(
+            for f in files:
+                if 'densities_' in f.name and not files[f] == PathStatus.OK:
+                    ds['missing_files'].append(f)
+            
+            return ds
+        
+
+        port = int(self.q.text(message='Enter a Port number:', default=f'{5678}', validate=check_port).ask())
+        ds_list = list([
+            check_files(ds) for ds in get_local_datasets()
+        ])
+        ds_list_missing = list([ds for ds in ds_list if len(ds['missing_files']) > 0])
+        if len(ds_list_missing) > 0:
+            for ds in ds_list_missing:
+                self.print_info(text_normal='The following dataset misses pre-generated densities: ', text_vital=f'{ds["name"]} [{ds["id"]}]', arrow='')
+                self.print_info(text_normal='The following densities are missing: ', text_vital=', '.join(list([f.name for f in ds['missing_files']])))
+        
+        ds_list_OK = list([ds for ds in ds_list if len(ds['missing_files']) == 0])
+        use_ds: LocalDataset = None
+        if len(ds_list_OK) == 0:
+            raise Exception('There are no datasets available or none of the datasets is complete.')
+        else:
+            use_ds = self.q.select(
                 message='Please choose one of the following Datasets:',
-                choices=list([Choice(title=f"{ds['name']} [{ds['id']}]", value=ds) for ds in get_local_datasets()])).ask())
+                choices=list([Choice(title=f"{ds['name']} [{ds['id']}]", value=ds) for ds in ds_list_OK])).ask()
+        
+        return port, use_ds
     
-    def _type_quit_to_exit(self) -> None:
[…]
