### 2023-02-21 206ac57
Change the description to make clear what from_dict() does
