### 2023-02-21 5aba9f1
Add checks to constructor of Parametric to only accept proper stat-tests
