### 2023-02-21 c5ee424
Properly test stat-tests of Parametric and Parametric_discrete
