### 2023-02-21 5ccbb17
Ignore decoding errors in process pipe
