+++ src/metrics_as_scores/tools/__init__.py
@@ -1,0 +1,3 @@
+"""
+This package contains tools that are commonly used in Metrics As Scores.
+"""
