+++ src/metrics_as_scores/cli/__init__.py
@@ -1,0 +1,6 @@
+"""
+This package contains the text-based user interface of Metrics As Scores.
+It is implemented as a command line interface (CLI). The main menu consists
+of numerous workflows for the user to choose from. Please refer to each
+workflow individually to understand what it does.
+"""
