+++ src/metrics_as_scores/data/__init__.py
@@ -1,0 +1,6 @@
+"""
+This package contains functions that are needed for the mass-wise fitting
+of continuous and discrete random variables to own data, as well as functions
+for pre-generating densities for own datasets that are then used in the
+interactive web application.
+"""
