### 2023-02-21 7a766f1
Add some commonly used poe-tasks to pyproject.toml
