+++ src/metrics_as_scores/cli/CreateDataset.py
@@ -8,6 +8,7 @@ from metrics_as_scores.cli.Workflow import Workflow
 from metrics_as_scores.cli.helpers import isint, isnumeric, get_known_datasets
 from metrics_as_scores.distribution.distribution import Dataset, LocalDataset
 from metrics_as_scores.tools.funcs import natsort
+from questionary import Choice
 from re import match
 from os import makedirs
 from json import dump
@@ -32,6 +33,7 @@ For a single sample, one column holds the numeric observation, one holds the
 ordinal type (name of feature), and one holds the group associated with it. A
 dataset can hold one or more features, but should hold at least two groups in
 order to compare distributions of a single sample type across groups.
+
 This workflow creates the entire dataset: The manifest JSON, the parametric fits,
 the pre-generated distributions. When done, the dataset is installed, such that it
 can be discovered and used by the local Web-Application. If you wish to publish
@@ -69,6 +71,51 @@ previous menu afterwards.
         
         return type_dict
 
+    def _transform_dataset(self) -> pd.DataFrame:
+        self.q.print('\n' + 10*'-')
+        file_type = self.ask(options=[
+            'CSV', 'Excel'
+        ], prompt='What kind of file do you want to read?', rtype=str)
+
+        path = self.q.text(message='Absolute file path or URL to original file:', validate=lambda s: len(s) > 0).ask().strip()
+        df: pd.DataFrame = None
+        if file_type == 'CSV':
+            df = self._read_csv(path_like=path)
+        elif file_type == 'Excel':
+            df = self._read_excel(path_like=path)
+        
+        available_cols = OrderedDict({ k: k for k in df.columns.values.tolist() })
+        if len(available_cols) < 2:
+            raise Exception(f'Need at least one group and one feature. The loaded data frame only has these columns: [{", ".join(available_cols.keys())}]')
+        col_ctx = self.ask(prompt='What is the name of the groups\' column?', options=list(available_cols.keys()), rtype=str)
+        del available_cols[col_ctx]
+
+        df_dtypes = df.dtypes.to_dict()
+        col_feats = self.q.checkbox(message='Please select all features you would like to include.', choices=[
+            Choice(title=f'{feat} [{df_dtypes[feat]}]', value=feat) for feat in available_cols.keys()
+        ], validate=lambda str_list: len(str_list) > 0).ask()
+
+        self.print_info(text_normal='Converting data frame ', text_vital='...')
+        # Let's stack the data frame:
+        target_df = pd.DataFrame(dict(
+            Feature = [],
+            Group = [],
+            Value = []
+        ))
+
+        nrow = len(df.index)
+        for col_feat in col_feats:
+            target_df = pd.concat([target_df, pd.DataFrame(dict(
+                Feature = nrow * [col_feat],
+                Group = df[col_ctx],
+                Value = df[col_feat]
+            ))])
+        
+        return target_df
+
+        
+
+
     def _read_csv(self, path_like: str) -> pd.DataFrame:
         sep = self.q.text(message='What is the separator used?', default=',', validate=lambda s: s in [',', ';', ' ', '\t']).ask()
         dec = self.q.text(message='What is the decimal point?', default='.', validate=lambda s: s in ['.', ',']).ask()
@@ -87,33 +134,46 @@ previous menu afterwards.
         # Check if all values are integer.
         return np.all(np.mod(vals, 1) == 0)
 
-    def _create_manifest(self) -> tuple[LocalDataset, pd.DataFrame]:
+    def _create_manifest(self, transformed_df: pd.DataFrame=None) -> tuple[LocalDataset, pd.DataFrame]:
         jsd: LocalDataset = {}
 
         df: pd.DataFrame = None
[…]
