### 2023-02-22 e51cc4a
Make helper function for transforming to MAS-compatible data frame
