### 2023-02-22 921c624
Also raise if group is not present in data frame
