### 2023-02-22 71f4bb9
Use helper function for transforming dataset to MAS-compatible
