### 2023-02-22 aaa8c85
Refactor workflow so that it can be used without CLI
