+++ src/metrics_as_scores/cli/CreateDataset.py
@@ -41,10 +41,64 @@ your dataset so that it can be used by others, start the Bundle-workflow from th
 previous menu afterwards.
     """.strip()
 
-    def __init__(self) -> None:
+    def __init__(self, manifest: LocalDataset=None, org_df: pd.DataFrame=None) -> None:
         super().__init__()
         self.regex_id = r'^[a-z]+[a-z-_\d]*?[a-z\d]+?$'
+        self.manifest: LocalDataset = manifest
+        self.org_df: pd.DataFrame = org_df
+        self.dataset: Dataset = None if manifest is None and org_df is None else Dataset(ds=manifest, df=org_df)
     
+
+    @property
+    def dataset_dir(self) -> Path:
+        if self.manifest is None:
+            raise Exception('A manifest is required.')
+        return DATASETS_DIR.joinpath(f'./{self.manifest["id"]}')
+    
+
+    @property
+    def fits_dir(self) -> Path:
+        return self.dataset_dir.joinpath('./fits')
+    
+
+    @property
+    def densities_dir(self) -> Path:
+        return self.dataset_dir.joinpath('./densities')
+    
+
+    @property
+    def tests_dir(self) -> Path:
+        return self.dataset_dir.joinpath('./stat-tests')
+    
+
+    @property
+    def web_dir(self) -> Path:
+        return self.dataset_dir.joinpath('./web')
+    
+
+    @property
+    def path_manifest(self) -> Path:
+        return self.dataset_dir.joinpath('./manifest.json')
+    
+
+    @property
+    def path_df_data(self) -> Path:
+        return self.dataset_dir.joinpath('./org-data.csv')
+
+
+    @property
+    def path_test_ANOVA(self) -> Path:
+        return self.tests_dir.joinpath('./anova.csv')
+    
+
+    @property
+    def path_test_TukeyHSD(self) -> Path:
+        return self.tests_dir.joinpath('./tukeyhsd.csv')
+    
+
+    @property
+    def path_test_ks2samp(self) -> Path:
+        return self.tests_dir.joinpath('./ks2samp.csv')
 
     def _set_ideal_values(self, qtypes: list[str]) -> dict[str, Union[float,int]]:
         type_dict = {}
@@ -70,6 +124,7 @@ previous menu afterwards.
             ready = not self.q.confirm(message='Continue setting ideal values?', default=True).ask()
         
         return type_dict
+
 
     def _transform_dataset(self) -> pd.DataFrame:
         self.q.print('\n' + 10*'-')
@@ -99,14 +154,13 @@ previous menu afterwards.
 
         return transform_to_MAS_dataset(df=df, group_col=col_ctx, feature_cols=col_feats)
 
-        
-
 
     def _read_csv(self, path_like: str) -> pd.DataFrame:
         sep = self.q.text(message='What is the separator used?', default=',', validate=lambda s: s in [',', ';', ' ', '\t']).ask()
         dec = self.q.text(message='What is the decimal point?', default='.', validate=lambda s: s in ['.', ',']).ask()
         return pd.read_csv(filepath_or_buffer=path_like, sep=sep, decimal=dec)
-    
+
+
     def _read_excel(self, path_like: str) -> pd.DataFrame:
         sheet = self.q.text(message='Which sheet would you like to read?', default='0').ask()
         if isint(sheet):
@@ -114,11 +168,13 @@ previous menu afterwards.
         header = self.q.text('Zero-based index of header column?', default='0', validate=isint).ask()
         return pd.read_excel(io=path_like, sheet_name=sheet, header=header)
 
+
     def _is_feature_discrete(self, df: pd.DataFrame, col_data: str, col_type: str, use_type: str) -> bool:
         temp = df.loc[df[col_type] == use_type, :]
[…]
