### 2023-02-22 3f9894b
Refactor workflow for usage without CLI
