### 2023-02-22 ca616db
Fix a bug related to printing
