### 2023-02-22 4c274ae
Make sure the range for stepping is positive
