### 2023-02-22 3743bf2
Remove overridden _min_max() method in Parametric
