### 2023-02-22 d9de6ba
Change implementation of _min_max() to limit f(x) rather than x
