### 2023-02-22 9115c5d
Override computing of practical domain to exploit parametric PPF and extended cutoff
