### 2023-02-22 8ef9d16
Return the data's range in Parametric when not fit
