### 2023-02-22 ccf7435
Add some coverage for fitting with dist transform != NONE
