### 2023-02-22 e51d8d6
Add date in long format to yaml for About.qmd
