+++ src/metrics_as_scores/datasets/_default/_quarto.yml
@@ -7,6 +7,8 @@ format:
   html:
     code-fold: true
   pdf:
+    date: today
+    date-format: long
     documentclass: scrartcl
     geometry: 'a4paper,margin=2.25cm'
     keep-tex: true
