+++ src/metrics_as_scores/datasets/known-datasets.bib
@@ -1,36 +1,43 @@
 @dataset{dataset_qcc,
   author       = {Sebastian Hönel},
-  title        = {{Metrics As Scores Dataset: Metrics and Domains 
-                   From the Qualitas.class corpus}},
-  month        = feb,
+  title        = {{Metrics As Scores Dataset: Metrics and Domains From the Qualitas.class Corpus}},
+  month        = 2,
   year         = 2023,
   publisher    = {Zenodo},
-  version      = {v1.0},
+  version      = {v1.2},
   doi          = {10.5281/zenodo.7633949},
   url          = {https://doi.org/10.5281/zenodo.7633949}
 }
 
 @dataset{dataset_elisa,
   author       = {Sebastian Hönel},
-  title        = {{Metrics As Scores Dataset: Elisa Spectrophotometer 
-                   Positive Samples}},
-  month        = feb,
+  title        = {{Metrics As Scores Dataset: Elisa Spectrophotometer Positive Samples}},
+  month        = 2,
   year         = 2023,
   publisher    = {Zenodo},
-  version      = {v1.0},
+  version      = {v1.2},
   doi          = {10.5281/zenodo.7633989},
   url          = {https://doi.org/10.5281/zenodo.7633989}
 }
 
 @dataset{dataset_diamonds-ideal-if,
   author       = {Sebastian Hönel},
-  title        = {{Metrics As Scores Dataset: Price, weight, and 
-                   other properties of over 1,200 ideal-cut and best-
-                   clarity diamonds}},
-  month        = feb,
+  title        = {{Metrics As Scores Dataset: Price, Weight, and Other Properties of Over 1,200 Ideal-Cut and Best-Clarity Diamonds}},
+  month        = 2,
   year         = 2023,
   publisher    = {Zenodo},
-  version      = {v1.0},
+  version      = {v1.2},
   doi          = {10.5281/zenodo.7647596},
   url          = {https://doi.org/10.5281/zenodo.7647596}
+}
+
+@dataset{dataset_iris,
+  author       = {Sebastian Hönel},
+  title        = {{Metrics As Scores Dataset: The Iris Flower Data Set}},
+  month        = 2,
+  year         = 2023,
+  publisher    = {Zenodo},
+  version      = {v1.1},
+  doi          = {10.5281/zenodo.7669664},
+  url          = {https://doi.org/10.5281/zenodo.7669664}
 }
