### 2023-02-27 307f116
Use unbuffered TextIOWrapper for child process
