### 2023-02-27 fdbb8c3
Acquire semaphore correctly, always kill process, and handle exceptions
