### 2023-02-27 a0061f6
Make the webapp exit forcefully when dataset cannot be loaded
