### 2023-02-28 e20fbbe
Fix the local webserver to use bokeh executable from environment
