### 2023-03-01 67f0eca
Use only one thread for reading both stdin and stderr; also keep on reading and piping
