+++ README.qmd
@@ -210,33 +210,35 @@ For each group, we can determine what might constitute a common/ideal value, and
 This is expressed in terms of a percentile (a standardized scale of `[0,1]`), which we call **score**.
 
 
+## Concrete Example Using the Qualitas.class Corpus Dataset
+
+The notebook [`notebooks/Example-webapp-qcc.ipynb`](https://github.com/MrShoenel/metrics-as-scores/blob/master/notebooks/Example-webapp-qcc.ipynb)
+holds a concrete example for using the web application to interactively obtain **scores**.
+In this example, we create a hypothetical application that ought to be in the application domain *SDK*.
+Using a concrete metric, *Number of Packages*, we find out that our hypothetical new SDK application scores poorly for what it is intended to be.
+
+
+This example illustrates the point that software metrics, when captured out of context, are meaningless [@gil2016software].
+For example, typical values for complexity metrics are vastly different, depending on the type of application.
+We find that, for example, applications of type SDK have a much lower *expected* complexity compared to Games (`1.9` vs. `3.1`) [@honel2022qrs].
+Software metrics are often used in software quality models.
+However, without knowledge of the application's context (here: domain), the deduced quality of these models is at least misleading, if not completely off.
+This becomes apparent if we examine how an application's complexity scores across certain domains.
+
+Since there are many software metrics that are captured simultaneously, we can also compare domains in their entirety:
+How many metrics are statistically significantly different from each other?
+Is there a set of domains that are not distinguishable from each other?
+Are there metrics that are always different across domains and must be used with care?
+In this example, we use a known and downloadable dataset [@dataset_qcc].
+It is based on software metrics and application domains of the "Qualitas.class corpus" [@terra2013qcc; @tempero2010qualitas].
+
+
 ## Concrete Example Using the Iris Dataset
 
 The notebook [`notebooks/Example-create-own-dataset.ipynb`](https://github.com/MrShoenel/metrics-as-scores/blob/master/notebooks/Example-create-own-dataset.ipynb)
 holds a concrete example for creating/importing/using one's own dataset.
 Although all necessary steps can be achieved using the **TUI**, this notebook demonstrates a complete example of implementing this in code.
 
-
-## Software Metrics Example
-
-Software metrics, when captured out of context, are meaningless [@gil2016software].
-For example, typical values for complexity metrics are vastly different, depending on the type of application.
-We find that, for example, applications of type SDK have a much lower *expected* complexity compared to Games (`1.9` vs. `3.1`) [@honel2022qrs].
-
-
-Software metrics are often used in software quality models.
-However, without knowledge of the application's context (here: domain), the deduced quality of these models is at least misleading, if not completely off.
-This becomes apparent if we examine how an application's complexity scores across certain domains.
-
-
-Since there are many software metrics that are captured simultaneously, we can also compare domains in their entirety:
-How many metrics are statistically significantly different from each other?
-Is there a set of domains that are not distinguishable from each other?
-Are there metrics that are always different across domains and must be used with care?
-
-
-This example is available as a downloadable dataset [@dataset_qcc].
-It is based on software metrics and application domains of the "Qualitas.class corpus" [@terra2013qcc; @tempero2010qualitas].
 
 
 ## Diamonds Example
