### 2023-03-01 bd002f5
Increase version to v2.4.2 and re-render readme
