### 2023-03-11 874d0a7
Add a toggle-button for the table and re-arrange the layout slightly
