+++ docs/index.rst
@@ -10,6 +10,8 @@ Welcome to Metrics As Scores's documentation!
    :maxdepth: 2
    :caption: Contents:
 
+   readme_link
+
    modules
 
 
