### 2023-03-13 7a514f8
Strip quotation marks from input path (not needed)
