### 2023-03-13 f2ef058
Make sure the transformed Group column is of type str
