### 2023-03-13 d9c1199
Implement two method for getting and checking the minimum amount of observations
