### 2023-03-13 90d7c2f
Add check for sufficiently many observations in new dataset
