### 2023-03-13 a3f56a8
Print concrete instance name of continuous distributions
