### 2023-03-28 #4913 @mdhaber
@mikldk You can see the status of my review at https://github.com/MrShoenel/metrics-as-scores/issues/1. We've been steadily working through things since mid-December. Most individual issues have been addressed, but a few are still open.

One of the review criteria is "Is the list of references complete...?" The paper cites some of MAS's software dependencies, but not all. How should the author decide which dependencies to cite? I am a maintainer of one such dependency, SciPy, which is used heavily [throughout MAS](https://github.com/MrShoenel/metrics-as-scores/search?q=scipy). Compared to other tools that are cited, it seems to be used enough to deserve a citation, but I'm not sure that I can claim to be completely objective about this, so I'd appreciate your input.
