### 2023-04-11 #4913 @mikldk
@mdhaber I think a citation to SciPy is in order, so please go ahead and add it.
