### 2023-04-13 #4913 @MrShoenel
I agree, too. I will address the other editing-related issues and add a proper citation to SciPy in a matching place.
