+++ pyproject.toml
@@ -1,6 +1,6 @@
 [tool.poetry]
 name = "metrics-as-scores"
-version = "2.5.1"
+version = "2.5.2"
 description = "Interactive web application, tool- and analysis suite for approximating, exploring, understanding, and sampling from conditional distributions."
 authors = ["Sebastian Hönel <development@hoenel.net>"]
 license = "Dual-licensed under GNU General Public License v3 (GPLv3) and closed-source"
@@ -58,7 +58,7 @@ myst-parser = "^1.0.0"
 [tool.poetry.group.dev.dependencies]
 twine = "^4.0.1"
 build = "^0.8.0"
-poetry = "^1.2.1"
+poetry = "^1.3.2"
 pytest = "^7.1.3"
 coverage = "^6.5.0"
 
