+++ README.md
@@ -1,50 +1,38 @@
-Metrics As Scores
+# Metrics As Scores
 [![DOI](https://zenodo.org/badge/524333119.svg)](https://zenodo.org/badge/latestdoi/524333119)
 [![status](https://joss.theoj.org/papers/eb549efe6c0111490395496c68717579/status.svg)](https://joss.theoj.org/papers/eb549efe6c0111490395496c68717579)
 [![codecov](https://codecov.io/github/MrShoenel/metrics-as-scores/branch/master/graph/badge.svg?token=HO1GYXVEUQ)](https://codecov.io/github/MrShoenel/metrics-as-scores)
-================
 
-- <a href="#usage" id="toc-usage"><span
-  class="toc-section-number">1</span> Usage</a>
-  - <a href="#text-based-user-interface-tui"
-    id="toc-text-based-user-interface-tui"><span
-    class="toc-section-number">1.1</span> Text-based User Interface
-    (TUI)</a>
-  - <a href="#web-application" id="toc-web-application"><span
-    class="toc-section-number">1.2</span> Web Application</a>
-  - <a href="#development-setup" id="toc-development-setup"><span
-    class="toc-section-number">1.3</span> Development Setup</a>
-    - <a href="#setting-up-a-virtual-environment"
-      id="toc-setting-up-a-virtual-environment"><span
-      class="toc-section-number">1.3.1</span> Setting Up a Virtual
-      Environment</a>
-    - <a href="#installing-packages" id="toc-installing-packages"><span
-      class="toc-section-number">1.3.2</span> Installing Packages</a>
-    - <a href="#running-tests" id="toc-running-tests"><span
-      class="toc-section-number">1.3.3</span> Running Tests</a>
-- <a href="#example-usage" id="toc-example-usage"><span
-  class="toc-section-number">2</span> Example Usage</a>
-  - <a href="#concrete-example-using-the-qualitas.class-corpus-dataset"
-    id="toc-concrete-example-using-the-qualitas.class-corpus-dataset"><span
-    class="toc-section-number">2.1</span> Concrete Example Using the
-    Qualitas.class Corpus Dataset</a>
-  - <a href="#concrete-example-using-the-iris-dataset"
-    id="toc-concrete-example-using-the-iris-dataset"><span
-    class="toc-section-number">2.2</span> Concrete Example Using the Iris
-    Dataset</a>
-  - <a href="#diamonds-example" id="toc-diamonds-example"><span
-    class="toc-section-number">2.3</span> Diamonds Example</a>
-- <a href="#datasets" id="toc-datasets"><span
-  class="toc-section-number">3</span> Datasets</a>
-  - <a href="#use-your-own" id="toc-use-your-own"><span
-    class="toc-section-number">3.1</span> Use Your Own</a>
-  - <a href="#known-datasets" id="toc-known-datasets"><span
-    class="toc-section-number">3.2</span> Known Datasets</a>
-- <a href="#personalizing-the-web-application"
-  id="toc-personalizing-the-web-application"><span
-  class="toc-section-number">4</span> Personalizing the Web
-  Application</a>
-- <a href="#references" id="toc-references">References</a>
+- [<span class="toc-section-number">1</span> Usage](#usage)
+  - [<span class="toc-section-number">1.1</span> Text-based User
+    Interface (TUI)](#text-based-user-interface-tui)
[…]
