+++ paper/refs.bib
@@ -16,13 +16,12 @@
 }
 
 @inproceedings{benlarbi2000thresh,
-  author    = {Sa{\"{\i}}da Benlarbi and
+  author    = {Saida Benlarbi and
                Khaled El Emam and
                Nishith Goel and
                Shesh N. Rai},
-  title     = {Thresholds for Object-Oriented Measures},
-  booktitle = {11th International Symposium on Software Reliability Engineering {(ISSRE}
-               2000), 8-11 October 2000, San Jose, CA, {USA}},
+  title     = {{Thresholds for Object-Oriented Measures}},
+  booktitle = {{11th International Symposium on Software Reliability Engineering {(ISSRE} 2000), 8-11 October 2000, San Jose, CA, {USA}}},
   pages     = {24--39},
   publisher = {{IEEE} Computer Society},
   year      = {2000},
@@ -41,8 +40,7 @@
                Michele Lanza and
                Andrian Marcus},
   title     = {Deriving metric thresholds from benchmark data},
-  booktitle = {26th {IEEE} International Conference on Software Maintenance {(ICSM}
-               2010), September 12-18, 2010, Timisoara, Romania},
+  booktitle = {{26th {IEEE} International Conference on Software Maintenance {(ICSM} 2010), September 12-18, 2010, Timisoara, Romania}},
   pages     = {1--10},
   publisher = {{IEEE} Computer Society},
   year      = {2010},
@@ -139,7 +137,7 @@
     publisher = {Association for Computing Machinery},
     address = {New York, NY, USA},
     doi = {10.1145/3338906.3341462},
-    booktitle = {Proceedings of the 2019 27th ACM Joint Meeting on European Software Engineering Conference and Symposium on the Foundations of Software Engineering},
+    booktitle = {{Proceedings of the 2019 27th ACM Joint Meeting on European Software Engineering Conference and Symposium on the Foundations of Software Engineering}},
     pages = {1144--1147},
     numpages = {4},
     keywords = {software fault prediction, software metrics, taxonomy},
@@ -151,8 +149,8 @@
   author    = {Syed Rashid Aziz and
                Tamim Ahmed Khan and
                Aamer Nadeem},
-  title     = {Experimental Validation of Inheritance Metrics' Impact on Software
-               Fault Prediction},
+  title     = {{Experimental Validation of Inheritance Metrics' Impact on Software
+               Fault Prediction}},
   journal   = {{IEEE} Access},
   volume    = {7},
   pages     = {85262--85275},
@@ -166,8 +164,8 @@
 @article{chidamber1994metrics,
   author    = {Shyam R. Chidamber and
                Chris F. Kemerer},
-  title     = {A Metrics Suite for Object Oriented Design},
-  journal   = {{IEEE} Trans. Software Eng.},
+  title     = {{A Metrics Suite for Object Oriented Design}},
+  journal   = {{{IEEE} Trans. Software Eng.}},
   volume    = {20},
   number    = {6},
   pages     = {476--493},
@@ -188,10 +186,8 @@
                Ivica Crnkovic and
                Marsha Chechik and
                Mark Harman},
-  title     = {Introducing quality models based on joint probabilities},
-  booktitle = {Proceedings of the 40th International Conference on Software Engineering:
-               Companion Proceeedings, {ICSE} 2018, Gothenburg, Sweden, May 27 -
-               June 03, 2018},
+  title     = {{Introducing quality models based on joint probabilities}},
[…]
