### 2023-04-21 0397680
Add refs for scipy, numpy, and Welch's t-test
