### 2023-04-21 241d851
Rewrite section about statistical tests and ideal-value transformations
