### 2023-04-21 1a1eede
Fix issues related to citing and quotes
