+++ paper/refs.bib
@@ -10,9 +10,14 @@
 }
 
 @inproceedings{kaner2004software,
-  title={Software Engineering Metrics: What Do They Measure and How Do We Know?},
   author={Cem Kaner and Walter P. Bond},
-  year={2004}
+  booktitle={{10th {IEEE} International Software Metrics Symposium {(METRICS} 2004), 11-17 September 2004, Chicago, IL, {USA}}},
+  publisher={{{IEEE} Computer Society}},
+  title={{Software Engineering Metrics: What Do They Measure and How Do We Know?}}, 
+  year={2004},
+  pages={1--12},
+  url={https://web.archive.org/web/20221002003552/https://kaner.com/pdfs/metrics2004.pdf},
+  urldate={2022-10-02}
 }
 
 @inproceedings{benlarbi2000thresh,
@@ -374,7 +379,7 @@
 
 @article{welch1947,
     author = {Welch, B. L.},
-    title = {The Generalization of ‘Student's’ Problem When Several Different Population Variances Are Involved},
+    title = {{The Generalization of ‘Student's’ Problem When Several Different Population Variances Are Involved}},
     journal = {Biometrika},
     volume = {34},
     number = {1-2},
