### 2023-04-25 3777d5b
Remove abbreviations
