+++ paper/refs.bib
@@ -72,7 +72,7 @@
                Morgan Ericsson and
                Anna Wingkvist},
   title     = {Copula-based software metrics aggregation},
-  journal   = {Softw. Qual. J.},
+  journal   = {{Software Quality Journal}},
   volume    = {29},
   number    = {4},
   pages     = {863--899},
@@ -88,7 +88,7 @@
   author    = {Julian Blank and
                Kalyanmoy Deb},
   title     = {Pymoo: Multi-Objective Optimization in Python},
-  journal   = {{IEEE} Access},
+  journal   = {{{IEEE} Access}},
   volume    = {8},
   pages     = {89497--89509},
   year      = {2020},
@@ -105,7 +105,7 @@
                Marco Tulio Valente and
                Roberto da Silva Bigonha},
   title     = {{Qualitas.class} corpus: a compiled version of the {qualitas} corpus},
-  journal   = {{ACM} {SIGSOFT} Softw. Eng. Notes},
+  journal   = {{{ACM} {SIGSOFT} Software Engineering Notes}},
   volume    = {38},
   number    = {5},
   pages     = {1--4},
@@ -122,7 +122,7 @@
                Lionel C. Briand and
                Walc{\'{e}}lio L. Melo},
   title     = {A Validation of Object-Oriented Design Metrics as Quality Indicators},
-  journal   = {{IEEE} Trans. Software Eng.},
+  journal   = {{{IEEE} Transactions on Software Engineering}},
   volume    = {22},
   number    = {10},
   pages     = {751--761},
@@ -170,7 +170,7 @@
   author    = {Shyam R. Chidamber and
                Chris F. Kemerer},
   title     = {{A Metrics Suite for Object Oriented Design}},
-  journal   = {{{IEEE} Trans. Software Eng.}},
+  journal   = {{{IEEE} Transactions on Software Engineering}},
   volume    = {20},
   number    = {6},
   pages     = {476--493},
@@ -346,7 +346,7 @@
             {van Mulbregt}, Paul and {SciPy 1.0 Contributors}},
   title   = {{{SciPy} 1.0: Fundamental Algorithms for Scientific
             Computing in Python}},
-  journal = {Nature Methods},
+  journal = {{Nature Methods}},
   year    = {2020},
   volume  = {17},
   pages   = {261--272},
