### 2023-04-25 8055212
Add references to Sturm and Fox
