+++ paper/refs.bib
@@ -355,26 +355,26 @@
 }
 
 @Article{numpy,
- title         = {{Array programming with {NumPy}}},
- author        = {Charles R. Harris and K. Jarrod Millman and St{\'{e}}fan J.
-                 van der Walt and Ralf Gommers and Pauli Virtanen and David
-                 Cournapeau and Eric Wieser and Julian Taylor and Sebastian
-                 Berg and Nathaniel J. Smith and Robert Kern and Matti Picus
-                 and Stephan Hoyer and Marten H. van Kerkwijk and Matthew
-                 Brett and Allan Haldane and Jaime Fern{\'{a}}ndez del
-                 R{\'{i}}o and Mark Wiebe and Pearu Peterson and Pierre
-                 G{\'{e}}rard-Marchant and Kevin Sheppard and Tyler Reddy and
-                 Warren Weckesser and Hameer Abbasi and Christoph Gohlke and
-                 Travis E. Oliphant},
- year          = {2020},
- month         = sep,
- journal       = {Nature},
- volume        = {585},
- number        = {7825},
- pages         = {357--362},
- doi           = {10.1038/s41586-020-2649-2},
- publisher     = {{Springer Science and Business Media {LLC}}},
- url           = {https://doi.org/10.1038/s41586-020-2649-2}
+  title         = {{Array programming with {NumPy}}},
+  author        = {Charles R. Harris and K. Jarrod Millman and St{\'{e}}fan J.
+                  van der Walt and Ralf Gommers and Pauli Virtanen and David
+                  Cournapeau and Eric Wieser and Julian Taylor and Sebastian
+                  Berg and Nathaniel J. Smith and Robert Kern and Matti Picus
+                  and Stephan Hoyer and Marten H. van Kerkwijk and Matthew
+                  Brett and Allan Haldane and Jaime Fern{\'{a}}ndez del
+                  R{\'{i}}o and Mark Wiebe and Pearu Peterson and Pierre
+                  G{\'{e}}rard-Marchant and Kevin Sheppard and Tyler Reddy and
+                  Warren Weckesser and Hameer Abbasi and Christoph Gohlke and
+                  Travis E. Oliphant},
+  year          = {2020},
+  month         = sep,
+  journal       = {Nature},
+  volume        = {585},
+  number        = {7825},
+  pages         = {357--362},
+  doi           = {10.1038/s41586-020-2649-2},
+  publisher     = {{Springer Science and Business Media {LLC}}},
+  url           = {https://doi.org/10.1038/s41586-020-2649-2}
 }
 
 @article{welch1947,
@@ -392,5 +392,42 @@
     eprint = {https://academic.oup.com/biomet/article-pdf/34/1-2/28/553093/34-1-2-28.pdf},
 }
 
+@article{sturm2005anova,
+    author = {Rachel Sturm-Beiss},
+    title = {{A Visualization Tool for One- and Two-Way Analysis of Variance}},
+    journal = {{Journal of Statistics Education}},
+    volume = {13},
+    number = {1},
+    pages = {1--7},
+    year  = {2005},
+    publisher = {Taylor \& Francis},
+    doi = {10.1080/10691898.2005.11910641}
+}
+
+@article{fox2009hyptests,
+  author       = {John Fox and
+                  Michael Friendly and
+                  Georges Monette},
+  title        = {{Visualizing hypothesis tests in multivariate linear models: the \emph{heplots}
+                  package for {R}}},
+  journal      = {Computational Statistics},
+  volume       = {24},
+  number       = {2},
+  pages        = {233--246},
+  year         = {2009},
[…]
