### 2023-04-25 c863be7
Replace latex markup with markdown markup
