### 2023-04-25 1fcf641
Add how we obtain a distance from a sample
