### 2023-04-25 MrShoenel/metrics-as-scores#1 @MrShoenel
@**Statement of need**: I did try to simplify some of the terms. 'raw data' does not appear any longer, the word 'metric' is now always used as 'software metric' to make it clear. 'score', however, needs to remain as-is because of the already published QRS paper (we cannot change the name in retrospect). It is made clear what the term means and how it is used in the first paragraph. Similarly, I have attempted to make it more explicit what the words 'context' and 'distance' mean.
