### 2023-04-25 #4913 @mdhaber
@mikldk I would like to wrap up my rather extensive (see https://github.com/MrShoenel/metrics-as-scores/issues?q=is%3Aissue) review, but the check mark for "Functionality" is very broad.

> Have the functional claims of the software been confirmed?

From a software perspective, I can't confirm that everything the software intends to do is done correctly, and from a domain perspective, I can't confirm that everything the software intends to do is appropriate. I attempted to verify the functionality and saw evidence that the software does some of the things it attempts to do. If this sufficient, I can check the box.
