### 2023-05-03 #4913 @mdhaber
Yes, my completed checklist is [above](https://github.com/openjournals/joss-reviews/issues/4913#issuecomment-1355604820).
