### 2023-05-04 #4913 @MrShoenel
I checked the missing DOI, it is *not* a valid DOI for this item. It is a book that still has no DOI associated and the entry is up-2-date.
