### 2023-05-25 #4819 @kostiantyn-kucher
Dear @mikldk ,

Yes, I think I will be able to act as a reviewer, although I cannot guarantee that I would submit my review until the beginning of July (but I will do my best).

For a full disclosure regarding potential conflicts of interest, please note that one of the submission co-authors, Morgan Ericsson, and I were among the co-editors of the SweDS21 workshop: https://urn.kb.se/resolve?urn=urn%3Anbn%3Ase%3Alnu%3Adiva-108364 , and I was also employed at Linnaeus University (the affiliation of all the submission authors) until August 2022  (I am currently not affiliated with LNU anymore and I have not had or currently have any co-authored publications, funding, etc. with the submission authors, though).
