### 2023-07-04 MrShoenel/metrics-as-scores#15 @kostiantyn-kucher
Additionally, please notice the following typo in the repository documentation under "Web Application":  "It feates".
