### 2023-07-16 c728c26
Add test for performing Kruskal-Walllis test
