### 2023-07-16 2c25bbe
Push version to v2.6.0
