### 2023-07-16 cd962af
Downgrade scipy to v1.10.1 to retain compatibility with existing datasets (renaming of 'gilbrat')
