### 2023-07-16 7993ad2
Add the context's name to each line glyph and refactor the on-hover tooltip to show it
