+++ README.qmd
@@ -126,7 +126,7 @@ There is no need to modify any source code, if you want to do one of the followi
 Metrics As Scores' main feature is perhaps the Web Application.
 It can be run directly and locally from the TUI using a selected dataset (you may download a known dataset or use your own).
 The Web Application allows to visually inspect each *feature* across all the defined *groups*.
-It feates the PDF/PMF, CDF and CCDF, as well as the PPF for each feature in each group.
+It features the PDF/PMF, CDF and CCDF, as well as the PPF for each feature in each group.
 It offers five different principal types of densities: Parametric, Parametric (discrete), Empirical, Empirical (discrete), and (approximate) Kernel Density Estimation.
 The Web Application includes a detailed [Help section](https://metrics-as-scores.ml/#help) that should answer most of your questions.
 
