### 2023-07-16 d47d6b2
Prepare for stable release of v2.7.0
