### 2023-07-16 f62e1e2
Address all issues related to issue #15
