### 2023-07-16 MrShoenel/metrics-as-scores#15 @MrShoenel
1. As you suggested, this design choice is indeed quite straightforward. I have added a sentence in section "MAS -- The Tool- and Analysis Suite".
2. I have made clear that studying the Qualitas.class corpus was the *initial* purpose of MAS. A list of the existing other datasets, including a brief discussion, was added.
3.
  - CorpusVis: Thanks for making me aware of this, I had not seen this before. The goal of that work is different from ours, but still related and relevant. Note that they study the ordinary corpus, not the compiled one though. I have added it to the related works section.
  - moreThanANOVA: This is somewhat related (the systematic approach to comparing quantities is). We have decided to add a reference in the introduction, rather than related work.
  - seaborn: seaborn is a visualization framework that is not actually used by MAS. Instead, we make use of bokeh, which is also cited. seaborn does not offer any other form of visualization that could have been more appropriate than the one implemented by us. Therefore, we have decided not to include or reference it at this point.
4.
  - 4.1) The tooltip now shows the X/Y coordinates (values), as well as the group's name that is associated with the currently hovered line.
  - 4.2) After discussion, we decided to keep these brief results, as MAS was principally created for the QCC dataset. Also, since we added a brief discussion for the other datasets, we feel that this should stay.
5.
  - 5.1) Changed to "allows the user to assess"
  - 5.2) We have added this and made clear that the application/package is self-sustained.
  - 5.3) Those abbreviations are now properly introduced.
  - 5.4) Both references are indeed required here, as the first one refers to the one-sample test and the second one to the two-sample test (which is an extension made by Anderson).
  - 5.5) We have manually checked every reference now. We used DBLP to obtain precise bibtex records and, if possible/available, even used the publisher's how-to-cite guidelines and the published paper as authoritative guidelines.
    - 5.5.1) That was corrected now!
    - 5.5.2) Also corrected now!

The additional issue of the misspelling in the repository documentation was corrected.

I will leave this issue open for you to close, once you are satisfied with out responses.
