### 2023-07-16 5e9e0e0
Fix some author names for the references
