+++ refs.bib
@@ -104,7 +104,7 @@
 }
 
 @article{cramer1928,
-  author = {Harald Cramér },
+  author = {Harald Cramér},
   title = {On the composition of elementary errors},
   journal = {Scandinavian Actuarial Journal},
   volume = {1928},
@@ -118,7 +118,7 @@
 }
 
 @article{stephens1974,
-  author = { M. A. Stephens },
+  author = {M. A. Stephens},
   title = {EDF Statistics for Goodness of Fit and Some Comparisons},
   journal = {Journal of the American Statistical Association},
   volume = {69},
@@ -145,7 +145,7 @@
 }
 
 @article{epps1986,
-  author = {T.W. Epps and Kenneth J. Singleton },
+  author = {Epps, T.W. and Singleton, Kenneth J.},
   title = {An omnibus test for the two-sample problem using the empirical characteristic function},
   journal = {Journal of Statistical Computation and Simulation},
   volume = {26},
@@ -159,7 +159,7 @@
 }
 
 @incollection{chambers2017,
-  author = {John M. Chambers, Anne E. Freeny, Richard M. Heiberger},
+  author = {Chambers, John M. and Freeny, Anne E. and Heiberger, Richard M.},
   doi = {10.1201/9780203738535},
   url = {https://doi.org/10.1201/9780203738535},
   year = {2017},
