### 2023-07-18 MrShoenel/metrics-as-scores#15 @kostiantyn-kucher
Thank you for the updates! I would say that most of my concerns and suggestions have been addressed.

Several further suggestions for minor improvements for the manuscript (I would not insist on these and could close the issue instead, if you prefer; but I believe these should be very easy for you to address):

1. Line 14 (in the PDF generated on July 16): "data visualization and -exploration" — I guess the dash is a typo here...
2. Figure 1: I believe this is still the old screenshot, as I do not see the group title within the tooltip. I would suggest updating the screenshot (with a visible tooltip), say, from the version currently deployed at https://metrics-as-scores.ml/
3. Regarding the bibliography, please note that the respective publishers' guidelines, etc. are not ideal and might contradict each other. I would suggest aiming for consistency in the resulting JOSS manuscript, not the contradictory guidelines and imperfect Bibtex entries... I would strongly suggest to at least edit the following parts of the references (w.r.t. capitalization): "... european ... tallin, estonia, august ..." (Caulo 2019); "... java ... asia pacific ... sydnes, australia, november 30 - december..." (Tempero 2010).
