### 2023-07-21 MrShoenel/metrics-as-scores#15 @MrShoenel
Thanks for your feedback, I think you make a valid point about the references, so I am going to fix that (and the other things, too)!
