### 2023-07-28 1018bc8
Enforce proper capitalization on all references
