+++ paper/refs.bib
@@ -12,7 +12,7 @@
 
 @inproceedings{kaner2004software,
   author={Cem Kaner and Walter P. Bond},
-  booktitle={10th {IEEE} International Software Metrics Symposium, {METRICS} 2004, Chicago, IL, {USA}, September 11--17, 2004},
+  booktitle={{10th {IEEE} International Software Metrics Symposium, {METRICS} 2004, Chicago, IL, {USA}, September 11--17, 2004}},
   publisher={{{IEEE} Computer Society}},
   title={{Software Engineering Metrics: What Do They Measure and How Do We Know?}}, 
   year={2004},
@@ -27,7 +27,7 @@
                Nishith Goel and
                Shesh N. Rai},
   title     = {{Thresholds for Object-Oriented Measures}},
-  booktitle = {11th International Symposium on Software Reliability Engineering, {ISSRE} 2000, San Jose, CA, {USA}, October 8--11, 2000},
+  booktitle = {{11th International Symposium on Software Reliability Engineering, {ISSRE} 2000, San Jose, CA, {USA}, October 8--11, 2000}},
   pages     = {24--39},
   publisher = {{IEEE} Computer Society},
   year      = {2000},
@@ -45,8 +45,8 @@
   editor    = {Radu Marinescu and
                Michele Lanza and
                Andrian Marcus},
-  title     = {Deriving Metric Thresholds from Benchmark Data},
-  booktitle = {26th {IEEE} International Conference on Software Maintenance, {ICSM} 2010, Timisoara, Romania, September 12--18, 2010},
+  title     = {{Deriving Metric Thresholds from Benchmark Data}},
+  booktitle = {{26th {IEEE} International Conference on Software Maintenance, {ICSM} 2010, Timisoara, Romania, September 12--18, 2010}},
   pages     = {1--10},
   publisher = {{IEEE} Computer Society},
   year      = {2010},
@@ -58,10 +58,10 @@
 }
 
 @book{grady1992practical,
-  title={Practical Software Metrics For Project Management And Process Improvement},
+  title={{Practical Software Metrics For Project Management And Process Improvement}},
   author={Grady, Robert B},
   year={1992},
-  publisher={Prentice Hall, Inc.},
+  publisher={{Prentice Hall, Inc.}},
   isbn10={0137203845},
   isbn={9780137203840},
   isbn13={9780137203840},
@@ -73,7 +73,7 @@
                Welf L{\"{o}}we and
                Morgan Ericsson and
                Anna Wingkvist},
-  title     = {Copula-based software metrics aggregation},
+  title     = {{Copula-based software metrics aggregation}},
   journal   = {{Software Quality Journal}},
   volume    = {29},
   number    = {4},
@@ -106,7 +106,7 @@
                Luis Fernando Miranda and
                Marco Tulio Valente and
                Roberto da Silva Bigonha},
-  title     = {{Qualitas.class} corpus: a compiled version of the {qualitas} corpus},
+  title     = {{{Qualitas.class} corpus: a compiled version of the {qualitas} corpus}},
   journal   = {{{ACM} {SIGSOFT} Software Engineering Notes}},
   volume    = {38},
   number    = {5},
@@ -123,7 +123,7 @@
   author    = {Victor R. Basili and
                Lionel C. Briand and
                Walc{\'{e}}lio L. Melo},
-  title     = {A Validation of Object-Oriented Design Metrics as Quality Indicators},
+  title     = {{A Validation of Object-Oriented Design Metrics as Quality Indicators}},
   journal   = {{{IEEE} Transactions on Software Engineering}},
   volume    = {22},
   number    = {10},
@@ -142,7 +142,7 @@
[…]
