### 2023-07-28 d552247
Update application screenshot and description in paper
