### 2023-07-28 MrShoenel/metrics-as-scores#15 @MrShoenel
I have made some changes to address your points:

1. This was not a typo, it is a hyphenated compound, essentially shortening the two concepts of "data visualization" and "data exploration" into "data visualization and -exploration", in order to avoid repetition of the word "data". Without the hyphen, those would be the two separate concepts of "data visualization" and "exploration".
2. I have updated the screenshot to reflect the latest version of the application. In the meantime, it appears that .ml domains [do no longer work](https://forum.infinityfree.net/t/all-ml-domains-are-down-due-to-registry-transfer/79379), so I have removed references to the hosted web application as well.
3. The venue titles now all follow the same common format. It was a bit tricky to enforce proper capitalization as the generation process for the paper's PDF is different from the one I use locally (where capitalization was correct).
