### 2023-07-28 #4913 @MrShoenel
@mikldk I believe all remaining comments are now properly addressed (see <https://github.com/MrShoenel/metrics-as-scores/issues/15>).

I looked over the generated PDF from the editorialbot, and I believe it is fine.
