### 2023-07-29 MrShoenel/metrics-as-scores#15 @kostiantyn-kucher
Thank you, I think you have addressed all of my comments now, and I am happy to close this issue!
