### 2023-07-31 #4913 @mikldk
@MrShoenel A minor comment:

I am not a native English speaker, but I am in doubt of the "-" in the title "Metrics As Scores: A Tool- and Analysis Suite". Should it be there? Similar for other places (e.g. "data visualization and -exploration").
