### 2023-07-31 #4913 @MrShoenel
The second reviewer actually also brought up the example of "data visualization and -exploration". In that case, as well as in the title, these are hyphenated compounds. If expanded, they would read "data visualization and data exploration" as well as "Metrics As Scores: A Tool Suite and Analysis Suite". So we make use of these amalgamations in order to avoid repeating words.
Long story short, it's fine :)
