### 2023-08-01 e0ea8fb
Remove duplicate references that are already available in paper
