+++ refs.bib
@@ -1,14 +1,3 @@
-@inproceedings{honel2022qrs,
-  author    = {Sebastian Hönel and Morgan Ericsson and Welf Löwe and Anna Wingkvist},
-  title     = {Contextual Operationalization of Metrics As Scores: Is My Metric Value Good?},
-  booktitle = {22nd {IEEE} International Conference on Software Quality, Reliability and Security, {QRS} 2022, Guangzhou, China, December 5-9, 2022},
-  pages     = {333--343},
-  publisher = {{IEEE}},
-  year      = {2022},
-  url       = {https://doi.org/10.1109/QRS57517.2022.00042},
-  doi       = {10.1109/QRS57517.2022.00042}
-}
-
 @Book{ggplot2,
 	author = {Hadley Wickham},
 	title = {{ggplot2}: Elegant Graphics for Data Analysis},
@@ -34,75 +23,6 @@
   bibsource = {dblp computer science bibliography, https://dblp.org}
 }
 
-@article{terra2013qcc,
-  author    = {Ricardo Terra and
-               Luis Fernando Miranda and
-               Marco T{\'{u}}lio Valente and
-               Roberto da Silva Bigonha},
-  title     = {Qualitas.class corpus: a compiled version of the qualitas corpus},
-  journal   = {{ACM} {SIGSOFT} Softw. Eng. Notes},
-  volume    = {38},
-  number    = {5},
-  pages     = {1--4},
-  year      = {2013},
-  url       = {https://doi.org/10.1145/2507288.2507314},
-  doi       = {10.1145/2507288.2507314},
-  timestamp = {Fri, 19 Aug 2022 14:05:12 +0200},
-  biburl    = {https://dblp.org/rec/journals/sigsoft/TerraMVB13.bib},
-  bibsource = {dblp computer science bibliography, https://dblp.org}
-}
-
-@inproceedings{tempero2010qualitas,
-  author    = {Ewan D. Tempero and
-               Craig Anslow and
-               Jens Dietrich and
-               Ted Han and
-               Jing Li and
-               Markus Lumpe and
-               Hayden Melton and
-               James Noble},
-  editor    = {Jun Han and
-               Tran Dan Thu},
-  title     = {The Qualitas Corpus: {A} Curated Collection of Java Code for Empirical
-               Studies},
-  booktitle = {17th Asia Pacific Software Engineering Conference, {APSEC} 2010, Sydney,
-               Australia, November 30 - December 3, 2010},
-  pages     = {336--345},
-  publisher = {{IEEE} Computer Society},
-  year      = {2010},
-  url       = {https://doi.org/10.1109/APSEC.2010.46},
-  doi       = {10.1109/APSEC.2010.46},
-  timestamp = {Fri, 27 Mar 2020 08:46:40 +0100},
-  biburl    = {https://dblp.org/rec/conf/apsec/TemperoADHLLMN10.bib},
-  bibsource = {dblp computer science bibliography, https://dblp.org}
-}
-
-@article{tukey1949anova,
- ISSN = {0006341X, 15410420},
- URL = {http://www.jstor.org/stable/3001913},
- doi = {10.2307/3001913},
- author = {John W. Tukey},
- journal = {Biometrics},
- number = {2},
- pages = {99--114},
- publisher = {[Wiley, International Biometric Society]},
- title = {Comparing Individual Means in the Analysis of Variance},
- urldate = {2023-02-16},
- volume = {5},
- year = {1949}
-}
-
-@ARTICLE{pymoo2020,
-  author={Blank, Julian and Deb, Kalyanmoy},
-  journal={IEEE Access}, 
-  title={Pymoo: Multi-Objective Optimization in Python}, 
[…]
