+++ README.md
@@ -53,7 +53,7 @@ between versions
 [**`v1.0.8`**](https://github.com/MrShoenel/metrics-as-scores/tree/v1.0.8)
 and **`v2.x.x`**.
 
-The current version is `v2.7.0`.
+The current version is `v2.8.0`.
 
 From version **`v2.x.x`** it has the following new features:
 
@@ -404,7 +404,7 @@ domain name coming soon) and one to this repository, that is,
 
 <div id="refs" class="references csl-bib-body hanging-indent">
 
-<div id="ref-anderson1962" class="csl-entry">
+<div id="ref-Anderson1962" class="csl-entry">
 
 Anderson, T. W. 1962. “<span class="nocase">On the Distribution of the
 Two-Sample Cramer-von Mises Criterion</span>.” *The Annals of
@@ -413,20 +413,21 @@ Mathematical Statistics* 33 (3): 1148–59.
 
 </div>
 
-<div id="ref-pymoo2020" class="csl-entry">
+<div id="ref-pymoo" class="csl-entry">
 
-Blank, Julian, and Kalyanmoy Deb. 2020. “Pymoo: Multi-Objective
-Optimization in Python.” *IEEE Access* 8: 89497–509.
-<https://doi.org/10.1109/ACCESS.2020.2990567>.
+Blank, Julian, and Kalyanmoy Deb. 2020. “<span class="nocase">pymoo:
+Multi-Objective Optimization in Python</span>.” *IEEE Access* 8:
+89497–509. <https://doi.org/10.1109/ACCESS.2020.2990567>.
 
 </div>
 
-<div id="ref-chambers2017" class="csl-entry">
+<div id="ref-chambers2017statistical" class="csl-entry">
 
 Chambers, John M., Anne E. Freeny, and Richard M. Heiberger. 2017.
-“Analysis of Variance; Designed Experiments.” In *Statistical Models in
-S*, edited by John M. Chambers and Trevor J. Hastie, 1st ed. Routledge.
-<https://doi.org/10.1201/9780203738535>.
+“<span class="nocase">Analysis of Variance; Designed
+Experiments</span>.” In *<span class="nocase">Statistical Models in
+S</span>*, edited by John M. Chambers and Trevor J. Hastie, 1st ed.
+Routledge. <https://doi.org/10.1201/9780203738535>.
 
 </div>
 
@@ -438,11 +439,12 @@ Cramér, Harald. 1928. “On the Composition of Elementary Errors.”
 
 </div>
 
-<div id="ref-epps1986" class="csl-entry">
+<div id="ref-Epps1986" class="csl-entry">
 
-Epps, T. W., and Kenneth J. Singleton. 1986. “An Omnibus Test for the
-Two-Sample Problem Using the Empirical Characteristic Function.”
-*Journal of Statistical Computation and Simulation* 26 (3-4): 177–203.
+Epps, T. W., and Kenneth J. Singleton. 1986. “<span class="nocase">An
+Omnibus Test for the Two-Sample Problem Using the Empirical
+Characteristic Function</span>.” *Journal of Statistical Computation and
+Simulation* 26 (3-4): 177–203.
 <https://doi.org/10.1080/00949658608810963>.
 
 </div>
@@ -486,48 +488,53 @@ Zenodo. <https://doi.org/10.5281/zenodo.7669664>.
 
 </div>
 
-<div id="ref-honel2022qrs" class="csl-entry">
+<div id="ref-honel2022mas" class="csl-entry">
 
 Hönel, Sebastian, Morgan Ericsson, Welf Löwe, and Anna Wingkvist. 2022.
-“Contextual Operationalization of Metrics as Scores: Is My Metric Value
[…]
