### 2023-08-01 #4913 @MrShoenel
I have read the paper in full once more and checked the generated PDF, too. All versions and the authors' list seem to be correct.

I did create a tagged release: [`v2.8.0-JOSS`](https://github.com/MrShoenel/metrics-as-scores/releases/tag/v2.8.0-JOSS). This release is now archived at Zenodo under the DOI [10.5281/zenodo.8202326](https://doi.org/10.5281/zenodo.8202326).
