### 2023-08-01 #4913 @MrShoenel
That one missing DOI came up previously. However, it is not valid, as it refers to a review of the book, not the book itself (which has no DOI).
