### 2023-08-01 #4913 @mikldk
@openjournals/joss-eics Note that the missing DOI is not correct for the book (it's for a review of the book). The book does apparently not have a DOI as far as @MrShoenel and me can find.
