### 2023-08-04 #4913 @gkthiruvathukal
@mikldk This is fine. Most books have ISBN/ISSN. I'm ready to accept here but am going to ask @arfon whether this can or should be used as the DOI.
