### 2023-08-08 #4913 @danielskatz
If there is no DOI for the book (or any other item), you don't need to have one. This can proceed as is.
