### 2023-08-25 98b6d7a
Create CITATION.cff
